% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotdata.R
\name{addAnnotToDataLabel}
\alias{addAnnotToDataLabel}
\title{Adds html code to the data labels include the annotation}
\usage{
addAnnotToDataLabel(
  data.label.text,
  annotation,
  tmp.dat,
  prepend = FALSE,
  tspan = FALSE
)
}
\arguments{
\item{data.label.text}{A character vector containing the original data labels
which is to be annotated}

\item{annotation}{An element of the \code{annotation.list} passed to the
top level charting function. The is usually a list with named elements
such as "type", "size", "font.family", "format". Note that this
function will not handle annotation of type "Circle - xxx" or "Marker border"
because these are implemented as additional traces.}

\item{tmp.dat}{A slice of \code{annot.dat} which matches data.label.text
It is used when \code{annotation$type} is "Text".}

\item{prepend}{Logical; when true, the annotation will be added to the
beginning of data.label.text instead of the end.}

\item{tspan}{Whether to use tspan instead of span. tspan is used when the
annotation is directly inserted into an svg text element, whereas span is
used with Plotly-drawn labels (Plotly automatically converts it to tspan).}
}
\value{
The modified character vector \code{data.label.text}.
}
\description{
Adds html code to the data labels include the annotation
}
\keyword{internal}
