% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alpha.five}
\alias{alpha.five}
\title{Character vector of five letters, A to E.}
\format{Unnamed vector with five letters, A to E}
\usage{
alpha.five
}
\description{
Character vector of five letters, A to E.
}
\keyword{datasets}

