% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriescolorgradient.R
\name{AlphabeticRowNames}
\alias{AlphabeticRowNames}
\title{Sorts a matrix by row names.}
\usage{
AlphabeticRowNames(x, desc = TRUE)
}
\arguments{
\item{x}{A matrix.}

\item{desc}{Logic; if TRUE then sort order is descending, if FALSE ascending.}
}
\value{
A sorted matrix.
}
\description{
\code{AlphabeticRowNames} returns a matrix where the rows have been sorted
alphabetically by the row names
}
\examples{
data("z")
AlphabeticRowNames(z, desc = FALSE)
}

