% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areachart.R
\name{Area}
\alias{Area}
\title{Area}
\usage{
Area(x, type = "Area", fit.type = "None", fit.ignore.last = FALSE,
  fit.line.type = "dot", fit.line.width = 1, fit.line.name = "Fitted",
  global.font.family = "Arial", global.font.color = rgb(44, 44, 44,
  maxColorValue = 255), title = "", title.font.family = global.font.family,
  title.font.color = global.font.color, title.font.size = 16,
  subtitle = "", subtitle.font.family = global.font.family,
  subtitle.font.color = global.font.color, subtitle.font.size = 12,
  footer = "", footer.font.family = global.font.family,
  footer.font.color = global.font.color, footer.font.size = 8,
  footer.wrap = TRUE, footer.wrap.nchar = 100, colors = ChartColors(max(1,
  ncol(x), na.rm = TRUE)), fit.line.colors = colors, opacity = NULL,
  background.fill.color = rgb(255, 255, 255, maxColorValue = 255),
  background.fill.opacity = 0,
  charting.area.fill.color = background.fill.color,
  charting.area.fill.opacity = 0, legend.show = TRUE,
  legend.fill.color = background.fill.color, legend.fill.opacity = 0,
  legend.border.color = rgb(44, 44, 44, maxColorValue = 255),
  legend.border.line.width = 0, legend.font.color = global.font.color,
  legend.font.family = global.font.family, legend.font.size = 10,
  legend.position.x = 1.02, legend.position.y = 1, legend.ascending = NA,
  margin.top = NULL, margin.bottom = NULL, margin.left = NULL,
  margin.right = NULL, margin.inner.pad = NULL, grid.show = TRUE,
  y.title = "", y.title.font.color = global.font.color,
  y.title.font.family = global.font.family, y.title.font.size = 12,
  y.line.width = 0, y.line.color = rgb(0, 0, 0, maxColorValue = 255),
  y.tick.mark.length = 5, y.bounds.minimum = NULL,
  y.bounds.maximum = NULL, y.tick.distance = NULL, y.zero = TRUE,
  y.zero.line.width = 0, y.zero.line.color = rgb(225, 225, 225,
  maxColorValue = 255), y.data.reversed = FALSE, y.grid.width = 1 *
  grid.show, y.grid.color = rgb(225, 225, 225, maxColorValue = 255),
  y.tick.show = TRUE, y.tick.suffix = "", y.tick.prefix = "",
  y.tick.format = "", y.hovertext.format = y.tick.format,
  y.tick.angle = NULL, y.tick.font.color = global.font.color,
  y.tick.font.family = global.font.family, y.tick.font.size = 10,
  x.title = "", x.title.font.color = global.font.color,
  x.title.font.family = global.font.family, x.title.font.size = 12,
  x.line.width = 0, x.line.color = rgb(0, 0, 0, maxColorValue = 255),
  x.tick.marks = "", x.tick.mark.length = 5, x.bounds.minimum = NULL,
  x.bounds.maximum = NULL, x.tick.distance = NULL, x.zero = FALSE,
  x.zero.line.width = 0, x.zero.line.color = rgb(225, 225, 225,
  maxColorValue = 255), x.data.reversed = FALSE, x.grid.width = 0 *
  grid.show, x.grid.color = rgb(225, 225, 225, maxColorValue = 255),
  x.tick.show = TRUE, x.tick.format = "",
  x.hovertext.format = x.tick.format, x.tick.angle = NULL,
  x.tick.font.color = global.font.color,
  x.tick.font.family = global.font.family, x.tick.font.size = 10,
  x.tick.label.wrap = TRUE, x.tick.label.wrap.nchar = 21,
  series.marker.show = NULL, series.marker.colors = colors,
  series.marker.opacity = 1, series.marker.size = 6,
  series.line.width = NULL, series.line.colors = colors,
  series.line.opacity = 1, series.marker.border.width = 1,
  series.marker.border.colors = colors, series.marker.border.opacity = 1,
  tooltip.show = TRUE, modebar.show = FALSE, data.label.show = FALSE,
  data.label.font.family = global.font.family, data.label.font.size = 10,
  data.label.font.color = global.font.color, data.label.format = "",
  data.label.prefix = "", data.label.suffix = "",
  data.label.position = "top middle")
}
\arguments{
\item{x}{Input data may be a matrix or a vector, containing the height of the columns
to be plotted, with the name/rownames used as the column names of the chart. Numeric and date labels
will be parsed automatically.}

\item{type}{One of "Column", "Stacked Column" or "100\% Stacked Column"}

\item{fit.type}{Character; type of line of best fit. Can be one of "None", "Linear" or "Smooth" (loess local polynomial fitting).}

\item{fit.ignore.last}{Logical; whether to ignore the last data point in the fit.}

\item{fit.line.type}{Character; One of "solid", "dot", "dash, "dotdash", or length of dash "2px", "5px".}

\item{fit.line.width}{Numeric; Line width of line of best fit.}

\item{fit.line.name}{Character; Name of the line of best fit, which will appear in the hovertext.}

\item{global.font.family}{Character; font family for all occurrences of any
font attribute for the chart unless specified individually.}

\item{global.font.color}{Global font color as a named color in character format
(e.g. "black") or an rgb value (e.g. #' rgb(0, 0, 0, maxColorValue = 255)).}

\item{title}{Character; chart title.}

\item{title.font.family}{Character; title font family. Can be "Arial Black",
"Arial", "Comic Sans MS", "Courier New", "Georgia", "Impact",
"Lucida Console", "Lucida Sans Unicode", "Marlett", "Symbol", "Tahoma",
"Times New Roman", "Trebuchet MS", "Verdana", "Webdings"}

\item{title.font.color}{Title font color as a named color in character
format (e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}

\item{title.font.size}{Integer; Title font size; default = 10.}

\item{subtitle}{Character}

\item{subtitle.font.family}{Character; subtitle font family}

\item{subtitle.font.color}{subtitle font color as a named color in
character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{subtitle.font.size}{Integer; subtitle font size}

\item{footer}{Character}

\item{footer.font.family}{Character; footer font family}

\item{footer.font.color}{footer font color as a named color in
character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{footer.font.size}{Integer; footer font size}

\item{footer.wrap}{Logical; whether the footer text should be wrapped.}

\item{footer.wrap.nchar}{Number of characters (approximately) in each line of the footer when \code{footer.wordwrap} \code{TRUE}.}

\item{colors}{Character; a vector containing one or more named
colors from grDevices OR one or more specified hex value colors OR a single
named palette from grDevices, RColorBrewer, colorspace, or colorRamps.}

\item{fit.line.colors}{Character; a vector containing one or more named
colors from grDevices OR one or more specified hex value colors OR a single
named palette from grDevices, RColorBrewer, colorspace, or colorRamps.}

\item{opacity}{Opacity of area fill colors as an alpha value (0 to 1).}

\item{background.fill.color}{Background color in character format
(e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}

\item{background.fill.opacity}{Background opacity as an alpha value
(0 to 1).}

\item{charting.area.fill.color}{Charting area background color as
a named color in character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{charting.area.fill.opacity}{Charting area background opacity as an alpha value (0 to 1).}

\item{legend.show}{Logical; show the legend.}

\item{legend.fill.color}{Legend fill color as a named color in character format
(e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}

\item{legend.fill.opacity}{Legend fill opacity as an alpha value (0 to 1).}

\item{legend.border.color}{Legend border color as a named color in character
format (e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}

\item{legend.border.line.width}{Width in pixels of the border
around the legend.  0 = no border.}

\item{legend.font.color}{Legend font color as a named color in character
format (e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}

\item{legend.font.family}{Character; legend font family.}

\item{legend.font.size}{Integer; Legend font size.}

\item{legend.position.x}{A numeric controlling the position of the legend.
Values range from -0.5 (left) to 1.5 (right).}

\item{legend.position.y}{A numeric controlling the position of the legend.
Values range from 0 (bottom) to 1 (top).}

\item{legend.ascending}{Logical; TRUE for ascending, FALSE for descending.
By default, we set it to to FALSE if the chart is stacked and TRUE otherwise.}

\item{margin.top}{Margin between plot area and the top of the
graphic in pixels}

\item{margin.bottom}{Margin between plot area and the bottom of the
graphic in pixels}

\item{margin.left}{Margin between plot area and the left of the
graphic in pixels}

\item{margin.right}{Margin between plot area and the right of the
graphic in pixels}

\item{margin.inner.pad}{Padding in pixels between plot proper
and axis lines}

\item{grid.show}{Logical; whether to show grid lines.}

\item{y.title}{Character, y-axis title; defaults to chart input values;
to turn off set to "FALSE".}

\item{y.title.font.color}{y-axis title font color as a named color in
character format (e.g. "black") or an rgb value (e.g. rgb(0, 0, 0,
max = 255)).}

\item{y.title.font.family}{Character; y-axis title font family}

\item{y.title.font.size}{Integer; y-axis title font size}

\item{y.line.width}{y-axis line width in pixels (0 = no line).}

\item{y.line.color}{y-axis line color as a named color in character format
(e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}

\item{y.tick.mark.length}{Length of tick marks in pixels. Ticks are only shown when \code{y.line.width > 0}.}

\item{y.bounds.minimum}{Minimum of range for plotting; NULL = no manual range set.}

\item{y.bounds.maximum}{Maximum of range for plotting; NULL = no manual range set.}

\item{y.tick.distance}{Distance between tick marks. Requires that \code{values.bounds.minimum} and \code{values.bounds.maximum} have been set.}

\item{y.zero}{Whether the y-axis should include zero.}

\item{y.zero.line.width}{Width in pixels of zero line;}

\item{y.zero.line.color}{Color of horizontal zero line as a named
color in character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{y.data.reversed}{Logical; whether to reverse y-axis or not}

\item{y.grid.width}{Width of y-grid lines in pixels; 0 = no line}

\item{y.grid.color}{Color of y-grid lines as a named color in character
format (e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}

\item{y.tick.show}{Whether to display the y-axis tick labels}

\item{y.tick.suffix}{y-axis tick label suffix}

\item{y.tick.prefix}{y-axis tick label prefix}

\item{y.tick.format}{A string representing a d3 formatting code.
See https://github.com/d3/d3/blob/master/API.md#number-formats-d3-format}

\item{y.hovertext.format}{A string representing a d3 formatting code
See https://github.com/d3/d3/blob/master/API.md#number-formats-d3-format}

\item{y.tick.angle}{y-axis tick label angle in degrees.
90 = vertical; 0 = horizontal}

\item{y.tick.font.color}{y-axis tick label font color as a named color
in character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{y.tick.font.family}{Character; y-axis tick label font family}

\item{y.tick.font.size}{Integer; y-axis tick label font size}

\item{x.title}{Character, x-axis title; defaults to chart input values;
to turn off set to "FALSE".}

\item{x.title.font.color}{x-axis title font color as a named color in
character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{x.title.font.family}{Character; x-axis title font family}

\item{x.title.font.size}{Integer; x-axis title font size}

\item{x.line.width}{x-axis line in pixels, 0 = no line}

\item{x.line.color}{x-axis line color as a named color in character format
(e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}

\item{x.tick.marks}{Character; whether and where to show tick marks on the
x-axis.  Can be "outside", "inside", "none"}

\item{x.tick.mark.length}{Length of tick marks in pixels.}

\item{x.bounds.minimum}{Minimum of range for plotting;
NULL = no manual range set.  Must be less than x.bounds.maximum}

\item{x.bounds.maximum}{Maximum of range for
plotting; NULL = no manual range set.  Must be greater than x.bounds.minimum}

\item{x.tick.distance}{Tick mark distance in
x-axis units between minimum and maximum for plotting; NULL = no manual
range set.}

\item{x.zero}{Whether the x-axis should include zero.}

\item{x.zero.line.width}{Width in pixels of zero line.}

\item{x.zero.line.color}{Color of horizontal zero (origin) line as a named
color in character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{x.data.reversed}{Logical; whether to reverse x-axis or not}

\item{x.grid.width}{Width of y-grid lines in pixels; 0 = no line}

\item{x.grid.color}{Color of y-grid lines as a named color in character
format (e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}

\item{x.tick.show}{Whether to display the x-axis tick labels}

\item{x.tick.format}{A string representing a d3 formatting code.
See https://github.com/mbostock/d3/wiki/Formatting#numbers}

\item{x.hovertext.format}{A string representing a d3 formatting code.
See https://github.com/mbostock/d3/wiki/Formatting#numbers}

\item{x.tick.angle}{x-axis tick label angle in degrees.
90 = vertical; 0 = horizontal}

\item{x.tick.font.color}{X-axis tick label font color as a named color in
character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{x.tick.font.family}{Character; x-axis tick label font family}

\item{x.tick.font.size}{Integer; x-axis tick label font size}

\item{x.tick.label.wrap}{Logical; whether to wrap long labels on the x-axis.}

\item{x.tick.label.wrap.nchar}{Integer; number of characters in each line when \code{label.wrap} is \code{TRUE}.}

\item{series.marker.show}{Can be "none", "automatic" or a vector referencing
the plotly symbol dictionary using either numerics or strings.}

\item{series.marker.colors}{Character; a vector containing one or more named
colors from grDevices OR one or more specified hex value colors OR a single
named palette from grDevices, RColorBrewer, colorspace, or colorRamps.
be reversed. Only used if \code{series.marker.show} is \code{TRUE}.}

\item{series.marker.opacity}{Opacity for series markers as an alpha value (0 to 1).}

\item{series.marker.size}{Size in pixels of marker}

\item{series.line.width}{Thickness, in pixels, of the series line}

\item{series.line.colors}{Character; a vector containing one or more named
colors from grDevices OR one or more specified hex value colors OR a single
named palette from grDevices, RColorBrewer, colorspace, or colorRamps.}

\item{series.line.opacity}{Opacity for series lines as an alpha value (0 to 1).}

\item{series.marker.border.width}{Width in pixels of border/line
around series markers; 0 is no line}

\item{series.marker.border.colors}{Character; a vector containing one or more named
colors from grDevices OR one or more specified hex value colors OR a single
named palette from grDevices, RColorBrewer, colorspace, or colorRamps.}

\item{series.marker.border.opacity}{Opacity of border/line around
series markers as an alpha value (0 to 1).}

\item{tooltip.show}{Logical; whether to show a tooltip on hover.}

\item{modebar.show}{Logical; whether to show the zoom menu buttons or not.}

\item{data.label.show}{Logical; whether to show data labels.}

\item{data.label.font.family}{Character; font family for data label.}

\item{data.label.font.size}{Integer; Font size for data label.px.}

\item{data.label.font.color}{Font color as a named color
in character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{data.label.format}{A string representing a d3 formatting code.
See https://github.com/mbostock/d3/wiki/Formatting#numbers}

\item{data.label.prefix}{Character; prefix for data values.}

\item{data.label.suffix}{Character; suffix for data values.}

\item{data.label.position}{Character; where to place the source data
value in relation to the marker icon.  Can be "top left", "top center", "top
right", "middle left", "middle center", "middle right", "bottom left",
"bottom center", "bottom right".}
}
\description{
Area chart
}
\examples{
z <- structure(c(1L, 2L, 3L, 4L, 5L, 2L, 3L, 4L, 5L, 6L),  .Dim = c(5L, 2L),
      .Dimnames = list(c("T", "U", "V", "W", "X"), c("A", "B")))
Area(z)
}
