% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographicmaphelper.R
\name{AreasInCountry}
\alias{AreasInCountry}
\title{Get the areas codes in a country}
\usage{
AreasInCountry(country)
}
\arguments{
\item{country}{Character; name of the country}
}
\description{
When mapping the areas of a country you need to match the areas exactly.
You can use this function to look up the full list of areas in Australia.
}
\seealso{
\code{\link{CountriesOrContinents}}

\code{\link{StatesInCountry}}

\code{\link{ZipcodesInCountry}}
}
