% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aschartmatrix.R
\name{AsChartMatrix}
\alias{AsChartMatrix}
\title{Converts one or more objects to a chart matrix.}
\usage{
AsChartMatrix(y, x = NULL, transpose = FALSE, aggregate.period = "month")
}
\arguments{
\item{y}{A vector, matrix, list of vectors, data frame, or table.}

\item{x}{A vector over which y will be aggregated. Must have the same
number of elements as y.}

\item{transpose}{Logical; should the final output be transposed?}

\item{aggregate.period}{Period over which date varaibles passed to the x
argument will be aggregated. Defaults to "month", and can also be "quarter"
or "year".}
}
\value{
A chart matrix with named rows and columns.
}
\description{
\code{AsChartMatrix} checks if data is in the appropriate format and
attempts to coerce it to the appropriate format if so required.
}
\examples{
data("y.data")
data("x.data")
data("var1")
data("x.dates")
AsChartMatrix(y = y.data, x = x.data, transpose = FALSE)
AsChartMatrix(y = var1, x = x.dates, transpose = TRUE, aggregate.period = "year")
}

