% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartutils.R
\name{autoFontColor}
\alias{autoFontColor}
\title{use black or white for good contrast against colors}
\usage{
autoFontColor(colors)
}
\arguments{
\item{colors}{vector of colors which will be the background color of the}
}
\description{
use black or white for good contrast against colors
}
