% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartutils.R
\name{autoFormatLongLabels}
\alias{autoFormatLongLabels}
\title{Format long labels for html by truncating and wrapping}
\usage{
autoFormatLongLabels(
  x,
  wordwrap = FALSE,
  n = 21,
  truncate = FALSE,
  remove.empty = TRUE
)
}
\arguments{
\item{x}{Character; the text to format.}

\item{wordwrap}{Logical; whether line breaks ('<br>') should be inserted.}

\item{n}{Integer; the maximum number of characters before a line breaks is added. The break is added at the nearest word boundary.}

\item{truncate}{Logical; whether long labels should be truncated and appended with '...' instead.}

\item{remove.empty}{Logical; whether to remove zero-length strings}
}
\description{
Format long labels for html by truncating and wrapping
}
