% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{available.fonts}
\alias{available.fonts}
\title{Vector of all available fonts in Plotly}
\format{All valid Plotly marker symbols in character format.}
\usage{
available.fonts
}
\description{
Vector of all available fonts in Plotly
}
\keyword{datasets}

