% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bad.examples}
\alias{bad.examples}
\title{List of 15 test examples.}
\format{15 test scenarios returning F.}
\usage{
bad.examples
}
\description{
List of 15 test examples.
}
\keyword{datasets}

