% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barmulticolor.R
\name{BarMultiColor}
\alias{BarMultiColor}
\title{BarMultiColor}
\usage{
BarMultiColor(
  x,
  annotation.list = NULL,
  colors = ChartColors(max(1, length(x))),
  opacity = NULL,
  global.font.family = "Arial",
  global.font.color = rgb(44, 44, 44, maxColorValue = 255),
  title = "",
  title.font.family = global.font.family,
  title.font.color = global.font.color,
  title.font.size = 16,
  subtitle = "",
  subtitle.font.family = global.font.family,
  subtitle.font.color = global.font.color,
  subtitle.font.size = 12,
  footer = "",
  footer.font.family = global.font.family,
  footer.font.color = global.font.color,
  footer.font.size = 8,
  footer.wrap = TRUE,
  footer.wrap.nchar = 100,
  background.fill.color = "transparent",
  background.fill.opacity = 1,
  charting.area.fill.color = background.fill.color,
  charting.area.fill.opacity = 0,
  margin.top = NULL,
  margin.bottom = NULL,
  margin.left = NULL,
  margin.right = NULL,
  margin.inner.pad = NULL,
  margin.autoexpand = TRUE,
  grid.show = TRUE,
  y.title = "",
  y.title.font.color = global.font.color,
  y.title.font.family = global.font.family,
  y.title.font.size = 12,
  y.line.width = 0,
  y.line.color = rgb(0, 0, 0, maxColorValue = 255),
  y.tick.mark.length = 3,
  y.bounds.minimum = NULL,
  y.bounds.maximum = NULL,
  y.tick.distance = NULL,
  y.tick.maxnum = 11,
  y.zero = FALSE,
  y.zero.line.width = 0,
  y.zero.line.color = rgb(225, 225, 225, maxColorValue = 255),
  y.data.reversed = FALSE,
  y.grid.width = 0 * grid.show,
  y.grid.color = rgb(225, 225, 225, maxColorValue = 255),
  y.tick.show = TRUE,
  y.tick.suffix = "",
  y.tick.prefix = "",
  y.tick.format = "",
  y.hovertext.format = y.tick.format,
  y.tick.angle = NULL,
  y.tick.font.color = global.font.color,
  y.tick.font.family = global.font.family,
  y.tick.font.size = 10,
  x.title = "",
  x.title.font.color = global.font.color,
  x.title.font.family = global.font.family,
  x.title.font.size = 12,
  x.line.width = 0,
  x.line.color = rgb(0, 0, 0, maxColorValue = 255),
  x.tick.marks = "",
  x.tick.mark.length = 0,
  x.bounds.minimum = NULL,
  x.bounds.maximum = NULL,
  x.tick.distance = NULL,
  x.tick.maxnum = NULL,
  x.zero = TRUE,
  x.zero.line.width = 0,
  x.zero.line.color = rgb(225, 225, 225, maxColorValue = 255),
  x.data.reversed = FALSE,
  x.grid.width = 1 * grid.show,
  x.grid.color = rgb(225, 225, 225, maxColorValue = 255),
  x.tick.show = TRUE,
  x.tick.suffix = "",
  x.tick.prefix = "",
  x.tick.format = "",
  x.tick.angle = NULL,
  x.tick.font.color = global.font.color,
  x.tick.font.family = global.font.family,
  x.tick.font.size = 10,
  y.tick.label.wrap = TRUE,
  y.tick.label.wrap.nchar = 21,
  hovertext.font.family = global.font.family,
  hovertext.font.size = 11,
  marker.border.width = 1,
  marker.border.colors = colors,
  marker.border.opacity = NULL,
  data.label.show = FALSE,
  data.label.font.autocolor = FALSE,
  data.label.font.family = global.font.family,
  data.label.font.size = 10,
  data.label.font.color = global.font.color,
  data.label.format = "",
  data.label.prefix = "",
  data.label.suffix = "",
  data.label.threshold = NULL,
  x.hovertext.format = data.label.format,
  tooltip.show = TRUE,
  modebar.show = FALSE,
  zoom.enable = TRUE,
  bar.gap = 0.15
)
}
\arguments{
\item{x}{Input data may be a matrix or a vector, containing the height of the columns
to be plotted, with the name/rownames used as the column names of the chart. Numeric and date labels
will be parsed automatically.}

\item{annotation.list}{Optional list of annotations to modify the data labels.}

\item{colors}{Character; a vector containing one or more colors specified as hex codes.}

\item{opacity}{Opacity of bars as an alpha value (0 to 1).}

\item{global.font.family}{Character; font family for all occurrences of any
font attribute for the chart unless specified individually.}

\item{global.font.color}{Global font color as a named color in character format
(e.g. "black") or an a hex code.}

\item{title}{Character; chart title.}

\item{title.font.family}{Character; title font family. Can be "Arial Black",
"Arial", "Comic Sans MS", "Courier New", "Georgia", "Impact",
"Lucida Console", "Lucida Sans Unicode", "Marlett", "Symbol", "Tahoma",
"Times New Roman", "Trebuchet MS", "Verdana", "Webdings"}

\item{title.font.color}{Title font color as a named color in character
format (e.g. "black") or a hex code.}

\item{title.font.size}{Integer; Title font size; default = 10.}

\item{subtitle}{Character}

\item{subtitle.font.family}{Character; subtitle font family}

\item{subtitle.font.color}{subtitle font color as a named color in
character format (e.g. "black") or an a hex code.}

\item{subtitle.font.size}{Integer; subtitle font size}

\item{footer}{Character}

\item{footer.font.family}{Character; footer font family}

\item{footer.font.color}{footer font color as a named color in
character format (e.g. "black") or an a hex code.}

\item{footer.font.size}{Integer; footer font size}

\item{footer.wrap}{Logical; whether the footer text should be wrapped.}

\item{footer.wrap.nchar}{Number of characters (approximately) in each
line of the footer when \code{footer.wrap} \code{TRUE}.}

\item{background.fill.color}{Background color in character format (e.g. "black") or a hex code.}

\item{background.fill.opacity}{Background opacity as an alpha value (0 to 1).}

\item{charting.area.fill.color}{Charting area background color as
a named color in character format (e.g. "black") or a hex code.}

\item{charting.area.fill.opacity}{Charting area background opacity as an alpha value (0 to 1).}

\item{margin.top}{Margin between plot area and the top of the graphic in pixels}

\item{margin.bottom}{Margin between plot area and the bottom of the graphic in pixels}

\item{margin.left}{Margin between plot area and the left of the graphic in pixels}

\item{margin.right}{Margin between plot area and the right of the graphic in pixels}

\item{margin.inner.pad}{Padding in pixels between plot proper
and axis lines}

\item{margin.autoexpand}{Logical; Whether extra space can be added to the margins
to allow space for axis/legend/data labels or other chart elements.}

\item{grid.show}{Logical; whether to show grid lines.}

\item{y.title}{Character, y-axis title; defaults to chart input values;
to turn off set to "FALSE".}

\item{y.title.font.color}{y-axis title font color as a named color in
character format (e.g. "black") or a hex code.}

\item{y.title.font.family}{Character; y-axis title font family}

\item{y.title.font.size}{Integer; y-axis title font size}

\item{y.line.width}{y-axis line width in pixels (0 = no line).}

\item{y.line.color}{y-axis line color as a named color in character format
(e.g. "black") or a hex code.}

\item{y.tick.mark.length}{Length of tick marks in pixels. Ticks are only shown when \code{y.line.width > 0}.}

\item{y.bounds.minimum}{Minimum of range for plotting; For a date axis this should be supplied as a date string.
For a categorical axis, the index of the category (0-based) should be used.}

\item{y.bounds.maximum}{Maximum of range for plotting; NULL = no manual range set.}

\item{y.tick.distance}{Distance between tick marks. Requires that \code{y.bounds.minimum} and \code{y.bounds.maximum} have been set.}

\item{y.tick.maxnum}{Maximum number of ticks shown on the axis.
This setting is ignored if \code{y.tick.distance} is set or
if the axis is categorical}

\item{y.zero}{Whether the y-axis should include zero.}

\item{y.zero.line.width}{Width in pixels of zero line;}

\item{y.zero.line.color}{Color of horizontal zero line as a named
color in character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{y.data.reversed}{Logical; whether to reverse y-axis or not}

\item{y.grid.width}{Width of y-grid lines in pixels; 0 = no line}

\item{y.grid.color}{Color of y-grid lines as a named color in character
format (e.g. "black") or a hex code.}

\item{y.tick.show}{Whether to display the y-axis tick labels}

\item{y.tick.suffix}{y-axis tick label suffix}

\item{y.tick.prefix}{y-axis tick label prefix}

\item{y.tick.format}{A string representing a d3 formatting code.
See https://github.com/d3/d3/blob/master/API.md#number-formats-d3-format}

\item{y.hovertext.format}{A string representing a d3 formatting code
See https://github.com/d3/d3/blob/master/API.md#number-formats-d3-format}

\item{y.tick.angle}{y-axis tick label angle in degrees.
90 = vertical; 0 = horizontal}

\item{y.tick.font.color}{y-axis tick label font color as a named color
in character format (e.g. "black") or an a hex code.}

\item{y.tick.font.family}{Character; y-axis tick label font family}

\item{y.tick.font.size}{Integer; y-axis tick label font size}

\item{x.title}{Character, x-axis title; defaults to chart input values;
to turn off set to "FALSE".}

\item{x.title.font.color}{x-axis title font color as a named color in
character format (e.g. "black") or an a hex code.}

\item{x.title.font.family}{Character; x-axis title font family}

\item{x.title.font.size}{Integer; x-axis title font size}

\item{x.line.width}{x-axis line in pixels, 0 = no line}

\item{x.line.color}{x-axis line color as a named color in character format
(e.g. "black") or a hex code.}

\item{x.tick.marks}{Character; whether and where to show tick marks on the
x-axis.  Can be "outside", "inside", "none"}

\item{x.tick.mark.length}{Length of tick marks in pixels.}

\item{x.bounds.minimum}{Minimum of range for plotting; For a date axis this should be supplied as a date string.
For a categorical axis, the index of the category (0-based) should be used.}

\item{x.bounds.maximum}{Maximum of range for
plotting; NULL = no manual range set.  Must be greater than x.bounds.minimum}

\item{x.tick.distance}{Tick mark distance in
x-axis units between minimum and maximum for plotting; NULL = no manual
range set.}

\item{x.tick.maxnum}{Maximum number of ticks shown on the axis.
It defaults to 11 which gives the same output from plotly as NULL.
This setting is ignored if \code{x.tick.distance} is set or
if the axis is categorical}

\item{x.zero}{Whether the x-axis should include zero.}

\item{x.zero.line.width}{Width in pixels of zero line.}

\item{x.zero.line.color}{Color of horizontal zero (origin) line as a named
color in character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{x.data.reversed}{Logical; whether to reverse x-axis or not}

\item{x.grid.width}{Width of y-grid lines in pixels; 0 = no line}

\item{x.grid.color}{Color of y-grid lines as a named color in character
format (e.g. "black") or a hex code.}

\item{x.tick.show}{Whether to display the x-axis tick labels}

\item{x.tick.suffix}{x-axis tick label suffix}

\item{x.tick.prefix}{x-axis tick label prefix}

\item{x.tick.format}{A string representing a d3 formatting code.
See https://github.com/mbostock/d3/wiki/Formatting#numbers}

\item{x.tick.angle}{x-axis tick label angle in degrees.
90 = vertical; 0 = horizontal}

\item{x.tick.font.color}{X-axis tick label font color as a named color in
character format (e.g. "black") or an a hex code.}

\item{x.tick.font.family}{Character; x-axis tick label font family}

\item{x.tick.font.size}{Integer; x-axis tick label font size}

\item{y.tick.label.wrap}{Logical; whether to wrap long labels on the y-axis (vertical).}

\item{y.tick.label.wrap.nchar}{Integer; number of characters in each line when \code{y.tick.label.wrap} is \code{TRUE}.}

\item{hovertext.font.family}{Font family of hover text.}

\item{hovertext.font.size}{Font size of hover text.}

\item{marker.border.width}{Width in pixels of border/line
around series bars; 0 is no line}

\item{marker.border.colors}{Character; a vector containing one or more colors specified as hex codes.}

\item{marker.border.opacity}{Opacity of border around bars as an alpha value (0 to 1).}

\item{data.label.show}{Logical; whether to show data labels.}

\item{data.label.font.autocolor}{Logical; Whether font color should be
automatically determined. For Line and Radar charts, the data labels will
colored in the series color. For stacked bar and column charts the
data labels will be black or white depending on the color of the
bar (which background the data label). For non-stacked bar and column
charts, this option is ignored.}

\item{data.label.font.family}{Character; font family for data label.}

\item{data.label.font.size}{Integer; Font size for data label.px.}

\item{data.label.font.color}{Font color as a named color
in character format (e.g. "black") or an a hex code. This can be a single
color, a vector of colors (1 for each series/column), or a comma separated list
of colors}

\item{data.label.format}{A string representing a d3 formatting code.
See https://github.com/mbostock/d3/wiki/Formatting#numbers}

\item{data.label.prefix}{Character; prefix for data values.}

\item{data.label.suffix}{Character; suffix for data values.}

\item{data.label.threshold}{The proportion of the total range below which
data labels should not be displayed. Only applicable for pie, bar and column
charts.}

\item{x.hovertext.format}{A string representing a d3 formatting code.
See https://github.com/mbostock/d3/wiki/Formatting#numbers}

\item{tooltip.show}{Logical; whether to show a tooltip on hover.}

\item{modebar.show}{Logical; whether to show the zoom menu buttons or not.}

\item{zoom.enable}{Logical; whether to enable zoom on the chart.
For Bar and Column charts with data labels it may be useful to turn off zoom
because data labels and annotations can be misplace on zoom.}

\item{bar.gap}{Chart proportion between each bar or column if using
bar or column charts, or between each cluster of bars or columns.}
}
\description{
Bar charts with many colors
}
\examples{
z <- structure(c(1L, 2L, 3L, 4L, 5L, 2L, 3L, 4L, 5L, 6L),  .Dim = c(5L, 2L),
      .Dimnames = list(c("T", "U", "V", "W", "X"), c("A", "B")))
Bar(z, type="100\% Stacked Bar")
}
