% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barpictograph.R
\name{BarPictograph}
\alias{BarPictograph}
\title{Create bar chart composed of icons}
\usage{
BarPictograph(x, fill.direction = "From left", data.label.show = FALSE,
  data.label.position = "Next to bar", data.label.format = "",
  data.label.prefix = "", data.label.suffix = "",
  data.label.font.family = global.font.family,
  data.label.font.color = global.font.color, data.label.font.size = 8,
  global.font.family = "Arial", global.font.color = rgb(44, 44, 44,
  maxColorValue = 255), categories.tick.show = TRUE,
  categories.tick.font.family = global.font.family,
  categories.tick.font.color = global.font.color,
  categories.tick.font.size = 12, categories.tick.align = "right", ...)
}
\arguments{
\item{x}{A \code{\link{list}}, where each vector contains data to be plotted. If the data is not numeric
it is coerced to be numeric.
Input data may be a matrix or a vector, containing the height of the columns
to be plotted, with the name/rownames used as the column names of the chart. Numeric and date labels
will be parsed automatically.}

\item{fill.direction}{Direction in which icons are filled. One of \code{"From left", "From right", "From top", "From bottom"}.}

\item{data.label.show}{Logical; whether to show data labels.}

\item{data.label.position}{When \code{show.label.data}, the position of the data labels can be one of \code{"Above icons", "Below icons"} (all modes) or \code{"Next to bar", "Above row label", "Below row label"} (bar mode only). Note that the last two options will overrride \code{sublabel.left} and \code{sublabel.right}}

\item{data.label.format}{A string representing a d3 formatting code.}

\item{data.label.prefix}{Character; prefix for data values.}

\item{data.label.suffix}{Character; suffix for data values.}

\item{data.label.font.family}{Character; font family for data label.}

\item{data.label.font.color}{Font color as a named color
in character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{data.label.font.size}{Font size for data label.px.}

\item{global.font.family}{Character; font family for all occurrences of any
font attribute for the chart unless specified individually.}

\item{global.font.color}{Global font color as a named color in character format
(e.g. "black") or an rgb value (e.g. #' rgb(0, 0, 0, maxColorValue = 255)).}

\item{categories.tick.show}{Logical; Whether to show tick labels for each bar}

\item{categories.tick.font.family}{Character; x-axis tick label font family}

\item{categories.tick.font.color}{X-axis tick label font color as a named color in
character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{categories.tick.font.size}{x-axis tick label font size}

\item{categories.tick.align}{Horizontal alignment of tick labels for each bar. One of "left", "right", "middle".}

\item{...}{Other arguments passed to PictographChart.}
}
\value{
A \code{plotly} chart.
}
\description{
Create bar chart composed of icons
}
