% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemap.R
\name{BaseMap}
\alias{BaseMap}
\title{Base mapping function}
\usage{
BaseMap(table, coords, treat.NA.as.0 = FALSE,
  only.show.regions.in.table = FALSE, add.detail = FALSE,
  remove.last.column = FALSE, remove.last.row = FALSE,
  colors = c("#CCF3FF", "#23B0DB"), ocean.color = "#DDDDDD",
  color.NA = "#808080", legend.show = TRUE, legend.title = "",
  values.hovertext.format = "", remove.regions = NULL,
  unmatched.regions.is.error = TRUE, name.map = NULL,
  mapping.package = "leaflet", high.resolution = FALSE,
  map.type = "countries")
}
\arguments{
\item{table}{A matrix, two-dimensional array, table or vector, containing the
data to be plotted. The \code{\link{rownames}} (or \code{\link{names}} in
the case of a vector) should contain the names of the geographic entities
to be plotted.}

\item{coords}{The coordinates to be mapped.}

\item{treat.NA.as.0}{Plots any \code{NA} values in the data and any
geographical entities without data as having a zero value.}

\item{only.show.regions.in.table}{When TRUE, only geographic entities that
are included in the table are shown on the map.}

\item{add.detail}{Display names of geographical entities on the map. When
TRUE, it also changes the appearance of the map, making the map wrap
around. The only way to prevent this is to resize the map.}

\item{remove.last.column}{Deletes the last column of the table prior to
creating the map, unless the table is a vector or only has one column.}

\item{remove.last.row}{Deletes the bottom row of the table prior to creating
the map, unless the table is a vector or only has one row.}

\item{colors}{A vector of two colors, which are used as endpoints in
interpolating colors.}

\item{ocean.color}{The color used for oceans, used only by \code{plotly}.}

\item{color.NA}{The color used to represent missing values. Not used when
\code{treat.NA.as.0}, is set to missing.}

\item{legend.show}{Logical; Whether to display a legend with the color scale.}

\item{legend.title}{The text to appear above the legend.}

\item{values.hovertext.format}{A string representing a d3 formatting code.
See https://github.com/d3/d3/blob/master/API.md#number-formats-d3-format}

\item{remove.regions}{The regions to remove, even if they are in the table.}

\item{unmatched.regions.is.error}{If there are regions in \code{table} that
are not found in \code{coords}, if this is \code{TRUE} it will cause an
error, otherwise just print a message.}

\item{name.map}{A mapping between incorrect and correct names, useful for
automatically fixing names that a commonly misspecified. Should be a list
where the keys are the correct names and the values are vectors of
incorrect names that should be changed.}

\item{mapping.package}{Either \code{"leaflet"} (better graphics, more country
maps) or \code{"plotly"} (faster).}

\item{high.resolution}{Specifically request a high resolution map. Otherwise
the resolution of the map is chosen automatically based on the resolution required
for the requested countries or regions.}

\item{map.type}{One of \code{"continents"}, \code{"countries"}, \code{"regions"} or
the name of a country, which respectively plot a world map by continent, a world
map by country, a map of USA by region, or a single country map by state.}
}
\description{
Creates a map with a table as an input, using shading to represent the values
of regions on the map.
}
\details{
This function is based on the \code{leaflet} package. See
  \url{https://rstudio.github.io/leaflet/} for an overview of this package
  and how to use it without using these functions.
}
