% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{Box}
\alias{Box}
\title{Box}
\usage{
Box(
  x,
  weights = NULL,
  vertical = TRUE,
  show.values = FALSE,
  density.color = "#008000",
  values.color = "#008000",
  box.points = "Suspected outliers",
  global.font.family = "Arial",
  global.font.color = rgb(44, 44, 44, maxColorValue = 255),
  title = "",
  title.font.family = global.font.family,
  title.font.color = global.font.color,
  title.font.size = 16,
  subtitle = "",
  subtitle.font.family = global.font.family,
  subtitle.font.color = global.font.color,
  subtitle.font.size = 12,
  footer = "",
  footer.font.family = global.font.family,
  footer.font.color = global.font.color,
  footer.font.size = 8,
  footer.wrap = TRUE,
  footer.wrap.nchar = 100,
  background.fill.color = "transparent",
  background.fill.opacity = 1,
  charting.area.fill.color = background.fill.color,
  charting.area.fill.opacity = 0,
  margin.top = NULL,
  margin.bottom = NULL,
  margin.left = NULL,
  margin.right = NULL,
  grid.show = FALSE,
  values.title = "",
  values.title.font.color = global.font.color,
  values.title.font.family = global.font.family,
  values.title.font.size = 12,
  values.line.width = 0,
  values.line.color = rgb(0, 0, 0, maxColorValue = 255),
  values.tick.mark.length = 5,
  values.bounds.minimum = NULL,
  values.bounds.maximum = NULL,
  values.tick.distance = NULL,
  values.zero.line.width = 0,
  values.zero.line.color = rgb(44, 44, 44, maxColorValue = 255),
  values.grid.width = 1 * grid.show,
  values.grid.color = rgb(225, 225, 225, maxColorValue = 255),
  values.tick.show = TRUE,
  values.tick.suffix = "",
  values.tick.prefix = "",
  values.tick.format = "",
  values.hovertext.format = "",
  values.tick.angle = NULL,
  values.tick.font.color = global.font.color,
  values.tick.font.family = global.font.family,
  values.tick.font.size = 10,
  categories.tick.font.color = global.font.color,
  categories.tick.font.family = global.font.family,
  categories.tick.font.size = 10,
  categories.tick.label.wrap = TRUE,
  categories.tick.label.wrap.nchar = 21,
  hovertext.font.family = global.font.family,
  hovertext.font.size = 11,
  hover.on = c("all", "points")[1],
  modebar.show = FALSE
)
}
\arguments{
\item{x}{A \code{\link{list}}, where each vector contains data to be plotted. If the data is not numeric
it is coerced to be numeric.
Input data may be a matrix or a vector, containing the height of the columns
to be plotted, with the name/rownames used as the column names of the chart. Numeric and date labels
will be parsed automatically.}

\item{weights}{An optional \code{\link{list}}, where each element is a vector containing weights corresponding to
the values of \code{x}, or, a vector where the weights is assumed applicable for each element in \code{x}.}

\item{vertical}{Display the densities vertically.}

\item{show.values}{Produces a rug plot of individual values.}

\item{density.color}{Defaults to "Green"}

\item{values.color}{Defaults to "Green"}

\item{box.points}{How outliers are displayed boxplots. \code{"All"} plots all the points. \code{"Suspected outliers"} plots points
between 1.5 and 3 IQR from the 1st and 3rd quartile with un-filled circles. \code{"Outliers"} does not plot points between 1.5 and 3 IQR from the 1st and 3rd quartiles.}

\item{global.font.family}{Character; font family for all occurrences of any
font attribute for the chart unless specified individually.}

\item{global.font.color}{Global font color as a named color in character format
(e.g. "black") or an rgb value (e.g. #' rgb(0, 0, 0, maxColorValue = 255)).}

\item{title}{Character; chart title.}

\item{title.font.family}{Character; title font family. Can be "Arial Black",
"Arial", "Comic Sans MS", "Courier New", "Georgia", "Impact",
"Lucida Console", "Lucida Sans Unicode", "Marlett", "Symbol", "Tahoma",
"Times New Roman", "Trebuchet MS", "Verdana", "Webdings"}

\item{title.font.color}{Title font color as a named color in character
format (e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}

\item{title.font.size}{Title font size; default = 10.}

\item{subtitle}{Character}

\item{subtitle.font.family}{Character; subtitle font family}

\item{subtitle.font.color}{subtitle font color as a named color in
character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{subtitle.font.size}{Integer; subtitle font size}

\item{footer}{Character}

\item{footer.font.family}{Character; footer font family}

\item{footer.font.color}{footer font color as a named color in
character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{footer.font.size}{Integer; footer font size}

\item{footer.wrap}{Logical; whether the footer text should be wrapped.}

\item{footer.wrap.nchar}{Number of characters (approximately) in each line of the footer when \code{footer.wordwrap} \code{TRUE}.}

\item{background.fill.color}{Background color in character format
(e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}

\item{background.fill.opacity}{Background opacity as an alpha value
(0 to 1).}

\item{charting.area.fill.color}{Charting area background color as
a named color in character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{charting.area.fill.opacity}{Charting area background
opacity as an alpha value (0 to 1).}

\item{margin.top}{Margin between plot area and the top of the
graphic in pixels}

\item{margin.bottom}{Margin between plot area and the bottom of the
graphic in pixels}

\item{margin.left}{Margin between plot area and the left of the
graphic in pixels}

\item{margin.right}{Margin between plot area and the right of the
graphic in pixels}

\item{grid.show}{Logical; whether to show grid lines.}

\item{values.title}{Character, y-axis title; defaults to chart input values;
to turn off set to "FALSE".}

\item{values.title.font.color}{y-axis title font color as a named color in
character format (e.g. "black") or an rgb value (e.g. rgb(0, 0, 0,
max = 255)).}

\item{values.title.font.family}{Character; y-axis title font family}

\item{values.title.font.size}{y-axis title font size}

\item{values.line.width}{y-axis line in pixels, 0 = no line}

\item{values.line.color}{y-axis line color as a named color in character format
(e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}

\item{values.tick.mark.length}{Length of tick marks in pixels.}

\item{values.bounds.minimum}{Minimum of range for plotting;
NULL = no manual range set.  Must be less than values.bounds.maximum}

\item{values.bounds.maximum}{Maximum of range for
plotting; NULL = no manual range set.  Must be greater than values.bounds.minimum}

\item{values.tick.distance}{The distance between the ticks. Requires that \code{values.bounds.minimum} and \code{values.bounds.maximum} have been set.}

\item{values.zero.line.width}{Width in pixels of zero line; 0 = no zero line
shown}

\item{values.zero.line.color}{Color of horizontal zero line as a named
color in character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{values.grid.width}{Width of y-grid lines in pixels; 0 = no line}

\item{values.grid.color}{Color of y-grid lines as a named color in character
format (e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}

\item{values.tick.show}{Whether to display the y-axis tick labels}

\item{values.tick.suffix}{y-axis tick label suffix}

\item{values.tick.prefix}{y-axis tick label prefix}

\item{values.tick.format}{d3 formatting string applied to the tick labels.
See https://github.com/mbostock/d3/wiki/Formatting#numbers}

\item{values.hovertext.format}{d3 formatting string applied to the hover text.
https://github.com/mbostock/d3/wiki/Formatting#numbers or}

\item{values.tick.angle}{y-axis tick label angle in degrees.
90 = vertical; 0 = horizontal}

\item{values.tick.font.color}{y-axis tick label font color as a named color
in character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{values.tick.font.family}{Character; y-axis tick label font family}

\item{values.tick.font.size}{y-axis tick label font size}

\item{categories.tick.font.color}{X-axis tick label font color as a named color in
character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{categories.tick.font.family}{Character; x-axis tick label font family}

\item{categories.tick.font.size}{x-axis tick label font size}

\item{categories.tick.label.wrap}{Logical; whether to wrap long labels on the x-axis.}

\item{categories.tick.label.wrap.nchar}{Integer; number of characters in each line when \code{categories.tick.label.wrap} is \code{TRUE}.}

\item{hovertext.font.family}{Font family of hover text.}

\item{hovertext.font.size}{Font size of hover text.}

\item{hover.on}{Only used for boxplot. If "all", then all the hovertext (for median,
for the element near the cursor will be shown.}

\item{modebar.show}{Logical; whether to show the zoom menu buttons or not.}
}
\value{
A \code{plotly} chart.
}
\description{
Box plot.
}
\examples{
Box(rnorm(100))
Box(list(rnorm(100), rexp(100)))
}
