% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{character.matrix}
\alias{character.matrix}
\title{Character matrix for unit testing}
\format{A character matrix}
\usage{
character.matrix
}
\description{
Character matrix for unit testing
}
\keyword{datasets}

