% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartColors.R
\name{ChartColors}
\alias{ChartColors}
\title{Generates a vector of colors for the number of rows in the passed-in
df or matrix chart.matrix; where colors can be either a single named
color or a hex color, any code that generates a vector of named colors
or a vector of hex colors (but not a mixed vector), or a single named
color palette from either of the packages grDevices, RColorBrewer,
colorspace, or colorRamps.  If a single color is provided, and more are
needed, then a gradient will be calculated towards white. If more than
one but fewer than needed are provided, a gradient will be calculated
between the provided colors.}
\usage{
ChartColors(number.colors.needed = NULL, given.colors = qColors,
  reverse = FALSE)
}
\arguments{
\item{number.colors.needed}{Integer; number of colors to generate.}

\item{given.colors}{Character; a vector containing one or more named
colors from grDevices OR one or more specified hex value colors OR a single
named palette from grDevices, RColorBrewer, colorspace, or colorRamps.}

\item{reverse}{Logical; if the output color vector shour be reversed.}
}
\description{
Alpha channels are ignored and set to 255/FF.
}
\details{
\code{ChartColors} generates a vector of colors.
}
\examples{
ChartColors(number.colors.needed = 5, given.colors = c("blue", "orange",
"green"))
ChartColors(number.colors.needed = 5, given.colors = "blue")
ChartColors(number.colors.needed = 5, given.colors = "#9CFF73")
ChartColors(number.colors.needed = 5, given.colors = "Set3", reverse = TRUE)
}

