% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartutils.R
\name{charToNumeric}
\alias{charToNumeric}
\title{Convert text to numeric data}
\usage{
charToNumeric(x)
}
\arguments{
\item{x}{The text to be converted to numbers}
}
\description{
This differs from as.numeric in that it returns NULL
instead of NA if there is no valid output.
Also performs basic substitution such as removing spaces and commas.
}
