% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinedscatterplot.R
\name{CombinedScatter}
\alias{CombinedScatter}
\title{CombinedScatter}
\usage{
CombinedScatter(
  x = NULL,
  y = NULL,
  scatter.x.column = 1,
  scatter.y.column = 2,
  scatter.labels = NULL,
  scatter.labels.name = NULL,
  scatter.sizes = NULL,
  scatter.sizes.name = NULL,
  scatter.sizes.column = 3,
  scatter.sizes.as.diameter = FALSE,
  scatter.colors = NULL,
  scatter.colors.name = NULL,
  scatter.colors.column = 4,
  scatter.colors.as.categorical = TRUE,
  scatter.labels.as.hovertext = TRUE,
  scatter.max.labels = 50,
  annotation.list = NULL,
  colors = ChartColors(12),
  trend.lines = FALSE,
  logos = NULL,
  logo.size = 0.5,
  fit.type = "None",
  fit.window.size = 3,
  fit.ignore.last = FALSE,
  fit.line.type = "dot",
  fit.line.width = 1,
  fit.line.colors = colors,
  fit.line.opacity = 1,
  fit.CI.show = FALSE,
  fit.CI.colors = fit.line.colors,
  fit.CI.opacity = 0.4,
  legend.show = TRUE,
  legend.orientation = "Vertical",
  legend.wrap = TRUE,
  legend.wrap.nchar = 30,
  global.font.family = "Arial",
  global.font.color = rgb(44, 44, 44, maxColorValue = 255),
  title = "",
  title.font.family = global.font.family,
  title.font.color = global.font.color,
  title.font.size = 16,
  subtitle = "",
  subtitle.font.family = global.font.family,
  subtitle.font.color = global.font.color,
  subtitle.font.size = 12,
  footer = "",
  footer.font.family = global.font.family,
  footer.font.color = global.font.color,
  footer.font.size = 8,
  footer.wrap = TRUE,
  footer.wrap.nchar = 100,
  data.label.font.family = global.font.family,
  data.label.font.color = global.font.color,
  data.label.font.autocolor = FALSE,
  data.label.font.size = 10,
  data.label.format = "",
  data.label.prefix = "",
  data.label.suffix = "",
  opacity = NULL,
  background.fill.color = "transparent",
  charting.area.fill.color = background.fill.color,
  legend.font.color = global.font.color,
  legend.font.family = global.font.family,
  legend.font.size = 10,
  legend.position.y = 1,
  legend.position.x = 1.02,
  margin.autoexpand = TRUE,
  margin.top = NULL,
  margin.bottom = NULL,
  margin.left = NULL,
  margin.right = NULL,
  grid.show = TRUE,
  y.title = "",
  y.title.font.color = global.font.color,
  y.title.font.family = global.font.family,
  y.title.font.size = 12,
  y.tick.mark.length = 0,
  y.tick.mark.color = "transparent",
  y.bounds.minimum = NULL,
  y.bounds.maximum = NULL,
  y.tick.distance = NULL,
  y.tick.maxnum = NULL,
  y.zero.line.width = 0,
  y.zero.line.color = rgb(225, 225, 225, maxColorValue = 255),
  y.grid.width = 1 * grid.show,
  y.grid.color = rgb(225, 225, 225, maxColorValue = 255),
  y.tick.show = TRUE,
  y.tick.suffix = "",
  y.tick.prefix = "",
  y.tick.format = "",
  y.hovertext.format = "",
  y.tick.font.color = global.font.color,
  y.tick.font.family = global.font.family,
  y.tick.font.size = 10,
  x.title = "",
  x.title.font.color = global.font.color,
  x.title.font.family = global.font.family,
  x.title.font.size = 12,
  x.line.width = 0,
  x.line.color = rgb(0, 0, 0, maxColorValue = 255),
  x.tick.mark.length = 3,
  x.tick.mark.color = "transparent",
  x.bounds.minimum = NULL,
  x.bounds.maximum = NULL,
  x.tick.distance = NULL,
  x.tick.maxnum = NULL,
  x.zero.line.width = 0,
  x.zero.line.color = rgb(225, 225, 225, maxColorValue = 255),
  x.grid.width = 1 * grid.show,
  x.grid.color = rgb(225, 225, 225, maxColorValue = 255),
  x.tick.show = TRUE,
  x.tick.suffix = "",
  x.tick.prefix = "",
  x.tick.format = "",
  x.hovertext.format = "",
  x.tick.angle = NULL,
  x.tick.font.color = global.font.color,
  x.tick.font.family = global.font.family,
  x.tick.font.size = 10,
  x.tick.label.wrap = TRUE,
  x.tick.label.wrap.nchar = 21,
  hovertext.font.family = global.font.family,
  hovertext.font.size = 11,
  marker.size = 6,
  swap.x.and.y = FALSE,
  legend.bubbles.show = TRUE,
  label.auto.placement = TRUE
)
}
\arguments{
\item{x}{A numeric vector for the x-axis coordinates (which may be named); or a matrix or datarame; or a list of matrices where each matrix share the same row and column names.}

\item{y}{Optional numeric vector for the y-axis coordinates. Should contain the same number of observations as x. If not provided, will use x instead.}

\item{scatter.x.column}{When \code{x} is a dataframe or matrix, the index of the column (1-based) which contains the x-coordinate data.}

\item{scatter.y.column}{When \code{x} is a dataframe or matrix, the index of the column (1-based) which contains the y-coordinate data.}

\item{scatter.labels}{Optional vector for labelling scatter points. This should be the same length as the number of observations in x and y.}

\item{scatter.labels.name}{Character; Used for labelling subtitles and footers.}

\item{scatter.sizes}{Numeric vector determining of the size of each observation. These can alternatively be provided as a column in \code{x}.}

\item{scatter.sizes.name}{Character; Used for labelling footers and legends.}

\item{scatter.sizes.column}{When \code{x} is a dataframe or matrix, the index of the column (1-based) which contains \code{scatter.sizes} data.}

\item{scatter.sizes.as.diameter}{Whether to show the points with diameter (instead of area, which is the default) proportional to the sizes variable.}

\item{scatter.colors}{Numeric, character, or categorical vector determining the color of each observation. These can alternatively be provided as a column in \code{x}.}

\item{scatter.colors.name}{Character; Used for labelling footers.}

\item{scatter.colors.column}{When \code{x} is a dataframe or matrix, the index of the column (1-based) which contains \code{scatter.colors} data.}

\item{scatter.colors.as.categorical}{Boolean; Whether to treat colors as a categorical groups, or a numeric scale.}

\item{scatter.labels.as.hovertext}{Logical; if TRUE, labels are shown has hovers; otherwise, as a labeled scatterplot.}

\item{scatter.max.labels}{Integer; the maximum number of labels to show on a Labeled Scatterplot.
If the number of labels is greater than this parameter, extra labels will be hidden by default but
can be toggled on by clicking on the marker.}

\item{colors}{A vector of colors to use in the chart. When \code{scatter.colors.as.categorical}, the vector of colors should have the length as the number of categories in \code{scatter.colors}. If \code{scatter.colors} is used as numeric vector, then a color ramp is constructed from the colors listed.}

\item{trend.lines}{Boolean indicating whether to plot trend lines for multiple tables.}

\item{logos}{Optional list of images to be used to label scatterplot instead of the row names.
This should be input as a comma-seperated list of URLs.}

\item{logo.size}{Numeric controlling the size of the logos.}

\item{data.label.font.autocolor}{Boolean; If true, \code{data.label.font.color} is ignored and labels are colored
according to the series color.}

\item{opacity}{of scatter point colors as an alpha value (0 to 1).}

\item{marker.size}{Size in pixels of marker. This is overriden
if \code{scatter.sizes} is provided, but used for the legend
if \code{scatter.colors.as.categorical}.}

\item{swap.x.and.y}{Swap the x and y axis around on the chart.}

\item{legend.bubbles.show}{Logical; show legend for bubble sizes.}

\item{label.auto.placement}{Logical; whether the scatter plot labels are positioned automatically
to reduce overlap.}
}
\description{
Scatter plot (uses rhtmlCombinedScatter)
}
