% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartutils.R
\name{commaFromD3}
\alias{commaFromD3}
\title{Whether to format numbers with comma separation of thousands based on a d3 format string.}
\usage{
commaFromD3(format)
}
\arguments{
\item{format}{d3 formatting string}
}
\value{
logical
}
\description{
All chart functions should accept d3 formats. This is used by functions
that do not handle d3 to determine how to format labels and hovertext.
#' @noRd
}
