% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographicmaphelper.R
\name{CountriesOrContinents}
\alias{CountriesOrContinents}
\title{\code{CountriesOrContinents} Names of geographic regions.}
\usage{
CountriesOrContinents(type)
}
\arguments{
\item{type}{The name of the geographic region type or "country". See
\code{\link{GeographicRegionTypes}}}
}
\description{
Returns the list of unique geographic names that can be used when creating a
WorldMap.
}
\examples{
CountriesOrContinents("country")
CountriesOrContinents("continent")

}
