% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartutils.R
\name{decimalsFromD3}
\alias{decimalsFromD3}
\title{Extract the number of decimal places from a d3 format string.}
\usage{
decimalsFromD3(format, default = 0)
}
\arguments{
\item{format}{d3 formatting string}

\item{default}{The number of decimal places if \code{format} is
not provided (usually signifying automatic formatting).}
}
\value{
integer
}
\description{
All chart functions should accept d3 formats. This is used by functions
that do not handle d3 to find the number of decimal places used to
format labels and hovertext.
#' @noRd
}
