% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{errorAsChartMatrix.examples}
\alias{errorAsChartMatrix.examples}
\title{List of ten test examples.}
\format{Ten test scenarios returning errors.}
\usage{
errorAsChartMatrix.examples
}
\description{
List of ten test examples.
}
\keyword{datasets}

