% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartutils.R
\name{ErrorIfNotEnoughData}
\alias{ErrorIfNotEnoughData}
\title{ErrorIfNotEnoughData}
\usage{
ErrorIfNotEnoughData(x, require.tidy = TRUE)
}
\arguments{
\item{x}{The data to be plotted.}

\item{require.tidy}{The data is assumed to be a numeric vector, matrix, array, or data frame.}
}
\description{
Returns an error if there is not enough data for charting.
}
