% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{errorIsChartMatrix.examples}
\alias{errorIsChartMatrix.examples}
\title{List of one test example.}
\format{One test scenario returning error.}
\usage{
errorIsChartMatrix.examples
}
\description{
List of one test example.
}
\keyword{datasets}

