% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex18.1}
\alias{ex18.1}
\title{rhtmlDonut objects created from dput structure out of Q or R for
unit testing}
\format{rhtmlDonut object}
\usage{
ex18.1
}
\description{
rhtmlDonut objects created from dput structure out of Q or R for
unit testing
}
\keyword{datasets}

