% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex27.1}
\alias{ex27.1}
\title{rhtmlLabeledScatter objects created from dput structure out of
Q or R for #' unit testing}
\format{rhtmlLabeledScatter object}
\usage{
ex27.1
}
\description{
rhtmlLabeledScatter objects created from dput structure out of
Q or R for #' unit testing
}
\keyword{datasets}

