% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex34.1}
\alias{ex34.1}
\title{Vector data objects created from dput structure out of
Q or R for #' unit testing}
\format{Plotly object}
\usage{
ex34.1
}
\description{
Vector data objects created from dput structure out of
Q or R for #' unit testing
}
\keyword{datasets}

