% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex38.2}
\alias{ex38.2}
\title{Vector data objects created from dput structure out of
Q or R for #' unit testing}
\format{Plotly object}
\usage{
ex38.2
}
\description{
Vector data objects created from dput structure out of
Q or R for #' unit testing
}
\keyword{datasets}

