% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex6.2}
\alias{ex6.2}
\title{Plotly objects created from dput structure out of Q or R for
unit testing}
\format{Plotly object}
\usage{
ex6.2
}
\description{
Plotly objects created from dput structure out of Q or R for
unit testing
}
\keyword{datasets}

