% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aschartmatrix.R
\name{FilterData}
\alias{FilterData}
\title{Filter y and x by a third (logical) variable.}
\usage{
FilterData(y, x = NULL, filter)
}
\arguments{
\item{y}{A vector, matrix, or data frame.}

\item{x}{A vector}

\item{filter}{A logical (or binary 0/1) vector}
}
\value{
List of filtered y and x objects
}
\description{
\code{FilterData} returns the orignial objects filtered by filter.
}
\examples{
data("z")
FilterData(y = z, filter = sample(0:1, length(z), replace = TRUE))
}

