% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemap.R, R/geographicmaphelper.R
\name{FindCountryFromRegions}
\alias{FindCountryFromRegions}
\alias{FindCountryFromRegions}
\title{Find the name of a country based upon a vector of state names}
\usage{
FindCountryFromRegions(states)

FindCountryFromRegions(states)
}
\arguments{
\item{states}{Character vector of states.}

\item{states}{Character vector of states.}
}
\description{
Find the name of a country based upon a vector of state names

Find the name of a country based upon a vector of state names
}
