% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemap.R, R/geographicmaphelper.R
\name{GeographicRegionRowNames}
\alias{GeographicRegionRowNames}
\alias{GeographicRegionRowNames}
\title{\code{GeographicRegionRowNames} Names of geographic regions.}
\usage{
GeographicRegionRowNames(type)

GeographicRegionRowNames(type)
}
\arguments{
\item{type}{The name of the geographic region type. See
\code{\link{GeographicRegionTypes}}}

\item{type}{The name of the geographic region type. See
\code{\link{GeographicRegionTypes}}}
}
\description{
Returns the list of unique geographic names that can be used when creating a
WorldMap.

Returns the list of unique geographic names that can be used when creating a
WorldMap.
}
\examples{
GeographicRegionRowNames("name")
GeographicRegionRowNames("continent")

GeographicRegionRowNames("name")
GeographicRegionRowNames("continent")

}
