% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{good.examples}
\alias{good.examples}
\title{List of eleven test examples.}
\format{Eleven test scenarios returning T.}
\usage{
good.examples
}
\description{
List of eleven test examples.
}
\keyword{datasets}

