% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat.R
\name{Heat}
\alias{Heat}
\title{Generates a heatmap from a table of data.}
\usage{
Heat(x, sort.rows = "None", sort.columns = "None", colors = c("#0066ff",
  "#ff0000"), standardization = "None", global.font.family = "sans-serif",
  global.font.color = "#000000", title = "",
  title.font.family = global.font.family,
  title.font.color = global.font.color, title.font.size = 14,
  x.title = "", x.title.font.family = global.font.family,
  x.title.font.color = global.font.color, x.title.font.size = 11,
  y.title = "", y.title.font.family = global.font.family,
  y.title.font.color = global.font.color, y.title.font.size = 11,
  x.tick.show = TRUE, x.tick.font.family = global.font.family,
  x.tick.font.color = global.font.color, x.tick.font.size = 11,
  y.tick.show = TRUE, y.tick.font.family = global.font.family,
  y.tick.font.color = global.font.color, y.tick.font.size = 11,
  legend.show = TRUE, legend.font.family = global.font.family,
  legend.font.color = global.font.color, legend.font.size = 11,
  data.label.show = TRUE, data.label.font.size = 11,
  data.label.font.family = global.font.family, data.label.format = "",
  data.label.prefix = "", data.label.suffix = "", left.columns = NULL,
  left.column.headings = "", right.columns = NULL,
  right.column.headings = "")
}
\arguments{
\item{x}{A matrix of data to be displayed.}

\item{sort.rows}{Whether to sort rows by their averages or link as a dendrogram. Options are \code{"None"},
\code{"Sort by averages (ascending)"}, \code{"Sort by averages (descending)"} and \code{"Dendrogram"}.}

\item{sort.columns}{Whether to sort columns by their averages or link as a dendrogram. Options are \code{"None"},
\code{"Sort by averages (ascending)"}, \code{"Sort by averages (descending)"} and \code{"Dendrogram"}.}

\item{colors}{A vector of 2 colors.}

\item{standardization}{Whether to standardize the shading of rows or columns. Options are \code{"None"},
\code{"Standardize rows"} and \code{"Standardize columns"}.}

\item{global.font.family}{= "sans-serif",}

\item{global.font.color}{= "#000000",}

\item{title}{= "",}

\item{title.font.family}{= global.font.family,}

\item{title.font.color}{= global.font.color,}

\item{title.font.size}{= 14,}

\item{x.title}{= "",}

\item{x.title.font.family}{= global.font.family,}

\item{x.title.font.color}{= global.font.color,}

\item{x.title.font.size}{= 11,}

\item{y.title}{= "",}

\item{y.title.font.family}{= global.font.family,}

\item{y.title.font.color}{= global.font.color,}

\item{y.title.font.size}{= 11,}

\item{x.tick.show}{= TRUE,}

\item{x.tick.font.family}{= global.font.family,}

\item{x.tick.font.color}{= global.font.color,}

\item{x.tick.font.size}{= 11,}

\item{y.tick.show}{= TRUE,}

\item{y.tick.font.family}{= global.font.family,}

\item{y.tick.font.color}{= global.font.color,}

\item{y.tick.font.size}{= 11,}

\item{legend.show}{= TRUE,}

\item{legend.font.family}{= global.font.family,}

\item{legend.font.color}{= global.font.color,}

\item{legend.font.size}{= 11,}

\item{data.label.show}{= TRUE,}

\item{data.label.font.size}{= 11,}

\item{data.label.font.family}{= global.font.family,}

\item{data.label.format}{= "",}

\item{data.label.prefix}{= "",}

\item{data.label.suffix}{= "",}

\item{left.columns}{An optional list of vectors or matrices to be appended to the left
of the heatmap.}

\item{left.column.headings}{An optional comma separated string containing headings for
\code{left.columns}. If not supplied, colnames of the items in \code{left.columns} are used.}

\item{right.columns}{An optional list of vectors or matrices to be appended to the right
of the heatmap.}

\item{right.column.headings}{An optional comma separated string containing headings for
\code{right.columns}. If not supplied, colnames of the items in \code{right.columns} are used.}
}
\description{
This function wraps the Heatmap function in the rhtmlHeatmap package.
}
