% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{HeatMap}
\alias{HeatMap}
\title{\code{HeatMap} generates a heatmap from a table of data. This function wraps the Heatmap function
in the rhtmlHeatmap package and performs additional data preparation.}
\usage{
HeatMap(table, data.type = "Use an existing table",
  ignore.rows = "NET, Total, SUM", ignore.columns = "NET, Total, SUM",
  transpose = FALSE, sort.rows = "None", sort.columns = "None",
  color = "Blues", standardization = "None",
  show.cell.values = "Automatic", show.row.labels = "Yes",
  show.column.labels = "Yes")
}
\arguments{
\item{table}{A matrix of data to be displayed.}

\item{data.type}{Whether \code{table} values are numeric or strings.
Options are \code{"Use an existing table"} or \code{"Type or paste data"}.}

\item{ignore.rows}{A string of rows to ignore, delimited by commas.}

\item{ignore.columns}{A string of rows to ignore, delimited by commas.}

\item{transpose}{Whether to transpose the table. \code{TRUE} or \code{FALSE}.}

\item{sort.rows}{Whether to sort rows by their averages or link as a dendrogram. Options are \code{"None"},
\code{"Sort by averages (ascending)"}, \code{"Sort by averages (descending)"} and \code{"Dendrogram"}.}

\item{sort.columns}{Whether to sort columns by their averages or link as a dendrogram. Options are \code{"None"},
\code{"Sort by averages (ascending)"}, \code{"Sort by averages (descending)"} and \code{"Dendrogram"}.}

\item{color}{Options are \code{"Blues"}, \code{"Reds"}, \code{"Greens"}, \code{"Greys"},
\code{"Purples"}, \code{"Oranges"}, \code{"Heat"}, \code{"Blues and reds"} or \code{"Greys and reds"}.}

\item{standardization}{Whether to standardize the shading of rows or columns. Options are \code{"None"},
\code{"Standardize rows"} and \code{"Standardize columns"}.}

\item{show.cell.values}{Whether to show values in cells. Options are \code{"Yes"}, \code{"No"} or
\code{"Automatic"} (<= 20 rows and <= 10 columns).}

\item{show.row.labels}{Whether to label the rows. \code{"Yes"} or \code{"No"}.}

\item{show.column.labels}{Whether to label the columns. \code{"Yes"} or \code{"No"}.}
}
\description{
\code{HeatMap} generates a heatmap from a table of data. This function wraps the Heatmap function
in the rhtmlHeatmap package and performs additional data preparation.
}
