% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aschartmatrix.R
\name{IsChartMatrix}
\alias{IsChartMatrix}
\title{Test if a matrix is appropriate for charting.}
\usage{
IsChartMatrix(x, n.rows, n.columns)
}
\arguments{
\item{x}{A matrix.}

\item{n.rows}{The number of rows in the matrix.}

\item{n.columns}{The number of columns in the matrix.}
}
\value{
Logical; if matrix is a chart matrix.
}
\description{
\code{IsChartMatrix} returns TRUE if the given matrix is suitable for charting, else FALSE.
}
\examples{
data("z")
IsChartMatrix(z, nrow(z), ncol(z))
}

