% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{junk}
\alias{junk}
\title{CSV file of data for junk chart demo}
\format{13 columns by 20 rows; first col alpha, remainder numeric}
\usage{
junk
}
\description{
CSV file of data for junk chart demo
}
\keyword{datasets}

