% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeledscatterplot.R
\name{labeledScatterplot}
\alias{labeledScatterplot}
\title{\code{labeledScatterplot}}
\usage{
labeledScatterplot(chart.matrix, colors = NULL, colors.reverse = FALSE,
  colors.custom.color = NA, colors.custom.gradient.start = NA,
  colors.custom.gradient.end = NA, colors.custom.palette = NA,
  type = "Labeled Scatterplot", group.labels.text = "",
  group.indices.text = "", origin = FALSE, transpose = FALSE,
  rows.to.ignore = "", cols.to.ignore = "", legend.show = TRUE,
  x.title = "", y.title = "")
}
\arguments{
\item{chart.matrix}{Input data with 2 or 3 columns.}

\item{colors}{Group colors.}

\item{colors.reverse}{Whether to reverse the colors.}

\item{colors.custom.color}{Character; a single color which is used if \code{colors} is set to \code{"Custom color"}.}

\item{colors.custom.gradient.start}{Character; starting color of gradient if \code{colors} is set to \code{"Custom gradient"}.}

\item{colors.custom.gradient.end}{Character; last color of gradient if \code{colors} is set to \code{"Custom gradient"}.}

\item{colors.custom.palette}{Character; comma separated list of colors to be used if \code{colors} is set to \code{"Custom palette"}.}

\item{type}{Either "Labeled Scatterplot" or "Labeled Bubbleplot".}

\item{group.labels.text}{Vector or text of comma-separated group labels.}

\item{group.indices.text}{Vector or text of comma-separated group indices corresponding to each row.}

\item{origin}{Whether to display the origin.}

\item{transpose}{Whether to switch the first and second columns before plotting.}

\item{rows.to.ignore}{Text of comma-separated row labels to omit.}

\item{cols.to.ignore}{Text of comma-separated column labels to omit.}

\item{legend.show}{Whether to show the legend, if there is more than one group.}

\item{x.title}{If this is text, this overrides the x-axis title. If this is FALSE, the title is hidden.}

\item{y.title}{If this is text, this overrides the y-axis title. If this is FALSE, the title is hidden.}
}
\description{
\code{labeledScatterplot}
}
