% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{logic.vector.named}
\alias{logic.vector.named}
\title{Named logic vector of five elements}
\format{Named vector containing T, F, T, T, F named A to E.}
\usage{
logic.vector.named
}
\description{
Named logic vector of five elements
}
\keyword{datasets}

