% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{logic.vector}
\alias{logic.vector}
\title{Logic vector of five elements}
\format{Unnamed vector containing T, F, T, T, F}
\usage{
logic.vector
}
\description{
Logic vector of five elements
}
\keyword{datasets}

