% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriescolorgradient.R
\name{MeanRowValueDescendingSort}
\alias{MeanRowValueDescendingSort}
\title{Sorts a matrix in descending order by the mean row value.}
\usage{
MeanRowValueDescendingSort(x)
}
\arguments{
\item{x}{A matrix.}
}
\value{
A sorted matrix.
}
\description{
\code{MeanRowValueDescendingSort} returns a matrix where the rows
have been sorted in descdening order based on the mean value of
each row.
}
\examples{
data("z")
MeanRowValueDescendingSort(z)
}

