% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{named.vector.a}
\alias{named.vector.a}
\title{Named integer vector of three elements}
\format{Named vector containing 1 to 3, named A to C.}
\usage{
named.vector.a
}
\description{
Named integer vector of three elements
}
\keyword{datasets}

