% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{named.vector.b}
\alias{named.vector.b}
\title{Named integer vector of three elements}
\format{Named vector containing 3 to 1, named D to F.}
\usage{
named.vector.b
}
\description{
Named integer vector of three elements
}
\keyword{datasets}

