% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelcoords.R
\name{ParallelCoordinates}
\alias{ParallelCoordinates}
\title{Parallel coordinates}
\usage{
ParallelCoordinates(x, group = NULL, colors = ChartColors(5,
  "Spectral"), scale.show = TRUE, scale.reverse = TRUE,
  scale.nticks = 10, global.font.family = "Arial",
  global.font.color = rgb(44, 44, 44, maxColorValue = 255),
  label.font.family = global.font.family,
  label.font.color = global.font.color, label.font.size = 12,
  range.font.family = global.font.family,
  range.font.color = global.font.color, range.font.size = 10,
  range.decimals = 0, tick.font.family = global.font.family,
  tick.font.color = global.font.color, tick.font.size = 10,
  background.fill.color = rgb(255, 255, 255, maxColorValue = 255),
  background.fill.opacity = 0, charting.area.fill.color = rgb(255, 255,
  255, maxColorValue = 255), charting.area.fill.opacity = 0,
  margin.top = 40, margin.bottom = 40, margin.left = 50,
  margin.right = 40, font.unit = "px")
}
\arguments{
\item{x}{A dataframe containing multiple variables (as columns)}

\item{group}{An optional variable which is used to color the lines in plot.
It should be a vector with length equal to the number of rows in \code{x}. 
Can be numeric or categorical.}

\item{colors}{A vector of colors for the color in lines in plot. Note that only 6-digit
hex codes are accepted (8-digit hex results in black lines). If no \code{group} 
is provided, then the first color will be used for all the lines. If it is provided,
\code{colors} will be interpolated (linearly) to create a color scalebar}

\item{scale.show}{Logical; whether to show the color scalebar}

\item{scale.reverse}{Logical; whether to reverse the color scalebar}

\item{scale.nticks}{Integer; number of ticks to show on colorscalebar}

\item{global.font.family}{Character; font family for all occurrences of any
font attribute for the chart unless specified individually.}

\item{global.font.color}{Global font color as a named color in character format
(e.g. "black") or an a hex code.}

\item{label.font.family}{Character; label font family.}

\item{label.font.color}{Label font color as a named color in character
format (e.g. "black") or a hex code.}

\item{label.font.size}{Integer; Label font size.}

\item{range.font.family}{Character; range label font family.}

\item{range.font.color}{Range label font color as a named color in character
format (e.g. "black") or a hex code.}

\item{range.font.size}{Integer; Range label font size.}

\item{range.decimals}{The number of decimals to shown in the range labels (for numeric
variables) and the tick labels.}

\item{tick.font.family}{Character; tick label font family.}

\item{tick.font.color}{Tick label font color as a named color in character
format (e.g. "black") or a hex code.}

\item{tick.font.size}{Integer; Tick label font size.}

\item{background.fill.color}{Background color in character format (e.g. "black") or a hex code.}

\item{background.fill.opacity}{Background opacity as an alpha value (0 to 1).}

\item{charting.area.fill.color}{Charting area background color as
a named color in character format (e.g. "black") or a hex code.}

\item{charting.area.fill.opacity}{Charting area background opacity as an alpha value (0 to 1).}

\item{margin.top}{Margin between plot area and the top of the
graphic in pixels}

\item{margin.bottom}{Margin between plot area and the bottom of the
graphic in pixels}

\item{margin.left}{Margin between plot area and the left of the
graphic in pixels}

\item{margin.right}{Margin between plot area and the right of the
graphic in pixels}

\item{font.unit}{One of "px" of "pt". By default all font sizes are specified in terms of
pixels ("px"). But changing this to "pt" will mean that the font sizes will be in terms
points ("pt"), which will be consistent with font sizes in text boxes.}
}
\description{
Create a parallel coordinates plot to show multivariate data
}
