% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartutils.R
\name{percentFromD3}
\alias{percentFromD3}
\title{Whether to format as percentages based on a d3 format string.}
\usage{
percentFromD3(format)
}
\arguments{
\item{format}{d3 formatting string}
}
\value{
logical
}
\description{
All chart functions should accept d3 formats. This is used by functions
that do not handle d3 to determine whether to format labels and hovertext as percentages.
#' @noRd
}
