% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piechart.R
\name{Pie}
\alias{Pie}
\title{Pie}
\usage{
Pie(x, type = "Pie", colors = NULL, pie.subslice.colors = NULL,
  pie.subslice.colors.repeat = TRUE, title = "",
  title.font.family = global.font.family, title.font.size = 16,
  title.font.color = global.font.color, pie.data.threshold = NULL,
  pie.values.order = "initial", data.label.prefix = "",
  data.label.suffix = "", data.label.decimals = 2,
  data.label.font.size = 10, data.label.font.color = global.font.color,
  data.label.font.family = global.font.family, pie.groups.font.size = 10,
  pie.groups.font.color = global.font.color,
  pie.groups.font.family = global.font.family, pie.groups.order = "initial",
  pie.inner.radius = 70, pie.border.color = rgb(255, 255, 255, maxColorValue
  = 255), as.percentages = FALSE, global.font.family = "Arial",
  global.font.color = rgb(44, 44, 44, maxColorValue = 255))
}
\arguments{
\item{x}{Input data in the form of a vector, matrix or dataframe. If a dataframe is given, the first column is expected to contain the label names, the second column contains numeric values and the third (optional) column are the group labels.}

\item{type}{One of "Pie" or "Donut"}

\item{colors}{A vector containing a hex value colors.}

\item{pie.subslice.colors}{A vector containing hex value colors for the outer ring of the pie chart. If not supplied will default to the same colors as the inner ring.}

\item{pie.subslice.colors.repeat}{Logical; if, when a grouped
pie chart is displayed, the colors of the subslices should repeat
by group, or be different throughout; defaults to TRUE.}

\item{title}{Character; chart title.}

\item{title.font.family}{Character; title font family. Can be "Arial Black",
"Arial", "Comic Sans MS", "Courier New", "Georgia", "Impact",
"Lucida Console", "Lucida Sans Unicode", "Marlett", "Symbol", "Tahoma",
"Times New Roman", "Trebuchet MS", "Verdana", "Webdings"}

\item{title.font.size}{Title font size; default = 10.}

\item{title.font.color}{Title font color as a named color in character
format (e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}

\item{pie.data.threshold}{Labels with values smaller than the theshold are not shown.}

\item{pie.values.order}{Order of the labels shown. Can be one of 'descending', 'ascending' or 'initial'.}

\item{data.label.prefix}{Character; prefix for data values.}

\item{data.label.suffix}{Character; suffix for data values.}

\item{data.label.decimals}{Number of decimal places to show in
data labels.}

\item{data.label.font.size}{Font size for data label.}

\item{data.label.font.color}{Font color for data label.}

\item{data.label.font.family}{Character; font family for data label.}

\item{pie.groups.font.size}{Font size for group labels.}

\item{pie.groups.font.color}{Font color as a named color}

\item{pie.groups.font.family}{Character; font family for group labels.}

\item{pie.groups.order}{Order of the groups shown. Can be one of 'descending', 'ascending' or 'initial'.}

\item{pie.inner.radius}{The size of the inner radius of pie and
donut charts as a proportion out of 100. defaults to 70.}

\item{pie.border.color}{A single color for space around pie and between segments.}

\item{as.percentages}{Whether to show percentages in pie and donut
charts instead of original values.}

\item{global.font.family}{Character; font family for all occurrences of any
font attribute for the chart unless specified individually.}

\item{global.font.color}{Global font color as a named color in character format
(e.g. "black") or an rgb value (e.g. #' rgb(0, 0, 0, maxColorValue = 255)).}
}
\description{
Pie chart
}
\examples{
dat <- data.frame(labels=rep(LETTERS[24:26], each=3), vals=abs(rnorm(9)),
     groups=rep(LETTERS[1:3], 3), stringsAsFactors = FALSE)
Pie(dat)
Pie(dat, pie.subslice.colors=rainbow(9), pie.subslice.colors.repeat = FALSE)
}
