% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plotlySymbols}
\alias{plotlySymbols}
\title{Vector of all available Plotly symbols}
\format{All valid Plotly marker symbols in numeric format.}
\usage{
plotlySymbols
}
\description{
Vector of all available Plotly symbols
}
\keyword{datasets}

