% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricesensitivitymeter.R
\name{PriceSensitivityMeter}
\alias{PriceSensitivityMeter}
\title{Plots chart analysing price sensitivity}
\usage{
PriceSensitivityMeter(
  x,
  colors = c("#FF0000", "#FF0000", "#008000", "#008000"),
  line.type = c("dot", "solid", "solid", "dot"),
  line.thickness = c(1, 2, 2, 1),
  global.font.family = "Arial",
  global.font.color = rgb(44, 44, 44, maxColorValue = 255),
  x.title = "Price",
  x.tick.prefix = "$",
  x.hovertext.format = ".2f",
  y.title = "Proportion of respondents",
  y.tick.format = "\%",
  resolution = 0.05,
  intersection.show = TRUE,
  intersection.arrow.color = global.font.color,
  intersection.arrow.size = 1.6,
  intersection.arrow.width = 0.7,
  intersection.arrow.length = 10,
  intersection.arrow.standoff = 3,
  intersection.label.font.family = global.font.family,
  intersection.label.font.color = global.font.color,
  intersection.label.font.size = 10,
  intersection.label.wrap = TRUE,
  intersection.label.wrap.nchar = 21,
  ...
)
}
\arguments{
\item{x}{Input table containing survey responses in 4 columns
At what price would you consider this product/brand to be 
1) Very cheap, 2) Cheap, 3) Expensive, 4) Very expensive.}

\item{colors}{Character; a vector containing one or more colors specified as hex codes.}

\item{line.type}{Character; one of 'solid', 'dot', 'dashed'.}

\item{line.thickness}{Thickness, in pixels, of the series line}

\item{global.font.family}{Character; font family for all occurrences of any
font attribute for the chart unless specified individually.}

\item{global.font.color}{Global font color as a named color in character format
(e.g. "black") or an a hex code.}

\item{x.title}{Character, x-axis title; defaults to chart input values;
to turn off set to "FALSE".}

\item{x.tick.prefix}{x-axis tick label prefix}

\item{x.hovertext.format}{A string representing a d3 formatting code.
See https://github.com/mbostock/d3/wiki/Formatting#numbers}

\item{y.title}{Character, y-axis title; defaults to chart input values;
to turn off set to "FALSE".}

\item{y.tick.format}{A string representing a d3 formatting code.
See https://github.com/d3/d3/blob/master/API.md#number-formats-d3-format}

\item{resolution}{Controls how grid points on the x-axis.}

\item{intersection.show}{Logical; Whether to show labels to the intersection points of the lines.}

\item{intersection.arrow.color}{Color of the arrows to the intersection points.}

\item{intersection.arrow.size}{Size of the arrows to the intersection points.}

\item{intersection.arrow.width}{Width of the arrow heads.}

\item{intersection.arrow.length}{Scaling factor controlling length of arrows.}

\item{intersection.arrow.standoff}{Distance between arrowhead and intersection point.}

\item{intersection.label.font.family}{Character; intersection.label font family}

\item{intersection.label.font.color}{intersection.label font color as a named color in
character format (e.g. "black") or an a hex code.}

\item{intersection.label.font.size}{Integer; intersection label font size}

\item{intersection.label.wrap}{Logical; whether the intersection label text should be wrapped.}

\item{intersection.label.wrap.nchar}{Number of characters (approximately) in each
line of the intersection label when \code{intersection.label.wrap} \code{TRUE}.}

\item{...}{Other charting parameters passed to \link{Line}.}
}
\description{
Line chart
}
\examples{
z <- structure(c(1L, 2L, 3L, 4L, 5L, 2L, 3L, 4L, 5L, 6L),  .Dim = c(5L, 2L),
      .Dimnames = list(c("T", "U", "V", "W", "X"), c("A", "B")))
Line(z)
}
