% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.StandardChart}
\alias{print.StandardChart}
\title{Printing \code{StandardChart} objects}
\usage{
\method{print}{StandardChart}(x, ...)
}
\arguments{
\item{x}{Object of class \code{StandardChart}}

\item{...}{Not used}
}
\description{
All objects of class \code{StandardChart} are
   expected to contain a component called htmlwidget is the
   original htmlwidget (e.g. plotly, leaflet, etc)
}
