% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qColors}
\alias{qColors}
\title{Vector of the 15 standard Q colors}
\format{15 Q colors in rbga format.}
\usage{
qColors
}
\description{
Vector of the 15 standard Q colors
}
\keyword{datasets}

