% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qTab.bad.examples}
\alias{qTab.bad.examples}
\title{Bad examples of tables dput from Q}
\format{List of bad example tables exported using dput from Q}
\usage{
qTab.bad.examples
}
\description{
Bad examples of tables dput from Q
}
\keyword{datasets}

