% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qTab.examples}
\alias{qTab.examples}
\title{Examples of tables dput from Q}
\format{List of example tables exported using dput from Q}
\usage{
qTab.examples
}
\description{
Examples of tables dput from Q
}
\keyword{datasets}

