% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterchart.R
\name{Scatter}
\alias{Scatter}
\title{Scatter}
\usage{
Scatter(
  x = NULL,
  y = NULL,
  scatter.x.column = 1,
  scatter.y.column = 2,
  scatter.labels = NULL,
  scatter.labels.name = NULL,
  scatter.sizes = NULL,
  scatter.sizes.name = NULL,
  scatter.sizes.column = 3,
  scatter.sizes.as.diameter = FALSE,
  scatter.colors = NULL,
  scatter.colors.name = NULL,
  scatter.colors.column = 4,
  scatter.colors.as.categorical = TRUE,
  scatter.labels.as.hovertext = TRUE,
  scatter.max.labels = 50,
  scatter.max.groups = 50,
  annotation.list = NULL,
  colors = ChartColors(12),
  trend.lines = FALSE,
  logos = NULL,
  logo.size = 0.5,
  fit.type = "None",
  fit.window.size = 3,
  fit.ignore.last = FALSE,
  fit.line.type = "dot",
  fit.line.width = 1,
  fit.line.name = "Fitted",
  fit.line.colors = colors,
  fit.line.opacity = 1,
  fit.CI.show = FALSE,
  fit.CI.colors = fit.line.colors,
  fit.CI.opacity = 0.4,
  legend.show = TRUE,
  legend.orientation = "Vertical",
  legend.wrap = TRUE,
  legend.wrap.nchar = 30,
  tooltip.show = TRUE,
  modebar.show = FALSE,
  zoom.enable = TRUE,
  axis.drag.enable = FALSE,
  global.font.family = "Arial",
  global.font.color = rgb(44, 44, 44, maxColorValue = 255),
  title = "",
  title.font.family = global.font.family,
  title.font.color = global.font.color,
  title.font.size = 16,
  title.align = "center",
  subtitle = "",
  subtitle.font.family = global.font.family,
  subtitle.font.color = global.font.color,
  subtitle.font.size = 12,
  subtitle.align = "center",
  footer = "",
  footer.font.family = global.font.family,
  footer.font.color = global.font.color,
  footer.font.size = 8,
  footer.align = "center",
  footer.wrap = TRUE,
  footer.wrap.nchar = 100,
  data.label.show = FALSE,
  data.label.font.family = global.font.family,
  data.label.font.color = global.font.color,
  data.label.font.autocolor = FALSE,
  data.label.font.size = 10,
  data.label.position = "top middle",
  opacity = NULL,
  background.fill.color = "transparent",
  background.fill.opacity = 1,
  charting.area.fill.color = background.fill.color,
  charting.area.fill.opacity = 0,
  legend.fill.color = background.fill.color,
  legend.fill.opacity = 0,
  legend.border.color = rgb(44, 44, 44, maxColorValue = 255),
  legend.border.line.width = 0,
  legend.font.color = global.font.color,
  legend.font.family = global.font.family,
  legend.font.size = 10,
  legend.position.y = 1,
  legend.position.x = 1.02,
  legend.ascending = NA,
  margin.autoexpand = TRUE,
  margin.top = NULL,
  margin.bottom = NULL,
  margin.left = NULL,
  margin.right = NULL,
  margin.inner.pad = NULL,
  grid.show = TRUE,
  y.title = "",
  y.title.font.color = global.font.color,
  y.title.font.family = global.font.family,
  y.title.font.size = 12,
  y.line.width = 0,
  y.line.color = rgb(0, 0, 0, maxColorValue = 255),
  y.tick.mark.length = 0,
  y.tick.mark.color = "transparent",
  y.bounds.minimum = NULL,
  y.bounds.maximum = NULL,
  y.tick.distance = NULL,
  y.tick.maxnum = NULL,
  y.zero = FALSE,
  y.zero.line.width = 0,
  y.zero.line.color = rgb(225, 225, 225, maxColorValue = 255),
  y.data.reversed = FALSE,
  y.grid.width = 1 * grid.show,
  y.grid.color = rgb(225, 225, 225, maxColorValue = 255),
  y.tick.show = TRUE,
  y.tick.on.label = TRUE,
  y.tick.suffix = "",
  y.tick.prefix = "",
  y.tick.format = "",
  y.hovertext.format = "",
  y.tick.angle = NULL,
  y.tick.font.color = global.font.color,
  y.tick.font.family = global.font.family,
  y.tick.font.size = 10,
  x.title = "",
  x.title.font.color = global.font.color,
  x.title.font.family = global.font.family,
  x.title.font.size = 12,
  x.line.width = 0,
  x.line.color = rgb(0, 0, 0, maxColorValue = 255),
  x.tick.mark.length = 3,
  x.tick.mark.color = "transparent",
  x.bounds.minimum = NULL,
  x.bounds.maximum = NULL,
  x.tick.distance = NULL,
  x.tick.maxnum = 11,
  x.zero = FALSE,
  x.zero.line.width = 0,
  x.zero.line.color = rgb(225, 225, 225, maxColorValue = 255),
  x.data.reversed = FALSE,
  x.grid.width = 1 * grid.show,
  x.grid.color = rgb(225, 225, 225, maxColorValue = 255),
  x.tick.show = TRUE,
  x.tick.on.label = TRUE,
  x.tick.suffix = "",
  x.tick.prefix = "",
  x.tick.format = "",
  x.hovertext.format = "",
  x.tick.angle = NULL,
  x.tick.font.color = global.font.color,
  x.tick.font.family = global.font.family,
  x.tick.font.size = 10,
  x.tick.label.wrap = TRUE,
  x.tick.label.wrap.nchar = 21,
  hovertext.font.family = global.font.family,
  hovertext.font.size = 11,
  hovertext.align = "left",
  hovertext.template = NULL,
  line.thickness = 0,
  line.colors = colors,
  marker.border.width = 1,
  marker.border.colors = colors,
  marker.border.opacity = NULL,
  marker.size = 6,
  marker.symbols = "circle",
  swap.x.and.y = FALSE,
  small.mult.index = NULL,
  legend.bubbles.show = TRUE,
  label.auto.placement = TRUE,
  sz.min = NULL,
  sz.max = NULL,
  sz.scale = NULL,
  col.min = NULL,
  col.max = NULL
)
}
\arguments{
\item{x}{A numeric vector for the x-axis coordinates (which may be named); or a matrix or datarame; or a list of matrices where each matrix share the same row and column names.}

\item{scatter.sizes.as.diameter}{Whether to show the points with diameter (instead of area, which is the default) proportional to the sizes variable.}

\item{scatter.labels.as.hovertext}{Logical; if TRUE, labels are shown has hovers; otherwise, as a labeled scatterplot.}

\item{scatter.max.groups}{Maximum number of different color categories in data. The is
equivalent to the number of enteries in the legend. Increasing this too high may cause
performance issues.}

\item{data.label.position}{Character; where to place the source data value in relation
to the marker icon. Can be "top left", "top center", "top right", "middle left", "middle center",
"middle right", "bottom left", "bottom center", "bottom right".}

\item{y.tick.on.label}{Logical; Set to FALSE to move grid lines between categories.}

\item{x.tick.on.label}{Logical; Set to FALSE to move grid lines between categories.}

\item{line.thickness}{Thickness, in pixels, of the series line}

\item{line.colors}{Character; a vector containing one or more named}

\item{marker.border.width}{Width in pixels of border/line around markers; 0 is no line.}

\item{marker.border.colors}{Character; a vector containing one or more colors specified as hex codes.}

\item{marker.border.opacity}{Opacity of border/line around markers as an alpha value (0 to 1).}

\item{marker.size}{Size in pixels of marker. This is overriden
if \code{scatter.sizes} is provided, but used for the legend
if \code{scatter.colors.as.categorical}.}

\item{marker.symbols}{Character; a vector describing the symbol used for each data series.
See \url{https://plotly-r.com/working-with-symbols.html} for a list of symbol names.
Note there is no corresponding parameter for LabeledScatter.}

\item{swap.x.and.y}{Swap the x and y axis around on the chart.}

\item{small.mult.index}{Used by Small Multiples to add prefixes to warnings.}

\item{sz.min}{Parameter to control scaling of scatter.sizes, used by SmallMultiples}

\item{sz.max}{Parameter to control scaling of scatter.sizes, used by SmallMultiples}

\item{sz.scale}{Parameter to control scaling of scatter.sizes. Marker size (in pixels) of
the points with the largest value of \code{scatter.size}. This defaults to 100/6 of \code{marker.size}.}

\item{col.min}{Parameter to control scaling of scatter.colors, used by SmallMultiples}

\item{col.max}{Parameter to control scaling of scatter.colors, used by SmallMultiples}
}
\description{
Scatter plot
}
