% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot-default.R
\name{ScatterPlot}
\alias{ScatterPlot}
\title{Scatterplot chart}
\usage{
ScatterPlot(x, scatter.labels = NULL, show.labels.as.hovertext = FALSE, ...)
}
\arguments{
\item{x}{Input data for charting}

\item{scatter.labels}{Alternative way to provide data labels.}

\item{show.labels.as.hovertext}{Use a normal scatterplot even if labels are provided.}

\item{...}{Extra arguments that are ignored.}
}
\description{
Create a scatter plot or a labeled scatter plot if labels are provided.
}
