% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallmultiples.R
\name{SmallMultiples}
\alias{SmallMultiples}
\title{Plotting data as a small multiples}
\usage{
SmallMultiples(x, chart.type = "Area", nrows = 2, pad.left = 0.01,
  pad.right = 0.01, pad.top = 0.01, pad.bottom = 0.01, x.order = NULL,
  average.show = FALSE, average.color = rgb(80, 80, 80, maxColorValue =
  255), colors = ChartColors(max(1, ncol(x), na.rm = TRUE)),
  global.font.family = "Arial", global.font.color = rgb(44, 44, 44,
  maxColorValue = 255), title = "", title.font.family = global.font.family,
  title.font.color = global.font.color, title.font.size = 16,
  paneltitle.show = TRUE, paneltitle.font.family = global.font.family,
  paneltitle.font.color = global.font.color, paneltitle.font.size = 14,
  paneltitle.wrap = TRUE, paneltitle.wrap.nchar = 20, x.title = "",
  x.title.font.size = 12, y.title = "", y.title.font.size = 12,
  data.label.show = FALSE, grid.show = TRUE, x.tick.show = TRUE,
  legend.show = FALSE, margin.left = NULL, margin.right = NULL,
  margin.top = NULL, margin.bottom = NULL, subtitle = "",
  subtitle.font.family = global.font.family,
  subtitle.font.color = global.font.color, subtitle.font.size = 12,
  footer = "", footer.font.family = global.font.family,
  footer.font.color = global.font.color, footer.font.size = 8,
  footer.wrap = TRUE, footer.wrap.nchar = 100, mapping.package = "plotly",
  ...)
}
\arguments{
\item{x}{Input data as a matrix or dataframe}

\item{chart.type}{Can be one of "Area", "Column", "Bar", "Line", "Radar" or "Geographic Map".}

\item{nrows}{Integer; Number of rows to arrange the charts in}

\item{pad.left}{Numeric in [0,1]; Spacing to the left of chart (between panels)}

\item{pad.right}{Numeric in [0,1]; Spacing to the right chart (between panels)}

\item{pad.top}{Numeric in [0,1]; Spacing above chart (between panels)}

\item{pad.bottom}{Numeric in [0,1]; Spacing below chart (between panels)}

\item{x.order}{A vector containing the list index of the columns in the order which they are to be shown}

\item{average.show}{Logical; whether to show a second series in each panel containing
the data averaged across all series.}

\item{average.color}{The color in which the average series should be displayed}

\item{colors}{Character; a vector containing one or more named
colors from grDevices OR one or more specified hex value colors OR a single
named palette from grDevices, RColorBrewer, colorspace, or colorRamps.}

\item{global.font.family}{Character; font family for all occurrences of any
font attribute for the chart unless specified individually.}

\item{global.font.color}{Global font color as a named color in character format
(e.g. "black") or an rgb value (e.g. #' rgb(0, 0, 0, maxColorValue = 255)).}

\item{title}{Character; chart title.}

\item{title.font.family}{Character; title font family. Can be "Arial Black",
"Arial", "Comic Sans MS", "Courier New", "Georgia", "Impact",
"Lucida Console", "Lucida Sans Unicode", "Marlett", "Symbol", "Tahoma",
"Times New Roman", "Trebuchet MS", "Verdana", "Webdings"}

\item{title.font.color}{Title font color as a named color in character
format (e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}

\item{title.font.size}{Integer; Title font size; default = 10.}

\item{paneltitle.show}{Logical; whether to show a title for each panel.}

\item{paneltitle.font.family}{Font family of panel titles.}

\item{paneltitle.font.color}{Font color of panel titles.}

\item{paneltitle.font.size}{Font size of panel titles.}

\item{paneltitle.wrap}{Logical; whether the panel title should be wrapped.}

\item{paneltitle.wrap.nchar}{Number of characters (approximately) in each line
of the panel title when \code{paneltitle.wordwrap} \code{TRUE}.}

\item{x.title}{Character, x-axis title; defaults to chart input values;
to turn off set to "FALSE".}

\item{x.title.font.size}{Integer; x-axis title font size}

\item{y.title}{Character, y-axis title; defaults to chart input values;
to turn off set to "FALSE".}

\item{y.title.font.size}{Integer; y-axis title font size}

\item{data.label.show}{Logical; whether to show data labels.}

\item{grid.show}{Logical; whether to show grid lines.}

\item{x.tick.show}{Whether to display the x-axis tick labels}

\item{legend.show}{Ignored except for with \code{GeographicMap}.}

\item{margin.left}{Margin between plot area and the left of the
graphic in pixels}

\item{margin.right}{Margin between plot area and the right of the
graphic in pixels}

\item{margin.top}{Margin between plot area and the top of the
graphic in pixels}

\item{margin.bottom}{Margin between plot area and the bottom of the
graphic in pixels}

\item{subtitle}{Character}

\item{subtitle.font.family}{Character; subtitle font family}

\item{subtitle.font.color}{subtitle font color as a named color in
character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{subtitle.font.size}{Integer; subtitle font size}

\item{footer}{Character}

\item{footer.font.family}{Character; footer font family}

\item{footer.font.color}{footer font color as a named color in
character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{footer.font.size}{Integer; footer font size}

\item{footer.wrap}{Logical; whether the footer text should be wrapped.}

\item{footer.wrap.nchar}{Number of characters (approximately) in each line of the footer when \code{footer.wordwrap} \code{TRUE}.}

\item{mapping.package}{Not used.}

\item{...}{Extra arguments passed to the charting function}
}
\description{
A small multiple (sometimes called trellis chart, lattice chart,
   grid chart, or panel chart) is a series of similar graphs or charts using
   the same scale and axes, allowing them to be easily compared.
}
