% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkline.R
\name{Sparkline}
\alias{Sparkline}
\title{Sparkline charts}
\usage{
Sparkline(x, type = c("Area", "Line", "Curve", "Column", "Box")[1],
  fill.color = "red", fill.opacity = 0.5, line.thickness = 3,
  line.color = fill.color, line.opacity = 1, end.points.show = FALSE,
  end.points.symbol = "circle", end.points.size = 20,
  end.points.color = fill.color, end.points.opacity = 1,
  end.labels.position = c("None", "Above", "Below", "Next")[1],
  end.labels.decimals = 2, background.fill.color = rgb(255, 255, 255,
  maxColorValue = 255), background.fill.opacity = 0,
  x.axis.show = FALSE, x.tick.show = x.axis.show,
  y.axis.show = FALSE, y.tick.show = y.axis.show, margin.left = 0,
  margin.right = 0, margin.top = 0, margin.bottom = 0)
}
\arguments{
\item{x}{A numeric input vector to plot. Note that because sparklines is designed to be minimal 
there is no support for vector names, and the x-axis of these charts is always the numeric index.}

\item{type}{The chart output type. One of 'Area', 'Line' (straight lines connecting points), 'Curve',
'Column' or 'Box'.}

\item{fill.color}{Character; Color of chart}

\item{fill.opacity}{Opacity of \code{fill.color} as an alpha value (0 to 1).}

\item{line.thickness}{Thickness of line (in pixels) for Area, Line or Curve chart.}

\item{line.color}{Character; Line color in Area, Line and Curve chart. For a line or curve chart, 
setting either \code{fill.color} or \code{line.color} while retaining default values can be used.
but if both arguments are used, then \code{fill.color} will be ignored in favour of \code{line.color}.}

\item{line.opacity}{Opacity of \code{line.color} as an alpha value (0 to 1).}

\item{end.points.show}{Logical; whether to show markers at either end of the series. This only applies for}

\item{end.points.symbol}{Character; marker symbol used at end points. See \url{https://plot.ly/r/reference/#scatter-marker-symbol} for options.}

\item{end.points.size}{Numeric; marker size in pixels.}

\item{end.points.color}{Character; color of end point markers.}

\item{end.points.opacity}{Opacity of \code{end.points.color} as an alpha value (0 to 1).}

\item{end.labels.position}{Character; Whether to put labels above, below or next to end points.}

\item{end.labels.decimals}{Numeric; The number of decimals to be shown in the end labels.}

\item{background.fill.color}{Background color in character format (e.g. "black") or a hex code.}

\item{background.fill.opacity}{Background opacity as an alpha value (0 to 1).}

\item{x.axis.show}{Logical; Whether to show the x-axis.}

\item{x.tick.show}{Logical; Whether to show ticklabels. This only has an effect is \code{x.tick.show} is true.}

\item{y.axis.show}{Logical; Whether to show the y-axis.}

\item{y.tick.show}{Logical; Whether to show ticklabels. This only has an effect is \code{y.tick.show} is true}

\item{margin.left}{Margin between plot area and the left of the graphic in pixels}

\item{margin.right}{Margin between plot area and the right of the graphic in pixels}

\item{margin.top}{Margin between plot area and the top of the graphic in pixels}

\item{margin.bottom}{Margin between plot area and the bottom of the graphic in pixels}
}
\description{
Create small charts for displaying inline with text
}
\examples{
xx <- rnorm(10)
Sparkline(xx, background.fill.color = "black", background.fill.opacity = 1)
Sparkline(xx, type = "Curve", end.points.show = TRUE, end.labels.position = "Next", 
     end.labels.decimals = 5)
Sparkline(xx, type = "Line", x.axis.show = TRUE, y.axis.show = TRUE, 
     x.tick.show = FALSE, y.tick.show = FALSE)
}
