% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemap.R
\name{StatesInCountry}
\alias{StatesInCountry}
\title{Get the states in a country}
\usage{
StatesInCountry(country)
}
\arguments{
\item{country}{The country to look at}
}
\description{
When mapping sthe states of a country you need to match the state names exactly.
You can use this function to look up the correct names of the states for the
country that you are interested in.
}
\seealso{
\code{\link{GeographicRegionRowNames}}
}
