% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{TimeSeries}
\alias{TimeSeries}
\title{TimeSeries}
\usage{
TimeSeries(x = NULL, range.bars = FALSE, colors = ChartColors(1),
  window.start = NULL, global.font.family = "Arial",
  global.font.color = rgb(44, 44, 44, maxColorValue = 255), title = "",
  title.font.family = global.font.family,
  title.font.color = global.font.color, title.font.size = 16,
  x.title = "", x.title.font.color = global.font.color,
  x.title.font.family = global.font.family, x.title.font.size = 12,
  x.tick.font.color = global.font.color,
  x.tick.font.family = global.font.family, x.tick.font.size = 10,
  y.title = "", y.title.font.color = global.font.color,
  y.title.font.family = global.font.family, y.title.font.size = 12,
  y.tick.font.color = global.font.color,
  y.tick.font.family = global.font.family, y.tick.font.size = 10)
}
\arguments{
\item{x}{Input data may be a matrix or a vector, wth dates as the rownames and data series along the columns.}

\item{range.bars}{Logical; whether the data consists of a single series with low, value, high in the columns, or
multiple series.}

\item{colors}{Character; a named color from grDevices OR a hex value color.}

\item{window.start}{The number of days before the end of the data series to start the range selector window.}

\item{global.font.family}{Character; font family for all occurrences of any
font attribute for the chart unless specified individually.}

\item{global.font.color}{Global font color as a named color in character format
(e.g. "black") or an rgb value (e.g. #' rgb(0, 0, 0, maxColorValue = 255)).}

\item{title}{Character; chart title.}

\item{title.font.family}{Character; title font family.}

\item{title.font.color}{Title font color as a named color in character
format (e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}

\item{title.font.size}{Title font size; default = 16.}

\item{x.title}{Character, x-axis title.}

\item{x.title.font.color}{x-axis title font color as a named color in character format (e.g. "black")
or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}

\item{x.title.font.family}{Character; x-axis title font family.}

\item{x.title.font.size}{x-axis title font size.}

\item{x.tick.font.color}{X-axis tick label font color as a named color in
character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{x.tick.font.family}{Character; x-axis tick label font family}

\item{x.tick.font.size}{x-axis tick label font size}

\item{y.title}{Character, y-axis title.}

\item{y.title.font.color}{y-axis title font color as a named color in character format (e.g. "black")
or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}

\item{y.title.font.family}{Character; y-axis title font family.}

\item{y.title.font.size}{y-axis title font size.}

\item{y.tick.font.color}{y-axis tick label font color as a named color in
character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{y.tick.font.family}{Character; y-axis tick label font family}

\item{y.tick.font.size}{y-axis tick label font size}
}
\description{
Plots interactive time series. Either multiple series may be plotted, or a single series with high and low
range/error bars.
}
