% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{var1}
\alias{var1}
\title{Numeric vector of five integers, 1 to 5.}
\format{Unnamed vector with five integers, 1 to 5.}
\usage{
var1
}
\description{
Numeric vector of five integers, 1 to 5.
}
\keyword{datasets}

