% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{var2}
\alias{var2}
\title{Numeric vector of five integers, 5 to 1.}
\format{Unnamed vector with five integers, 5 to 1.}
\usage{
var2
}
\description{
Numeric vector of five integers, 5 to 1.
}
\keyword{datasets}

