% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{var3}
\alias{var3}
\title{Numeric vector of the number 3, repeated 5 times.}
\format{Unnamed vector of the number 3, repeated 5 times.}
\usage{
var3
}
\description{
Numeric vector of the number 3, repeated 5 times.
}
\keyword{datasets}

