% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{var4}
\alias{var4}
\title{Numeric vector of five integers, ordered 3, 4, 5, 2, 1.}
\format{Unnamed vector with five integers, ordered 3, 4, 5, 2, 1.}
\usage{
var4
}
\description{
Numeric vector of five integers, ordered 3, 4, 5, 2, 1.
}
\keyword{datasets}

