% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{var5}
\alias{var5}
\title{Numeric vector of five integers, ordered 5, 4, 2, 3, 1.}
\format{Unnamed vector with five integers, ordered 5, 4, 2, 3, 1.}
\usage{
var5
}
\description{
Numeric vector of five integers, ordered 5, 4, 2, 3, 1.
}
\keyword{datasets}

