% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{Venn}
\alias{Venn}
\title{\code{Venn}}
\usage{
Venn(x = NULL, weights = NULL, data.label.font.size = 20,
  as.percentages = TRUE, data.label.decimals = 0)
}
\arguments{
\item{x}{A \link{data.frame} \code{logical} (converted to logical using >= 1 if not) or a JSON-like list.}

\item{weights}{An optional vector of weights, or, the name or, the name of a variable in \code{x}. It may not be an expression.}

\item{data.label.font.size}{The font size of the labels. Defaults to 10.}

\item{as.percentages}{Show percentages in the Venn diagram (without the percentage signs). These do not necessarily add up to 100\%. This option only applies when x is a matrix.}

\item{data.label.decimals}{Number of decimal places.}
}
\description{
Venn diagram (also known as Euler diagram). Minimizes sum of squared residuals to find the best overlap. Where no diagram appears,
it is likely that there is a problem with the inputs. This is a wrapper for http://www.buildingwidgets.com/blog/2015/6/5/week-22-d3vennr,
which in turn is a wrapper for http://www.benfrederickson.com/venn-diagrams-with-d3.js/.
}
\examples{
Venn(list(
   list("sets"= list(0), "label"= "Like", "size"= 100),
   list("sets"= list(1), "label"= "Love", "size"= 50),
   list("sets"= list(2), "label"= "Dislike", "size"= 100),
   list("sets"= list(3), "label"= "Hate", "size"= 50),
   list("sets"= list(0, 1), "size"= 50),
   list("sets"= list(0, 2), "size"= 0),
   list("sets"= list(2, 3), "size"= 50)))
}
