% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{venn_tooltip}
\alias{venn_tooltip}
\title{venn_tooltip}
\usage{
venn_tooltip(venn, suffix = "")
}
\arguments{
\item{venn}{The JSON-like list.}

\item{suffix}{A string to append to the hovertext.}
}
\description{
Creating the tooltips
}
