% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall.R
\name{Waterfall}
\alias{Waterfall}
\title{Waterfall}
\usage{
Waterfall(y, cumulative = 1, top = NULL, decimals = 1, prefix = "",
  suffix = "", title = "", y.title = "", colors = c("red", "blue"),
  show.cumulative = TRUE, label.in.column.threshold = 0.025,
  label.offset = 0.025)
}
\arguments{
\item{y}{A vector, or, a matrix.}

\item{cumulative}{If \code{y} is a \code{vector}, this argument is ignored. Otherwise,
it can be a \code{\link{vector}} of length \code{length(y)}, or, an \code{integer} indicating the column
in \code{y} containing the cumulative values. It defaults to \code{1}.}

\item{top}{The position from which the blocks "hang" (i.e., the top of each column). Only used if
\code{y} is a \code{\link{matrix}}. If not supplied, it is assumed to be the same as \code{cumulative}.}

\item{decimals}{The number of decimals to show in value labels. Defaults to 1.}

\item{prefix}{The text to appear before after a value in labels. Defaults to nothing.}

\item{suffix}{The text to appear after a value in labels. Defaults to nothing.}

\item{title}{The title for the chart. Defaults to nothing.}

\item{y.title}{The title for the chart. Defaults to nothing.}

\item{colors}{The colors of the blocks, defaulting to red for negative values and blue for positive values.
Where \code{y} is a \code{\link{vector}}, negative values get the first color and positives the second, unless the length
of \code{colors} is the same as \code{y}, in which case each color represents each category of \code{y} if \code{y} is a \code{vector},
and each column otherwise.
a \code{\link{matrix}}, colors are assigned to each column of \code{y}.}

\item{show.cumulative}{If \code{TRUE}, which is the default, the cumulative labels are shown at the top of each bar.}

\item{label.in.column.threshold}{Where the height of a block as a proportion of the range of values on the y-axis
is less than this value, the size of the block is shown below (rather than within) the columns.
The defaut is 0.025.}

\item{label.offset}{When a label is not plotted in a box (which includes all times when the cumulative value is plotted above),
this parameter controls how far the labels is from the box, and it is intepretated as a proportion of the range of \code{y}.
The defaut is 0.025.
is less than this value, the size of the block is shown below (rather than within) the columns.}
}
\description{
Creates a waterfall chart showing the source(s) of change in sales for one period, relative to the previous period.
}
