% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{x.data}
\alias{x.data}
\title{Character vector of five letters, A to E.}
\format{Unnamed vector with five letters, A to E}
\usage{
x.data
}
\description{
Character vector of five letters, A to E.
}
\keyword{datasets}

