% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{x.dates}
\alias{x.dates}
\title{Numeric vector of five elements with POSIXct formatted dates.}
\format{Numeric vector of five elements with POSIXct formatted dates.}
\usage{
x.dates
}
\description{
Numeric vector of five elements with POSIXct formatted dates.
}
\keyword{datasets}

