% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{y.data}
\alias{y.data}
\title{Numeric vector of five integers, 1 to 5.}
\format{Unnamed vector with five integers, 1 to 5.}
\usage{
y.data
}
\description{
Numeric vector of five integers, 1 to 5.
}
\keyword{datasets}

