% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{z}
\alias{z}
\title{Numeric matrix with 5 rows and 1 column.}
\format{Numeric matrix with 5 rows and 1 column, rows named A to E and
column named "Series 1"}
\usage{
z
}
\description{
Numeric matrix with 5 rows and 1 column.
}
\keyword{datasets}

