% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographicmaphelper.R
\name{ZipcodesInCountry}
\alias{ZipcodesInCountry}
\title{Get the zip/post codes in a country}
\usage{
ZipcodesInCountry(country)
}
\arguments{
\item{country}{Character; name of the country}
}
\description{
When mapping the zip or postcodes of a country you need to match the codes exactly.
You can use this function to look up the full list of codes from USA, UK, or Australia.
}
\seealso{
\code{\link{CountriesOrContinents}}

\code{\link{StatesInCountry}}
}
