% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinedscatterplot.R
\name{CombinedScatter}
\alias{CombinedScatter}
\title{CombinedScatter}
\usage{
CombinedScatter(
  x = NULL,
  y = NULL,
  scatter.x.column = 1,
  scatter.y.column = 2,
  scatter.labels = NULL,
  scatter.labels.name = "",
  scatter.sizes = NULL,
  scatter.sizes.name = "",
  scatter.sizes.column = 3,
  scatter.sizes.as.diameter = FALSE,
  scatter.colors = NULL,
  scatter.colors.name = "",
  scatter.colors.column = 4,
  scatter.colors.as.categorical = TRUE,
  scatter.groups = NULL,
  scatter.groups.column = NULL,
  scatter.labels.as.hovertext = TRUE,
  scatter.max.labels = 50,
  scatter.max.groups = 50,
  annotation.list = NULL,
  colors = ChartColors(12),
  trend.lines = FALSE,
  logos = NULL,
  logo.size = 0.5,
  fit.type = "None",
  fit.window.size = 3,
  fit.ignore.last = FALSE,
  fit.line.type = "dot",
  fit.line.width = 1,
  fit.line.colors = colors,
  fit.line.opacity = 1,
  fit.CI.show = FALSE,
  fit.CI.colors = fit.line.colors,
  fit.CI.opacity = 0.4,
  legend.show = TRUE,
  legend.orientation = "Vertical",
  legend.wrap = TRUE,
  legend.wrap.nchar = 30,
  global.font.family = "Arial",
  global.font.color = rgb(44, 44, 44, maxColorValue = 255),
  title = "",
  title.font.family = global.font.family,
  title.font.color = global.font.color,
  title.font.size = 16,
  subtitle = "",
  subtitle.font.family = global.font.family,
  subtitle.font.color = global.font.color,
  subtitle.font.size = 12,
  panel.title.font.family = global.font.family,
  panel.title.font.color = global.font.color,
  panel.title.font.size = 14,
  nrows = 2,
  share.axes = TRUE,
  x.order = NULL,
  panel.x.gap = 0.2,
  panel.y.gap = 0.3,
  footer = "",
  footer.font.family = global.font.family,
  footer.font.color = global.font.color,
  footer.font.size = 8,
  footer.wrap = TRUE,
  footer.wrap.nchar = 100,
  data.label.font.family = global.font.family,
  data.label.font.color = global.font.color,
  data.label.font.autocolor = FALSE,
  data.label.font.size = 10,
  data.label.format = "",
  data.label.prefix = "",
  data.label.suffix = "",
  opacity = NULL,
  background.fill.color = "transparent",
  charting.area.fill.color = background.fill.color,
  legend.font.color = global.font.color,
  legend.font.family = global.font.family,
  legend.font.size = 10,
  legend.position.y = 1,
  legend.position.x = 1.02,
  legend.title = "",
  legend.title.font.color = global.font.color,
  legend.title.font.family = global.font.family,
  legend.title.font.size = 12,
  legend.bubble.title = "",
  legend.bubble.font.color = global.font.color,
  legend.bubble.font.family = global.font.family,
  legend.bubble.font.size = 10,
  legend.bubble.title.font.color = global.font.color,
  legend.bubble.title.font.family = global.font.family,
  legend.bubble.title.font.size = 12,
  margin.autoexpand = TRUE,
  margin.top = NULL,
  margin.bottom = NULL,
  margin.left = NULL,
  margin.right = NULL,
  grid.show = TRUE,
  y.title = "",
  y.title.font.color = global.font.color,
  y.title.font.family = global.font.family,
  y.title.font.size = 12,
  y.line.width = 0,
  y.line.color = rgb(0, 0, 0, maxColorValue = 255),
  y.tick.mark.length = 0,
  y.tick.mark.color = "transparent",
  y.bounds.minimum = NULL,
  y.bounds.maximum = NULL,
  y.tick.distance = NULL,
  y.tick.maxnum = NULL,
  y.zero.line.width = 0,
  y.zero.line.color = rgb(225, 225, 225, maxColorValue = 255),
  y.zero.line.dash = "Dash",
  y.grid.width = 1 * grid.show,
  y.grid.color = rgb(225, 225, 225, maxColorValue = 255),
  y.grid.dash = "Solid",
  y.tick.show = TRUE,
  y.tick.suffix = "",
  y.tick.prefix = "",
  y.tick.format = "",
  y.hovertext.format = "",
  y.tick.font.color = global.font.color,
  y.tick.font.family = global.font.family,
  y.tick.font.size = 10,
  x.title = "",
  x.title.font.color = global.font.color,
  x.title.font.family = global.font.family,
  x.title.font.size = 12,
  x.line.width = 0,
  x.line.color = rgb(0, 0, 0, maxColorValue = 255),
  x.tick.mark.length = 3,
  x.tick.mark.color = "transparent",
  x.bounds.minimum = NULL,
  x.bounds.maximum = NULL,
  x.tick.distance = NULL,
  x.tick.maxnum = NULL,
  x.zero.line.width = 0,
  x.zero.line.color = rgb(225, 225, 225, maxColorValue = 255),
  x.zero.line.dash = "Dash",
  x.grid.width = 1 * grid.show,
  x.grid.color = rgb(225, 225, 225, maxColorValue = 255),
  x.grid.dash = "Solid",
  x.tick.show = TRUE,
  x.tick.suffix = "",
  x.tick.prefix = "",
  x.tick.format = "",
  x.hovertext.format = "",
  x.tick.angle = NULL,
  x.tick.font.color = global.font.color,
  x.tick.font.family = global.font.family,
  x.tick.font.size = 10,
  x.tick.label.wrap = TRUE,
  x.tick.label.wrap.nchar = 21,
  hovertext.font.family = global.font.family,
  hovertext.font.size = 11,
  marker.size = 6,
  swap.x.and.y = FALSE,
  legend.bubbles.show = NULL,
  color.scale.show = NULL,
  label.auto.placement = TRUE
)
}
\arguments{
\item{x}{A numeric vector for the x-axis coordinates (which may be named); or a matrix or datarame; or a list of matrices where each matrix share the same row and column names.}

\item{y}{Optional numeric vector for the y-axis coordinates. Should contain the same number of observations as x. If not provided, will use x instead.}

\item{scatter.x.column}{When \code{x} is a dataframe or matrix, the index of the column (1-based) which contains the x-coordinate data.}

\item{scatter.y.column}{When \code{x} is a dataframe or matrix, the index of the column (1-based) which contains the y-coordinate data.}

\item{scatter.labels}{Optional vector for labelling scatter points. This should be the same length as the number of observations in x and y.}

\item{scatter.labels.name}{Character; Used for labelling subtitles and footers.}

\item{scatter.sizes}{Numeric vector determining of the size of each observation. These can alternatively be provided as a column in \code{x}.}

\item{scatter.sizes.name}{Character; Used for labelling footers and legends.}

\item{scatter.sizes.column}{When \code{x} is a dataframe or matrix, the index of the column (1-based) which contains \code{scatter.sizes} data.}

\item{scatter.sizes.as.diameter}{Whether to show the points with diameter (instead of area, which is the default) proportional to the sizes variable.}

\item{scatter.colors}{Numeric, character, or categorical vector determining the color of each observation. These can alternatively be provided as a column in \code{x}.}

\item{scatter.colors.name}{Character; Used for labelling footers.}

\item{scatter.colors.column}{When \code{x} is a dataframe or matrix, the index of the column (1-based) which contains \code{scatter.colors} data.}

\item{scatter.colors.as.categorical}{Boolean; Whether to treat colors as a categorical groups, or a numeric scale.}

\item{scatter.groups}{A factor of the same length as \code{x} which is
used to aggregate the data for small multiples.}

\item{scatter.groups.column}{The column of \code{x} which is used to aggregate
the data for small multiples (ignored when \code{scatter.groups} is provided)}

\item{scatter.labels.as.hovertext}{Logical; if TRUE, labels are shown has hovers; otherwise, as a labeled scatterplot.}

\item{scatter.max.labels}{Integer; the maximum number of labels to show on a Labeled Scatterplot.
If the number of labels is greater than this parameter, extra labels will be hidden by default but
can be toggled on by clicking on the marker.}

\item{scatter.max.groups}{Maximum number of different color categories in data. The is
equivalent to the number of enteries in the legend. Increasing this too high may cause
performance issues.}

\item{annotation.list}{Optional list of annotations to modify the data labels.}

\item{colors}{A vector of colors to use in the chart. When \code{scatter.colors.as.categorical}, the vector of colors should have the length as the number of categories in \code{scatter.colors}. If \code{scatter.colors} is used as numeric vector, then a color ramp is constructed from the colors listed.}

\item{trend.lines}{Boolean indicating whether to plot trend lines for multiple tables.}

\item{logos}{Optional list of images to be used to label scatterplot instead of the row names.
This should be input as a comma-seperated list of URLs.}

\item{logo.size}{Numeric controlling the size of the logos.}

\item{fit.type}{Character; type of line of best fit. Can be one of "None", "Linear", "LOESS",
"Friedman's super smoother", "Cubic spline", "Moving average", "Centered moving average".}

\item{fit.window.size}{Integer; Use to determine how the average is computed when \code{fit.type}
"Moving average" or "Centered moving average".}

\item{fit.ignore.last}{Logical; whether to ignore the last data point in the fit.}

\item{fit.line.type}{Character; One of "solid", "dot", "dash, "dotdash", or length of dash "2px", "5px".}

\item{fit.line.width}{Numeric; Line width of line of best fit.}

\item{fit.line.colors}{Character; a vector containing one or more colors specified as hex codes.}

\item{fit.line.opacity}{Opacity of trend line as an alpha value (0 to 1).}

\item{fit.CI.show}{Show 95\% confidence interval.}

\item{fit.CI.colors}{Character; a vector containing one or more colors specified as hex codes.}

\item{fit.CI.opacity}{Opacity of confidence interval ribbon as an alpha value (0 to 1).}

\item{legend.show}{is the toggle to show the legend. Can be logical or "Automatic", "Show" or "Hide".
When automatic, the legend is only shown when there is more than one group. Defaults to TRUE.
When FALSE or "Hide", the colorscale and bubble legends are also hidden
(if not overridden by their own "show" parameters).}

\item{legend.orientation}{Character; One of 'Vertical' or 'Horizontal'}

\item{legend.wrap}{Logical; whether the legend text should be wrapped.}

\item{legend.wrap.nchar}{Number of characters (approximately) in each
line of the legend when \code{legend.wrap} \code{TRUE}.}

\item{global.font.family}{Character; font family for all occurrences of any
font attribute for the chart unless specified individually.}

\item{global.font.color}{Global font color as a named color in character format
(e.g. "black") or an a hex code.}

\item{title}{Character; chart title.}

\item{title.font.family}{Character; title font family. Can be "Arial Black",
"Arial", "Comic Sans MS", "Courier New", "Georgia", "Impact",
"Lucida Console", "Lucida Sans Unicode", "Marlett", "Symbol", "Tahoma",
"Times New Roman", "Trebuchet MS", "Verdana", "Webdings"}

\item{title.font.color}{Title font color as a named color in character
format (e.g. "black") or a hex code.}

\item{title.font.size}{Integer; Title font size; default = 10.}

\item{subtitle}{Character}

\item{subtitle.font.family}{Character; subtitle font family}

\item{subtitle.font.color}{subtitle font color as a named color in
character format (e.g. "black") or an a hex code.}

\item{subtitle.font.size}{Integer; subtitle font size}

\item{panel.title.font.family}{Font family of panel titles.}

\item{panel.title.font.color}{Font color of panel titles.}

\item{panel.title.font.size}{Font size of panel titles.}

\item{nrows}{Integer; Number of rows to arrange the small multiple panels.}

\item{share.axes}{Force range of the plot to be the same across all panels.}

\item{x.order}{A vector containing the list index of the columns in the order
which they are to be shown, or a string with comma separated indices.}

\item{panel.x.gap}{A number between 0 and 1. Controls the horizontal space between panels.}

\item{panel.y.gap}{A number between 0 and 1. Controls the vertical space between panels.}

\item{footer}{Character}

\item{footer.font.family}{Character; footer font family}

\item{footer.font.color}{footer font color as a named color in
character format (e.g. "black") or an a hex code.}

\item{footer.font.size}{Integer; footer font size}

\item{footer.wrap}{Logical; whether the footer text should be wrapped.}

\item{footer.wrap.nchar}{Number of characters (approximately) in each
line of the footer when \code{footer.wrap} \code{TRUE}.}

\item{data.label.font.family}{Character; font family for data label.}

\item{data.label.font.color}{Font color as a named color
in character format (e.g. "black") or an a hex code. This can be a single
color, a vector of colors (1 for each series/column), or a comma separated list
of colors}

\item{data.label.font.autocolor}{Boolean; If true, \code{data.label.font.color} is ignored and labels are colored
according to the series color.}

\item{data.label.font.size}{Integer; Font size for data label.px.}

\item{data.label.format}{A string representing a d3 formatting code.
See https://github.com/mbostock/d3/wiki/Formatting#numbers}

\item{data.label.prefix}{Character; prefix for data values.}

\item{data.label.suffix}{Character; suffix for data values.}

\item{opacity}{of scatter point colors as an alpha value (0 to 1).}

\item{background.fill.color}{Background color in character format (e.g. "black") or a hex code.}

\item{charting.area.fill.color}{Charting area background color as
a named color in character format (e.g. "black") or a hex code.}

\item{legend.font.color}{Legend font color as a named color in character
format (e.g. "black") or a hex code.}

\item{legend.font.family}{Character; legend font family.}

\item{legend.font.size}{Integer; Legend font size.}

\item{legend.position.y}{A numeric controlling the position of the legend.
Values range from 0 (bottom) to 1 (top).}

\item{legend.position.x}{A numeric controlling the position of the legend.
Values range from -0.5 (left) to 1.5 (right).}

\item{legend.title}{Title to show above the legend (and color scale bar)}

\item{legend.title.font.color}{Font color of the legend (and color scale bar) title}

\item{legend.title.font.family}{Font family of the legend (and color scale bar) title}

\item{legend.title.font.size}{Font size of the legend (and color scale bar) title}

\item{legend.bubble.title}{Title to show above the bubble legend}

\item{legend.bubble.font.color}{Font color of the bubble legend}

\item{legend.bubble.font.family}{Font family of the bubble legend}

\item{legend.bubble.font.size}{Font size of the bubble legend}

\item{legend.bubble.title.font.color}{Font color of the bubble legend title}

\item{legend.bubble.title.font.family}{Font family of the bubble legend title}

\item{legend.bubble.title.font.size}{Font size of the bubble legend title}

\item{margin.autoexpand}{Logical; Whether extra space can be added to the margins
to allow space for axis/legend/data labels or other chart elements.}

\item{margin.top}{Margin between plot area and the top of the graphic in pixels}

\item{margin.bottom}{Margin between plot area and the bottom of the graphic in pixels}

\item{margin.left}{Margin between plot area and the left of the graphic in pixels}

\item{margin.right}{Margin between plot area and the right of the graphic in pixels}

\item{grid.show}{Logical; whether to show grid lines.}

\item{y.title}{Character, y-axis title; defaults to chart input values;
to turn off set to "FALSE".}

\item{y.title.font.color}{y-axis title font color as a named color in
character format (e.g. "black") or a hex code.}

\item{y.title.font.family}{Character; y-axis title font family}

\item{y.title.font.size}{Integer; y-axis title font size}

\item{y.line.width}{y-axis line width in pixels (0 = no line).}

\item{y.line.color}{y-axis line color as a named color in character format
(e.g. "black") or a hex code.}

\item{y.tick.mark.length}{Length of tick marks in pixels. Ticks are only shown when \code{y.line.width > 0}.}

\item{y.tick.mark.color}{Color of tick marks (default transparent).}

\item{y.bounds.minimum}{Minimum of range for plotting; For a date axis this should be supplied as a date string.
For a categorical axis, the index of the category (0-based) should be used.}

\item{y.bounds.maximum}{Maximum of range for plotting; NULL = no manual range set.}

\item{y.tick.distance}{Distance between tick marks. Requires that \code{y.bounds.minimum} and \code{y.bounds.maximum} have been set.}

\item{y.tick.maxnum}{Maximum number of ticks shown on the axis.
This setting is ignored if \code{y.tick.distance} is set or
if the axis is categorical}

\item{y.zero.line.width}{Width in pixels of zero line;}

\item{y.zero.line.color}{Color of horizontal zero line as a named
color in character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{y.zero.line.dash}{Line type of y zero line. Can be one of 'Solid', 'Dot', 'Dash'.}

\item{y.grid.width}{Width of y-grid lines in pixels; 0 = no line}

\item{y.grid.color}{Color of y-grid lines as a named color in character
format (e.g. "black") or a hex code.}

\item{y.grid.dash}{Line type of y grid line. Can be one of 'Solid', 'Dot', 'Dash'.}

\item{y.tick.show}{Whether to display the y-axis tick labels}

\item{y.tick.suffix}{y-axis tick label suffix}

\item{y.tick.prefix}{y-axis tick label prefix}

\item{y.tick.format}{A string representing a d3 formatting code.
See https://github.com/d3/d3/blob/master/API.md#number-formats-d3-format}

\item{y.hovertext.format}{A string representing a d3 formatting code
See https://github.com/d3/d3/blob/master/API.md#number-formats-d3-format}

\item{y.tick.font.color}{y-axis tick label font color as a named color
in character format (e.g. "black") or an a hex code.}

\item{y.tick.font.family}{Character; y-axis tick label font family}

\item{y.tick.font.size}{Integer; y-axis tick label font size}

\item{x.title}{Character, x-axis title; defaults to chart input values;
to turn off set to "FALSE".}

\item{x.title.font.color}{x-axis title font color as a named color in
character format (e.g. "black") or an a hex code.}

\item{x.title.font.family}{Character; x-axis title font family}

\item{x.title.font.size}{Integer; x-axis title font size}

\item{x.line.width}{x-axis line in pixels, 0 = no line}

\item{x.line.color}{x-axis line color as a named color in character format
(e.g. "black") or a hex code.}

\item{x.tick.mark.length}{Length of tick marks in pixels.}

\item{x.tick.mark.color}{Color of tick marks.}

\item{x.bounds.minimum}{Minimum of range for plotting; For a date axis this should be supplied as a date string.
For a categorical axis, the index of the category (0-based) should be used.}

\item{x.bounds.maximum}{Maximum of range for
plotting; NULL = no manual range set.  Must be greater than x.bounds.minimum}

\item{x.tick.distance}{Tick mark distance in
x-axis units between minimum and maximum for plotting; NULL = no manual
range set.}

\item{x.tick.maxnum}{Maximum number of ticks shown on the axis.
It defaults to 11 which gives the same output from plotly as NULL.
This setting is ignored if \code{x.tick.distance} is set or
if the axis is categorical}

\item{x.zero.line.width}{Width in pixels of zero line.}

\item{x.zero.line.color}{Color of horizontal zero (origin) line as a named
color in character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{x.zero.line.dash}{Line type of x zero line. Can be one of 'Solid', 'Dot', 'Dash'.}

\item{x.grid.width}{Width of y-grid lines in pixels; 0 = no line}

\item{x.grid.color}{Color of y-grid lines as a named color in character
format (e.g. "black") or a hex code.}

\item{x.grid.dash}{Line type of x grid line. Can be one of 'Solid', 'Dot', 'Dash'.}

\item{x.tick.show}{Whether to display the x-axis tick labels}

\item{x.tick.suffix}{x-axis tick label suffix}

\item{x.tick.prefix}{x-axis tick label prefix}

\item{x.tick.format}{A string representing a d3 formatting code.
See https://github.com/mbostock/d3/wiki/Formatting#numbers}

\item{x.hovertext.format}{A string representing a d3 formatting code.
See https://github.com/mbostock/d3/wiki/Formatting#numbers}

\item{x.tick.angle}{x-axis tick label angle in degrees.
90 = vertical; 0 = horizontal}

\item{x.tick.font.color}{X-axis tick label font color as a named color in
character format (e.g. "black") or an a hex code.}

\item{x.tick.font.family}{Character; x-axis tick label font family}

\item{x.tick.font.size}{Integer; x-axis tick label font size}

\item{x.tick.label.wrap}{Logical; whether to wrap long labels on the x-axis.}

\item{x.tick.label.wrap.nchar}{Integer; number of characters in each line when \code{label.wrap} is \code{TRUE}.}

\item{hovertext.font.family}{Font family of hover text.}

\item{hovertext.font.size}{Font size of hover text.}

\item{marker.size}{Size in pixels of marker. This is overriden
if \code{scatter.sizes} is provided, but used for the legend
if \code{scatter.colors.as.categorical}.}

\item{swap.x.and.y}{Swap the x and y axis around on the chart.}

\item{legend.bubbles.show}{Logical; show legend for bubble sizes.}

\item{color.scale.show}{is the toggle to show the color scale bar.}

\item{label.auto.placement}{Logical; whether the scatter plot labels are positioned automatically
to reduce overlap.}
}
\value{
an HTML widget for \code{"leaflet"} or a \code{"plotly"} object.
}
\description{
Scatter plot (uses rhtmlCombinedScatter)
}
\examples{
z <- structure(c(1L, 2L, 3L, 4L, 5L, 2L, 3L, 4L, 5L, 6L),  .Dim = c(5L, 2L),
      .Dimnames = list(c("T", "U", "V", "W", "X"), c("A", "B")))
Column(z, type="Stacked Column")
}
