% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartutils.R
\name{ErrorIfNotEnoughData}
\alias{ErrorIfNotEnoughData}
\title{ErrorIfNotEnoughData}
\usage{
ErrorIfNotEnoughData(x)
}
\arguments{
\item{x}{The data to be plotted.}
}
\description{
Returns an error if there is not enough data for charting/
}
