% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographicmap.R
\name{GeographicMap}
\alias{GeographicMap}
\title{Plot a Geographic Map}
\usage{
GeographicMap(x, country, high.resolution = FALSE,
  show.missing.regions = TRUE, treat.NA.as.0 = FALSE,
  colors = c("#CCF3FF", "#23B0DB"), ocean.color = "#DDDDDD",
  color.NA = "#808080", global.font.family = "Arial",
  global.font.color = rgb(44, 44, 44, maxColorValue = 255),
  legend.show = TRUE, legend.title = "",
  legend.font.family = global.font.family,
  legend.font.color = global.font.color, legend.font.size = 10,
  values.hovertext.format = "", values.bounds.minimum = NULL,
  values.bounds.maximum = NULL, mapping.package = "leaflet",
  background = FALSE, title = "",
  title.font.family = global.font.family,
  title.font.color = global.font.color, title.font.size = 16,
  subtitle = "", subtitle.font.family = global.font.family,
  subtitle.font.color = global.font.color, subtitle.font.size = 12,
  footer = "", footer.font.family = global.font.family,
  footer.font.color = global.font.color, footer.font.size = 8,
  footer.wrap = TRUE, footer.wrap.nchar = 100,
  zip.country = "Automatic")
}
\arguments{
\item{x}{A matrix, two-dimensional array, table or vector, containing the
data to be plotted. The \code{\link{rownames}} (or \code{\link{names}} in
the case of a vector) should contain the names of the geographic entities
to be plotted.}

\item{country}{Character string optionally stating the country that the states are from, if
\code{map.type} is \code{"states"}.}

\item{high.resolution}{Specifically request a high resolution map. Otherwise
the resolution of the map is chosen automatically based on the resolution required
for the requested countries or regions.}

\item{show.missing.regions}{Logical; Whether to plot regions not included in
\code{x} with values of \code{NA}. Used only by \code{"leaflet"}.}

\item{treat.NA.as.0}{Plots any \code{NA} values in the data and any
geographical entities without data as having a zero value.}

\item{colors}{A vector of two colors, which are used as endpoints in
interpolating colors.}

\item{ocean.color}{The color used for oceans (or background).}

\item{color.NA}{The color used to represent missing values. Not used when
\code{treat.NA.as.0}, is set to missing.}

\item{global.font.family}{Character; font family for all occurrences of any
font attribute for the chart unless specified individually.}

\item{global.font.color}{Global font color as a named color in character format
(e.g. "black") or an rgb value (e.g. #' rgb(0, 0, 0, maxColorValue = 255)).}

\item{legend.show}{Logical; Whether to display a legend with the color scale.}

\item{legend.title}{The text to appear above the legend.}

\item{legend.font.family}{Font family of legend. Only used with \code{plotly} object.}

\item{legend.font.color}{Font color of legend. Only used with \code{plotly} object.}

\item{legend.font.size}{Font size of legend. Only used with \code{plotly} object.}

\item{values.hovertext.format}{A string representing a d3 formatting code.
See https://github.com/d3/d3/blob/master/API.md#number-formats-d3-format}

\item{values.bounds.minimum}{Numeric; specifies the minimum value in the colorscale.
If not specified, this will be automatically determined from the data.
If the value specified is larger than the minimum in the data then
the specified value will be ignored.}

\item{values.bounds.maximum}{Numeric; specifies the maximum value in the colorscale.}

\item{mapping.package}{Either \code{"leaflet"} (better graphics, wider range of
maps) or \code{"plotly"} (faster).}

\item{background}{If \code{"mapping.package"} is \code{"leaflet"}, add a background
tile from opensteetmaps.}

\item{title}{Character; chart title.}

\item{title.font.family}{Character; title font family. Can be "Arial Black",
"Arial", "Comic Sans MS", "Courier New", "Georgia", "Impact",
"Lucida Console", "Lucida Sans Unicode", "Marlett", "Symbol", "Tahoma",
"Times New Roman", "Trebuchet MS", "Verdana", "Webdings"}

\item{title.font.color}{Title font color as a named color in character
format (e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}

\item{title.font.size}{Integer; Title font size; default = 10.}

\item{subtitle}{Character}

\item{subtitle.font.family}{Character; subtitle font family}

\item{subtitle.font.color}{subtitle font color as a named color in
character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{subtitle.font.size}{Integer; subtitle font size}

\item{footer}{Character}

\item{footer.font.family}{Character; footer font family}

\item{footer.font.color}{footer font color as a named color in
character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{footer.font.size}{Integer; footer font size}

\item{footer.wrap}{Logical; whether the footer text should be wrapped.}

\item{footer.wrap.nchar}{Number of characters (approximately) in each 
line of the footer when \code{footer.wrap} \code{TRUE}.}

\item{zip.country}{One of \code{"Automatic"}, \code{"USA"}, \code{"UK"} or \code{"Australia"}.
If \code{"Automatic"} an attempt is made to infer the country from the data.}
}
\value{
an HTML widget for \code{"leaflet"} or a \code{"plotly"} object.
}
\description{
Creates a map with a table as input, using shading to represent the values of
countries or states on the map.
}
\examples{
data <- seq(4)
names(data) <- c("France", "China", "Brazil", "Canada")
GeographicMap(data)
}
