% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographicmap.R
\name{GeographicMap}
\alias{GeographicMap}
\title{Geographic Map}
\usage{
GeographicMap(x, country, high.resolution = FALSE, treat.NA.as.0 = FALSE,
  colors = c("#CCF3FF", "#23B0DB"), ocean.color = "#DDDDDD",
  color.NA = "#808080", legend.show = TRUE, legend.title = "",
  values.hovertext.format = "", mapping.package = "leaflet")
}
\arguments{
\item{x}{A matrix, two-dimensional array, table or vector, containing the
data to be plotted. The \code{\link{rownames}} (or \code{\link{names}} in
the case of a vector) should contain the names of the geographic entities
to be plotted.}

\item{country}{Character string optionally stating the country that the states are from, if
\code{map.type} is \code{"states"}.}

\item{high.resolution}{Specifically request a high resolution map. Otherwise
the resolution of the map is chosen automatically based on the resolution required
for the requested countries or regions.}

\item{treat.NA.as.0}{Plots any \code{NA} values in the data and any
geographical entities without data as having a zero value.}

\item{colors}{A vector of two colors, which are used as endpoints in
interpolating colors.}

\item{ocean.color}{The color used for oceans, used only by \code{plotly}.}

\item{color.NA}{The color used to represent missing values. Not used when
\code{treat.NA.as.0}, is set to missing.}

\item{legend.show}{Logical; Whether to display a legend with the color scale.}

\item{legend.title}{The text to appear above the legend.}

\item{values.hovertext.format}{A string representing a d3 formatting code.
See https://github.com/d3/d3/blob/master/API.md#number-formats-d3-format}

\item{mapping.package}{Either \code{"leaflet"} (better graphics, more country
maps) or \code{"plotly"} (faster).}
}
\description{
Creates a map with a table as input, using shading to represent the values of
countries or states on the map.
}
