% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographicmaphelper.R
\name{GeographicRegionRowNames}
\alias{GeographicRegionRowNames}
\title{\code{GeographicRegionRowNames} Names of geographic regions.}
\usage{
GeographicRegionRowNames(type)
}
\arguments{
\item{type}{The name of the geographic region type. See
\code{\link{GeographicRegionTypes}}}
}
\description{
Deprecated - use \code{\link{CountriesOrContinents}}
}
