% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographicmaphelper.R
\name{GeographicRegionTypes}
\alias{GeographicRegionTypes}
\title{\code{GeographicRegionTypes} Types of Geographic Regions}
\usage{
GeographicRegionTypes()
}
\description{
The geographic region types that are available for referring in a map. E.g.,
\code{name}, \code{continent}
}
\examples{
GeographicRegionTypes()

}
