% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{HeatMap}
\alias{HeatMap}
\title{Generates a heatmap from a table of data.}
\usage{
HeatMap(table, data.type = "Use an existing table",
  ignore.rows = "NET, Total, SUM", ignore.columns = "NET, Total, SUM",
  transpose = FALSE, sort.rows = "None", sort.columns = "None",
  color = "Blues", standardization = "None",
  show.cell.values = "Automatic", cell.decimals = 2,
  show.row.labels = "Yes", show.column.labels = "Yes", show.legend = TRUE,
  chart.title = "", x.axis.title = "", y.axis.title = "",
  font.family = "sans-serif", font.color = "#000000",
  title.font.size = 24, xaxis.title.font.size = 14,
  yaxis.title.font.size = 14, legend.font.size = 11, value.font.size = 11,
  axis.label.font.size = 11, left.columns = NULL,
  left.column.headings = "", right.columns = NULL,
  right.column.headings = "")
}
\arguments{
\item{table}{A matrix of data to be displayed.}

\item{data.type}{Whether \code{table} values are numeric or strings.
Options are \code{"Use an existing table"} or \code{"Type or paste data"}.}

\item{ignore.rows}{A string of rows to ignore, delimited by commas.}

\item{ignore.columns}{A string of rows to ignore, delimited by commas.}

\item{transpose}{Whether to transpose the table. \code{TRUE} or \code{FALSE}.}

\item{sort.rows}{Whether to sort rows by their averages or link as a dendrogram. Options are \code{"None"},
\code{"Sort by averages (ascending)"}, \code{"Sort by averages (descending)"} and \code{"Dendrogram"}.}

\item{sort.columns}{Whether to sort columns by their averages or link as a dendrogram. Options are \code{"None"},
\code{"Sort by averages (ascending)"}, \code{"Sort by averages (descending)"} and \code{"Dendrogram"}.}

\item{color}{Options are \code{"Blues"}, \code{"Reds"}, \code{"Greens"}, \code{"Greys"},
\code{"Purples"}, \code{"Oranges"}, \code{"Heat"}, \code{"Blues and reds"} or \code{"Greys and reds"}.}

\item{standardization}{Whether to standardize the shading of rows or columns. Options are \code{"None"},
\code{"Standardize rows"} and \code{"Standardize columns"}.}

\item{show.cell.values}{Whether to show values in cells. Options are \code{"Yes"}, \code{"No"} or
\code{"Automatic"} (<= 20 rows and <= 10 columns).}

\item{cell.decimals}{The number of decimal points to use for formatting cell values.}

\item{show.row.labels}{Whether to label the rows. \code{"Yes"} or \code{"No"}.}

\item{show.column.labels}{Whether to label the columns. \code{"Yes"} or \code{"No"}.}

\item{show.legend}{Whether to show the legend.}

\item{chart.title}{Title of the chart.}

\item{x.axis.title}{Title of the x-axis.}

\item{y.axis.title}{Title of the y-axis.}

\item{font.family}{Font family to be used for all titles, axes, labels and values.}

\item{font.color}{Font color to be used for all titles, axes and labels.}

\item{title.font.size}{Font size of the title.}

\item{xaxis.title.font.size}{Font size of the x-axis title.}

\item{yaxis.title.font.size}{Font size of the y-axis title.}

\item{legend.font.size}{Font size of the legend.}

\item{value.font.size}{Font size of the cell values and tooltips.}

\item{axis.label.font.size}{Font size of the axis labels.}

\item{left.columns}{An optional list of vectors or matrices to be appended to the left
of the heatmap.}

\item{left.column.headings}{An optional comma separated string containing headings for
\code{left.columns}. If not supplied, colnames of the items in \code{left.columns} are used.}

\item{right.columns}{An optional list of vectors or matrices to be appended to the right
of the heatmap.}

\item{right.column.headings}{An optional comma separated string containing headings for
\code{right.columns}. If not supplied, colnames of the items in \code{right.columns} are used.}
}
\description{
This function wraps the Heatmap function in the rhtmlHeatmap package and performs additional data preparation.
}
