% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeledscatterchart.R
\name{LabeledScatter}
\alias{LabeledScatter}
\title{LabeledScatter}
\usage{
LabeledScatter(x = NULL, y = NULL, scatter.x.column = 1,
  scatter.y.column = 2, scatter.labels = NULL,
  scatter.labels.name = NULL, scatter.sizes = NULL,
  scatter.sizes.name = NULL, scatter.sizes.column = 3,
  scatter.colors = NULL, scatter.colors.name = NULL,
  scatter.colors.column = 4, scatter.colors.as.categorical = TRUE,
  trend.lines = FALSE, logos = NULL, logo.size = 0.5,
  colors = ChartColors(12), opacity = NULL, legend.show = TRUE,
  global.font.family = "Arial", global.font.color = rgb(44, 44, 44,
  maxColorValue = 255), title = "",
  title.font.family = global.font.family,
  title.font.color = global.font.color, title.font.size = 16,
  subtitle = "", subtitle.font.family = global.font.family,
  subtitle.font.color = global.font.color, subtitle.font.size = 12,
  footer = "", footer.font.family = global.font.family,
  footer.font.color = global.font.color, footer.font.size = 8,
  footer.wrap = TRUE, footer.wrap.nchar = 100,
  scatter.max.labels = 50, data.label.font.family = global.font.family,
  data.label.font.color = global.font.color,
  data.label.font.autocolor = NA, data.label.font.size = 10,
  data.label.format = "", data.label.prefix = "",
  data.label.suffix = "", legend.font.color = global.font.color,
  legend.font.family = global.font.family, legend.font.size = 10,
  grid.show = TRUE, y.title = "",
  y.title.font.color = global.font.color,
  y.title.font.family = global.font.family, y.title.font.size = 12,
  y.line.width = 0, y.line.color = rgb(0, 0, 0, maxColorValue = 255),
  y.bounds.minimum = NULL, y.bounds.maximum = NULL,
  y.tick.distance = NULL, y.grid.width = 1, y.grid.color = rgb(225,
  225, 225, maxColorValue = 255), y.tick.show = TRUE,
  y.tick.suffix = "", y.tick.prefix = "", y.tick.format = "",
  y.tick.font.color = global.font.color,
  y.tick.font.family = global.font.family, y.tick.font.size = 10,
  x.title = "", x.title.font.color = global.font.color,
  x.title.font.family = global.font.family, x.title.font.size = 12,
  x.line.width = 0, x.line.color = rgb(0, 0, 0, maxColorValue = 255),
  x.bounds.minimum = NULL, x.bounds.maximum = NULL,
  x.tick.distance = NULL, x.grid.width = 1, x.grid.color = rgb(225,
  225, 225, maxColorValue = 255), x.tick.show = TRUE,
  x.tick.suffix = "", x.tick.prefix = "", x.tick.format = "",
  x.tick.font.color = global.font.color,
  x.tick.font.family = global.font.family, x.tick.font.size = 10,
  marker.size = 6, swap.x.and.y = FALSE, ...)
}
\arguments{
\item{x}{A numeric vector for the x-axis coordinates (which may be named); or a matrix or dataframe; or a list of matrices, where each matrix share the same row and column names}

\item{y}{Optional numeric vector for the y-axis coordinates. Should contain the same number of observations as x. If not provided, will use x instead.}

\item{scatter.x.column}{When \code{x} is a dataframe or matrix, the index of the column (1-based) which contains the x-coordinate data.}

\item{scatter.y.column}{When \code{x} is a dataframe or matrix, the index of the column (1-based) which contains the y-coordinate data.}

\item{scatter.labels}{Optional vector for labelling scatter points. This should be the same length as the number of observations in x and y.}

\item{scatter.labels.name}{Character; Used for labelling subtitles and footers.}

\item{scatter.sizes}{Numeric vector determining of the size of each observation. These can alternatively be provided as a column in \code{x}.}

\item{scatter.sizes.name}{Character; Used for labelling footers and legends.}

\item{scatter.sizes.column}{When \code{x} is a dataframe or matrix, the index of the column (1-based) which contains \code{scatter.sizes} data.}

\item{scatter.colors}{Numeric, character, or categorical vector determining the color of each observation. These can alternatively be provided as a column in \code{x}.}

\item{scatter.colors.name}{Character; Used for labelling footers.}

\item{scatter.colors.column}{When \code{x} is a dataframe or matrix, the index of the column (1-based) which contains \code{scatter.colors} data.}

\item{scatter.colors.as.categorical}{Boolean; Whether to treat colors as a categorical groups, or a numeric scale.}

\item{trend.lines}{Boolean indicating whether to plot trend lines for multiple tables.}

\item{logos}{Optional list of images to be used to label scatterplot instead of the row names.
This should be input as a comma-seperated list of URLs.}

\item{logo.size}{Numeric controlling the size of the logos.}

\item{colors}{A vector of colors to use in the chart. When \code{scatter.colors.as.categorical}, the vector of colors should have the length as the number of categories in \code{scatter.colors}. If \code{scatter.colors} is used as numeric vector, then a color ramp is constructed from the colors listed.}

\item{opacity}{of scatter point colors as an alpha value (0 to 1).}

\item{legend.show}{Logical; show the legend.}

\item{global.font.family}{Character; font family for all occurrences of any
font attribute for the chart unless specified individually.}

\item{global.font.color}{Global font color as a named color in character format
(e.g. "black") or an a hex code.}

\item{title}{Character; chart title.}

\item{title.font.family}{Character; title font family. Can be "Arial Black",
"Arial", "Comic Sans MS", "Courier New", "Georgia", "Impact",
"Lucida Console", "Lucida Sans Unicode", "Marlett", "Symbol", "Tahoma",
"Times New Roman", "Trebuchet MS", "Verdana", "Webdings"}

\item{title.font.color}{Title font color as a named color in character
format (e.g. "black") or a hex code.}

\item{title.font.size}{Integer; Title font size; default = 10.}

\item{subtitle}{Character}

\item{subtitle.font.family}{Character; subtitle font family}

\item{subtitle.font.color}{subtitle font color as a named color in
character format (e.g. "black") or an a hex code.}

\item{subtitle.font.size}{Integer; subtitle font size}

\item{footer}{Character}

\item{footer.font.family}{Character; footer font family}

\item{footer.font.color}{footer font color as a named color in
character format (e.g. "black") or an a hex code.}

\item{footer.font.size}{Integer; footer font size}

\item{footer.wrap}{Logical; whether the footer text should be wrapped.}

\item{footer.wrap.nchar}{Number of characters (approximately) in each 
line of the footer when \code{footer.wrap} \code{TRUE}.}

\item{scatter.max.labels}{Integer; the maximum number of labels to show on a Labeled Scatterplot.}

\item{data.label.font.family}{Character; font family for data label.}

\item{data.label.font.color}{Font color as a named color
in character format (e.g. "black") or an a hex code. This can be a single
color, a vector of colors (1 for each series/column), or a comma separated list
of colors}

\item{data.label.font.autocolor}{Boolean; If true, \code{data.label.font.color} is ignored and labels are colored
according to the series color.}

\item{data.label.font.size}{Integer; Font size for data label.px.}

\item{data.label.format}{A string representing a d3 formatting code.
See https://github.com/mbostock/d3/wiki/Formatting#numbers}

\item{data.label.prefix}{Character; prefix for data values.}

\item{data.label.suffix}{Character; suffix for data values.}

\item{legend.font.color}{Legend font color as a named color in character
format (e.g. "black") or a hex code.}

\item{legend.font.family}{Character; legend font family.}

\item{legend.font.size}{Integer; Legend font size.}

\item{grid.show}{Logical; whether to show grid lines.}

\item{y.title}{Character, y-axis title; defaults to chart input values;
to turn off set to "FALSE".}

\item{y.title.font.color}{y-axis title font color as a named color in
character format (e.g. "black") or a hex code.}

\item{y.title.font.family}{Character; y-axis title font family}

\item{y.title.font.size}{Integer; y-axis title font size}

\item{y.line.width}{y-axis line width in pixels (0 = no line).}

\item{y.line.color}{y-axis line color as a named color in character format
(e.g. "black") or a hex code.}

\item{y.bounds.minimum}{Minimum of range for plotting; For a date axis this should be supplied as a date string.
For a categorical axis, the index of the category (0-based) should be used.}

\item{y.bounds.maximum}{Maximum of range for plotting; NULL = no manual range set.}

\item{y.tick.distance}{Distance between tick marks. Requires that \code{y.bounds.minimum} and \code{y.bounds.maximum} have been set.}

\item{y.grid.width}{Width of y-grid lines in pixels; 0 = no line}

\item{y.grid.color}{Color of y-grid lines as a named color in character
format (e.g. "black") or a hex code.}

\item{y.tick.show}{Whether to display the y-axis tick labels}

\item{y.tick.suffix}{y-axis tick label suffix}

\item{y.tick.prefix}{y-axis tick label prefix}

\item{y.tick.format}{A string representing a d3 formatting code.
See https://github.com/d3/d3/blob/master/API.md#number-formats-d3-format}

\item{y.tick.font.color}{y-axis tick label font color as a named color
in character format (e.g. "black") or an a hex code.}

\item{y.tick.font.family}{Character; y-axis tick label font family}

\item{y.tick.font.size}{Integer; y-axis tick label font size}

\item{x.title}{Character, x-axis title; defaults to chart input values;
to turn off set to "FALSE".}

\item{x.title.font.color}{x-axis title font color as a named color in
character format (e.g. "black") or an a hex code.}

\item{x.title.font.family}{Character; x-axis title font family}

\item{x.title.font.size}{Integer; x-axis title font size}

\item{x.line.width}{x-axis line in pixels, 0 = no line}

\item{x.line.color}{x-axis line color as a named color in character format
(e.g. "black") or a hex code.}

\item{x.bounds.minimum}{Minimum of range for plotting; For a date axis this should be supplied as a date string.
For a categorical axis, the index of the category (0-based) should be used.}

\item{x.bounds.maximum}{Maximum of range for
plotting; NULL = no manual range set.  Must be greater than x.bounds.minimum}

\item{x.tick.distance}{Tick mark distance in
x-axis units between minimum and maximum for plotting; NULL = no manual
range set.}

\item{x.grid.width}{Width of y-grid lines in pixels; 0 = no line}

\item{x.grid.color}{Color of y-grid lines as a named color in character
format (e.g. "black") or a hex code.}

\item{x.tick.show}{Whether to display the x-axis tick labels}

\item{x.tick.suffix}{x-axis tick label suffix}

\item{x.tick.prefix}{x-axis tick label prefix}

\item{x.tick.format}{A string representing a d3 formatting code.
See https://github.com/mbostock/d3/wiki/Formatting#numbers}

\item{x.tick.font.color}{X-axis tick label font color as a named color in
character format (e.g. "black") or an a hex code.}

\item{x.tick.font.family}{Character; x-axis tick label font family}

\item{x.tick.font.size}{Integer; x-axis tick label font size}

\item{marker.size}{Size in pixels of marker.}

\item{swap.x.and.y}{Swap the x and y axis around on the chart.}

\item{...}{Other arguments which are ignored.}
}
\description{
Labeled Scatter Chart
}
\examples{
z <- structure(c(1L, 2L, 3L, 4L, 5L, 2L, 3L, 4L, 5L, 6L),  .Dim = c(5L, 2L),
      .Dimnames = list(c("T", "U", "V", "W", "X"), c("A", "B")))
Column(z, type="Stacked Column")
}
