% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriescolorgradient.R
\name{MakeColorGradient}
\alias{MakeColorGradient}
\title{Generates a vector of colors based on the number of rows in a matrix}
\usage{
MakeColorGradient(x, red, green, blue)
}
\arguments{
\item{x}{A matrix.}

\item{red}{An integer between 0 and 255.}

\item{green}{An integer between 0 and 255.}

\item{blue}{An integer between 0 and 255.}
}
\value{
A named elements vector of colors with decreasing alpha values.
}
\description{
\code{MakeColorGradient} Gives a named-elements vector of colours for
each row in a matrix, where the base colour is specified, and the alpha
value gradually decreases based on the row-mean of the matrix,
and where the colour vector is returned sorted alphabetically
}
\examples{
data("z")
MakeColorGradient(z, red = 192, green = 35, blue = 220)
}

