% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palm.R
\name{Palm}
\alias{Palm}
\title{Generates a palm tree plot from a table of data.}
\usage{
Palm(table, x.tick.font.family = "", x.tick.font.size = 11, x.title = "",
  x.title.font.family = "", x.title.font.size = 12,
  legend.font.family = "", legend.font.size = 11, y.axis.show = TRUE,
  tooltip.show = TRUE, y.tick.format = "", y.tick.font.family = "",
  y.tick.font.size = 11, y.title = "", y.title.font.family = "",
  y.title.font.size = 12, y.tick.prefix = NULL, y.tick.suffix = NULL,
  colors = NULL)
}
\arguments{
\item{table}{A matrix of data to be displayed. Each row is represented by a palm tree, the value in each column
determines the area of a leaf on each palm tree.}

\item{x.tick.font.family}{Character; x-axis tick label font family}

\item{x.tick.font.size}{x-axis tick label font size}

\item{x.title}{Character; x-axis title}

\item{x.title.font.family}{Character; x-axis title font family}

\item{x.title.font.size}{x-axis title font size}

\item{legend.font.family}{Character; legend font family}

\item{legend.font.size}{Legend font size}

\item{y.axis.show}{Logical; whether t show the y-axis}

\item{tooltip.show}{Logical; whether to show a tooltip on hover.}

\item{y.tick.format}{A string representing a d3 formatting code.
See https://github.com/mbostock/d3/wiki/Formatting#numbers or
https://docs.python.org/release/3.1.3/library/string.html#formatspec}

\item{y.tick.font.family}{Character; y-axis tick label font family}

\item{y.tick.font.size}{y-axis tick label font size}

\item{y.title}{Character, y-axis title}

\item{y.title.font.family}{Character; y-axis title font family}

\item{y.title.font.size}{y-axis title font size}

\item{y.tick.prefix}{y-axis tick label prefix}

\item{y.tick.suffix}{y-axis tick label suffix}

\item{colors}{Character; a vector containing one or more named
colors from grDevices OR one or more specified hex value colors OR a single
named palette from grDevices, RColorBrewer, colorspace, or colorRamps.}
}
\description{
This function wraps the PalmTrees function in the rhtmlPalmTrees package and performs additional data preparation.
}
\examples{
z <- matrix(runif(20), nrow = 4)
rownames(z) <- c("oak", "elm", "birch", "fir")
colnames(z) <- c("leaf", "frond", "twig", "branch", "stick")
Palm(z)
}
