% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelcoords.R
\name{ParallelCoordinates}
\alias{ParallelCoordinates}
\title{Parallel coordinates}
\usage{
ParallelCoordinates(
  x,
  opacity = 0.4,
  group = NULL,
  colors = ChartColors(5, "Spectral"),
  reverse.axes = TRUE,
  global.font.family = "Arial",
  global.font.color = rgb(44, 44, 44, maxColorValue = 255),
  label.font.family = global.font.family,
  label.font.color = global.font.color,
  label.font.size = 12,
  label.rotate = FALSE,
  tick.font.family = global.font.family,
  tick.font.color = global.font.color,
  tick.font.size = 10,
  margin.top = 0,
  margin.bottom = 0,
  margin.left = 0,
  margin.right = 0,
  auto.resize = TRUE,
  interactive = TRUE,
  queue = FALSE,
  queue.rate = NULL,
  font.unit = "px",
  width = NULL,
  height = NULL,
  ...
)
}
\arguments{
\item{x}{A dataframe containing multiple variables (as columns)}

\item{opacity}{Opacity of the line colors as an alpha value (0 to 1).}

\item{group}{An optional variable which is used to color the lines in plot.
It should be a vector with length equal to the number of rows in \code{x}.
Can be numeric or categorical.}

\item{colors}{A vector of colors for the color in lines in plot. Note that only 6-digit
hex codes are accepted (8-digit hex results in black lines). If no \code{group}
is provided, then the first color will be used for all the lines. If it is provided,
\code{colors} will be interpolated (linearly) to create a color scalebar.}

\item{reverse.axes}{By default this is \code{true}, so that the smallest values are at the
top of the y-axes. Setting it to \code{false}, will mean the axes start at the bottom of the chart.}

\item{global.font.family}{Character; font family for all occurrences of any
font attribute for the chart unless specified individually.}

\item{global.font.color}{Global font color as a named color in character format
(e.g. "black") or an a hex code.}

\item{label.font.family}{Character; label font family.}

\item{label.font.color}{Label font color as a named color in character
format (e.g. "black") or a hex code.}

\item{label.font.size}{Integer; Label font size.}

\item{label.rotate}{Boolean; whether to rotate the variable names on the chart.}

\item{tick.font.family}{Character; tick label font family.}

\item{tick.font.color}{Tick label font color as a named color in character
format (e.g. "black") or a hex code.}

\item{tick.font.size}{Integer; Tick label font size.}

\item{margin.top}{Margin between plot area and the top of the
graphic in pixels}

\item{margin.bottom}{Margin between plot area and the bottom of the
graphic in pixels}

\item{margin.left}{Margin between plot area and the left of the
graphic in pixels}

\item{margin.right}{Margin between plot area and the right of the
graphic in pixels}

\item{auto.resize}{Whether the chart should resize automatically. It does not appear to work.}

\item{interactive}{Whether the lines should be hidden/shown depending on mouse.}

\item{queue}{Logical; whether lines should be added slowly to the plot}

\item{queue.rate}{Specifies the speed if \code{queue = TRUE}. This specifies the number
of lines drawn per second.}

\item{font.unit}{One of "px" of "pt". By default all font sizes are specified in terms of
pixels ("px"). But changing this to "pt" will mean that the font sizes will be in terms
points ("pt"), which will be consistent with font sizes in text boxes.}

\item{width}{Numeric; Width of chart in pixels.}

\item{height}{Numeric; Height of chart in pixels.}

\item{...}{Parameters which are ignored. This is included so calls expecting the
plotly implementation which had more parameters will not cause errors}
}
\description{
Create a parallel coordinates plot to show multivariate data
}
