% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radarchart.R
\name{Radar}
\alias{Radar}
\title{Radar}
\usage{
Radar(x, title = "", title.font.family = global.font.family,
  title.font.color = global.font.color, title.font.size = 16,
  colors = ChartColors(max(1, ncol(x), na.rm = TRUE)), opacity = 0.4,
  background.fill.color = rgb(255, 255, 255, maxColorValue = 255),
  background.fill.opacity = 0,
  charting.area.fill.color = background.fill.color,
  charting.area.fill.opacity = 0, legend.show = TRUE,
  legend.fill.color = background.fill.color, legend.fill.opacity = 0,
  legend.border.color = rgb(44, 44, 44, maxColorValue = 255),
  legend.border.line.width = 0, legend.font.color = global.font.color,
  legend.font.family = global.font.family, legend.font.size = 10,
  legend.ascending = NA, legend.position.y = 1, legend.position.x = 1.02,
  margin.top = NULL, margin.bottom = NULL, margin.left = NULL,
  margin.right = NULL, margin.inner.pad = NULL, pad.left = 0,
  pad.right = 0, line.thickness = 3, tooltip.show = TRUE,
  modebar.show = FALSE, global.font.family = "Arial",
  global.font.color = rgb(44, 44, 44, maxColorValue = 255),
  grid.show = TRUE, x.tick.show = TRUE,
  x.tick.font.color = global.font.color,
  x.tick.font.family = global.font.family, x.tick.font.size = 12,
  x.grid.width = 1 * grid.show, x.grid.color = rgb(225, 225, 225,
  maxColorValue = 255), y.bounds.maximum = NULL, y.tick.distance = NULL,
  y.grid.width = 1 * grid.show, y.grid.color = rgb(225, 225, 225,
  maxColorValue = 255), y.tick.show = TRUE, y.tick.suffix = "",
  y.tick.prefix = "", y.tick.format = "", hovertext.show = TRUE,
  y.hovertext.format = "", y.tick.font.color = global.font.color,
  y.tick.font.family = global.font.family, y.tick.font.size = 10,
  x.tick.label.wrap = TRUE, x.tick.label.wrap.nchar = 21,
  data.label.show = FALSE, data.label.font.family = global.font.family,
  data.label.font.size = 10, data.label.font.color = global.font.color,
  data.label.format = "", data.label.prefix = "", data.label.suffix = "",
  subtitle = "", subtitle.font.family = global.font.family,
  subtitle.font.color = global.font.color, subtitle.font.size = 12,
  footer = "", footer.font.family = global.font.family,
  footer.font.color = global.font.color, footer.font.size = 8,
  footer.wrap = TRUE, footer.wrap.nchar = 100)
}
\arguments{
\item{x}{Input data in the form of a vector or matrix. The categories used
to create the radar (i.e. the x-axis) is taken from the names/rownames of x.}

\item{title}{Character; chart title.}

\item{title.font.family}{Character; title font family. Can be "Arial Black",
"Arial", "Comic Sans MS", "Courier New", "Georgia", "Impact",
"Lucida Console", "Lucida Sans Unicode", "Marlett", "Symbol", "Tahoma",
"Times New Roman", "Trebuchet MS", "Verdana", "Webdings"}

\item{title.font.color}{Title font color as a named color in character
format (e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}

\item{title.font.size}{Title font size; default = 10.}

\item{colors}{Character; a vector containing one or more named
colors from grDevices OR one or more specified hex value colors OR a single
named palette from grDevices, RColorBrewer, colorspace, or colorRamps.}

\item{opacity}{Opacity of area fill colors as an alpha value (0 to 1).}

\item{background.fill.color}{Background color in character format
(e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}

\item{background.fill.opacity}{Background opacity as an alpha value
(0 to 1).}

\item{charting.area.fill.color}{Charting area background color as
a named color in character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{charting.area.fill.opacity}{Charting area background}

\item{legend.show}{Logical; show the legend.}

\item{legend.fill.color}{Legend fill color as a named color in character format
(e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}

\item{legend.fill.opacity}{Legend fill opacity as an alpha value
(0 to 1).}

\item{legend.border.color}{Legend border color as a named color in character
format (e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}

\item{legend.border.line.width}{Width in pixels of the border
around the legend.  0 = no border.}

\item{legend.font.color}{Legend font color as a named color in character
format (e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}

\item{legend.font.family}{Character; legend font family.}

\item{legend.font.size}{Legend font size.}

\item{legend.ascending}{Logical; TRUE for ascending, FALSE for descending.
By default, we set it to to FALSE if the chart is stacked and TRUE otherwise.}

\item{legend.position.y}{A numeric controlling the position of the legend.
Values range from 0 (bottom) to 1 (top).}

\item{legend.position.x}{A numeric controlling the position of the legend.
Values range from -0.5 (left) to 1.5 (right).}

\item{margin.top}{Margin between plot area and the top of the
graphic in pixels}

\item{margin.bottom}{Margin between plot area and the bottom of the
graphic in pixels}

\item{margin.left}{Margin between plot area and the left of the
graphic in pixels}

\item{margin.right}{Margin between plot area and the right of the
graphic in pixels}

\item{margin.inner.pad}{Padding in pixels between plot proper
and axis lines}

\item{pad.left}{Spacing on the left of the chart. Mainly used by SmallMultiples.}

\item{pad.right}{Spacing on the right of the chart. Mainly used by SmallMultiples.}

\item{line.thickness}{Thickness of outline of radar polygons.}

\item{tooltip.show}{Logical; whether to show a tooltip on hover.}

\item{modebar.show}{Logical; whether to show the zoom menu buttons or not.}

\item{global.font.family}{Character; font family for all occurrences of any
font attribute for the chart unless specified individually.}

\item{global.font.color}{Global font color as a named color in character format
(e.g. "black") or an rgb value (e.g. #' rgb(0, 0, 0, maxColorValue = 255)).}

\item{grid.show}{Logical; Whether to show grid lines.
opacity as an alpha value (0 to 1).}

\item{x.tick.show}{Whether to display the x-axis tick labels
(i.e. labels around the sides of the radar chart)}

\item{x.tick.font.color}{x-axis title font color as a named color in
character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{x.tick.font.family}{Character; x-axis title font family}

\item{x.tick.font.size}{x-axis title font size}

\item{x.grid.width}{Width of y-grid lines in pixels; 0 = no line}

\item{x.grid.color}{Color of y-grid lines as a named color in character
format (e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}

\item{y.bounds.maximum}{Maximum of range for
plotting; NULL = no manual range set.  The minimum is always 0.}

\item{y.tick.distance}{Tick mark distance.}

\item{y.grid.width}{Width of y-grid lines in pixels; 0 = no line}

\item{y.grid.color}{Color of y-grid lines as a named color in character
format (e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}

\item{y.tick.show}{Whether to display the y-axis tick labels
(i.e. radial distance from center)}

\item{y.tick.suffix}{y-axis tick label suffix}

\item{y.tick.prefix}{y-axis tick label prefix}

\item{y.tick.format}{A string representing a d3 formatting code.
See https://github.com/mbostock/d3/wiki/Formatting#numbers}

\item{hovertext.show}{Logical; whether to show hovertext.}

\item{y.hovertext.format}{A string representing a d3 formatting code.
See https://github.com/mbostock/d3/wiki/Formatting#numbers}

\item{y.tick.font.color}{y-axis tick label font color as a named color
in character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{y.tick.font.family}{Character; y-axis tick label font family}

\item{y.tick.font.size}{y-axis tick label font size}

\item{x.tick.label.wrap}{Logical; whether to wrap long labels on the x-axis.}

\item{x.tick.label.wrap.nchar}{Integer; number of characters in each line when \code{label.wrap} is \code{TRUE}.}

\item{data.label.show}{Logical; whether to show data labels.}

\item{data.label.font.family}{Character; font family for data label.}

\item{data.label.font.size}{Font size for data label.}

\item{data.label.font.color}{Font color as a named color
in character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{data.label.format}{A string representing a d3 formatting code.
See https://github.com/mbostock/d3/wiki/Formatting#numbers}

\item{data.label.prefix}{Character; prefix for data values.}

\item{data.label.suffix}{Character; suffix for data values.}

\item{subtitle}{Character}

\item{subtitle.font.family}{Character; subtitle font family}

\item{subtitle.font.color}{subtitle font color as a named color in
character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{subtitle.font.size}{subtitle font size}

\item{footer}{Character}

\item{footer.font.family}{Character; footer font family}

\item{footer.font.color}{footer font color as a named color in
character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{footer.font.size}{footer font size}

\item{footer.wrap}{Logical; whether the footer text should be wrapped.}

\item{footer.wrap.nchar}{Number of characters (approximately) in each line of the footer when \code{footer.wordwrap} \code{TRUE}.}

\item{...}{Extra arguments that are ignored.}
}
\description{
Radar chart, also known as web chart, spider chart, star chart, star plot, cobweb chart, irregular polygon, polar chart, or Kiviat diagram
}
