% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterchart.R
\name{Scatter}
\alias{Scatter}
\title{Scatter}
\usage{
Scatter(x = NULL, y = NULL, scatter.x.column = 1, scatter.y.column = 2,
  scatter.labels = NULL, scatter.labels.name = NULL, scatter.sizes = NULL,
  scatter.sizes.name = NULL, scatter.sizes.column = 3,
  scatter.colors = NULL, scatter.colors.name = NULL,
  scatter.colors.column = 4, scatter.colors.as.categorical = FALSE,
  colors = ChartColors(12), fit.type = "None", fit.ignore.last = FALSE,
  fit.line.type = "dot", fit.line.width = 1, fit.line.name = "Fitted",
  fit.line.colors = colors, legend.show = TRUE, tooltip.show = TRUE,
  modebar.show = FALSE, global.font.family = "Arial",
  global.font.color = rgb(44, 44, 44, maxColorValue = 255), title = "",
  title.font.family = global.font.family,
  title.font.color = global.font.color, title.font.size = 16,
  subtitle = "", subtitle.font.family = global.font.family,
  subtitle.font.color = global.font.color, subtitle.font.size = 12,
  footer = "", footer.font.family = global.font.family,
  footer.font.color = global.font.color, footer.font.size = 8,
  footer.wrap = TRUE, footer.wrap.nchar = 100, data.label.show = FALSE,
  data.label.font.family = global.font.family,
  data.label.font.color = global.font.color, data.label.font.size = 10,
  data.label.decimals = 2, data.label.prefix = "", data.label.suffix = "",
  data.label.position = "top middle", opacity = NULL,
  background.fill.color = rgb(255, 255, 255, maxColorValue = 255),
  background.fill.opacity = 1, charting.area.fill.color = rgb(255, 255, 255,
  maxColorValue = 255), charting.area.fill.opacity = 1,
  legend.fill = rgb(255, 255, 255, maxColorValue = 255),
  legend.fill.color = legend.fill, legend.fill.opacity = 1,
  legend.border.color = rgb(44, 44, 44, maxColorValue = 255),
  legend.border.line.width = 0, legend.font.color = global.font.color,
  legend.font.family = global.font.family, legend.font.size = 10,
  legend.position.y = 1, legend.position.x = 1.02, legend.ascending = NA,
  margin.top = NULL, margin.bottom = NULL, margin.left = NULL,
  margin.right = NULL, margin.inner.pad = NULL, grid.show = TRUE,
  y.title = "", y.title.font.color = global.font.color,
  y.title.font.family = global.font.color, y.title.font.size = 12,
  y.line.width = 0, y.line.color = rgb(0, 0, 0, maxColorValue = 255),
  y.tick.marks = "", y.tick.mark.length = 5, y.bounds.minimum = NULL,
  y.bounds.maximum = NULL, y.tick.distance = NULL, y.zero = TRUE,
  y.zero.line.width = 0, y.zero.line.color = rgb(44, 44, 44, maxColorValue =
  255), y.position = "left", y.data.reversed = FALSE, y.grid.width = 1 *
  grid.show, y.grid.color = rgb(225, 225, 225, maxColorValue = 255),
  y.tick.show = TRUE, y.tick.suffix = "", y.tick.prefix = "",
  y.tick.format = "", y.hovertext.format = y.tick.format,
  y.tick.angle = NULL, y.tick.font.color = global.font.color,
  y.tick.font.family = global.font.family, y.tick.font.size = 10,
  x.title = "", x.title.font.color = global.font.color,
  x.title.font.family = global.font.family, x.title.font.size = 12,
  x.line.width = 0, x.line.color = rgb(0, 0, 0, maxColorValue = 255),
  x.tick.marks = "", x.tick.mark.length = 5, x.bounds.minimum = NULL,
  x.bounds.maximum = NULL, x.tick.distance = NULL, x.zero.line.width = 0,
  x.zero.line.color = rgb(44, 44, 44, maxColorValue = 255),
  x.position = "bottom", x.data.reversed = FALSE, x.grid.width = 1 *
  grid.show, x.grid.color = rgb(225, 225, 225, maxColorValue = 255),
  x.tick.show = TRUE, x.tick.suffix = "", x.tick.prefix = "",
  x.tick.format = "", x.hovertext.format = x.tick.format,
  x.tick.angle = NULL, x.tick.font.color = global.font.color,
  x.tick.font.family = global.font.family, x.tick.font.size = 10,
  x.tick.label.wrap = TRUE, x.tick.label.wrap.nchar = 21,
  series.line.width = 0, series.marker.border.width = 1,
  series.marker.size = 6, series.marker.opacity = 1,
  series.marker.show = "none", series.marker.colors = NULL,
  swap.x.and.y = FALSE)
}
\arguments{
\item{x}{A numeric vector for the x-axis coordinates (which may be named); or a matrix or dataframe.}

\item{y}{Optional numeric vector for the y-axis coordinates. Should contain the same number of observations as x. If not provided, will use x instead.}

\item{scatter.x.column}{When \code{x} is a dataframe or matrix, the index of the column (1-based) which contains the x-coordinate data.}

\item{scatter.y.column}{When \code{x} is a dataframe or matrix, the index of the column (1-based) which contains the y-coordinate data.}

\item{scatter.labels}{Optional vector for labelling scatter points. This should be the same length as the number of observations in x andy. This is used in the hovertext and data labels.}

\item{scatter.labels.name}{Character; Used for labelling subtitles and footers.}

\item{scatter.sizes}{Numeric vector determining of the size of each observation. These can alternatively be provided as a column in \code{x}.}

\item{scatter.sizes.name}{Character; Used for labelling footers and legends.}

\item{scatter.sizes.column}{When \code{x} is a dataframe or matrix, the index of the column (1-based) which contains \code{scatter.sizes} data.}

\item{scatter.colors}{Numeric, character, or categorical vector determining the color of each observation. These can alternatively be provided as a column in \code{x}.}

\item{scatter.colors.name}{Character; Used for labelling footers.}

\item{scatter.colors.column}{When \code{x} is a dataframe or matrix, the index of the column (1-based) which contains \code{scatter.colors} data.}

\item{scatter.colors.as.categorical}{Whether to treat colors as a categorical groups, or a numeric scale.}

\item{colors}{A vector of colors to use in the chart. When \code{scatter.colors.as.categorical}, the vector of colors should have the length as the number of categories in \code{scatter.colors}. If \code{scatter.colors} is used as numeric vector, then a color ramp is constructed from the colors listed.}

\item{fit.type}{Character; type of line of best fit. Can be one of "None", "Linear" or "Smooth" (loess local polynomial fitting).}

\item{fit.ignore.last}{Boolean; whether to ignore the last data point in the fit.}

\item{fit.line.type}{Character; One of "solid", "dot", "dash, "dotdash", or length of dash "2px", "5px".}

\item{fit.line.width}{Numeric; Line width of line of best fit.}

\item{fit.line.name}{Character; Name of the line of best fit, which will appear in the hovertext.}

\item{fit.line.colors}{Character; a vector containing one or more named
colors from grDevices OR one or more specified hex value colors OR a single
named palette from grDevices, RColorBrewer, colorspace, or colorRamps.}

\item{legend.show}{Logical; show the legend.}

\item{tooltip.show}{Logical; whether to show a tooltip on hover.}

\item{modebar.show}{Logical; whether to show the zoom menu buttons or not.}

\item{global.font.family}{Character; font family for all occurrences of any
font attribute for the chart unless specified individually.}

\item{global.font.color}{Global font color as a named color in character format
(e.g. "black") or an rgb value (e.g. #' rgb(0, 0, 0, maxColorValue = 255)).}

\item{title}{Character; chart title.}

\item{title.font.family}{Character; title font family. Can be "Arial Black",
"Arial", "Comic Sans MS", "Courier New", "Georgia", "Impact",
"Lucida Console", "Lucida Sans Unicode", "Marlett", "Symbol", "Tahoma",
"Times New Roman", "Trebuchet MS", "Verdana", "Webdings"}

\item{title.font.color}{Title font color as a named color in character
format (e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}

\item{title.font.size}{Title font size; default = 10.}

\item{subtitle}{Character}

\item{subtitle.font.family}{Character; subtitle font family}

\item{subtitle.font.color}{subtitle font color as a named color in
character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{subtitle.font.size}{subtitle font size}

\item{footer}{Character}

\item{footer.font.family}{Character; footer font family}

\item{footer.font.color}{footer font color as a named color in
character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{footer.font.size}{footer font size}

\item{footer.wrap}{Logical; whether the footer text should be wrapped.}

\item{footer.wrap.nchar}{Number of characters (approximately) in each line of the footer when \code{footer.wordwrap} \code{TRUE}.}

\item{data.label.show}{Logical; whether to show data labels.}

\item{data.label.font.family}{Character; font family for data label.}

\item{data.label.font.color}{Font color as a named color
in character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{data.label.font.size}{Font size for data label.}

\item{data.label.decimals}{Number of decimal places to show in
data labels.}

\item{data.label.prefix}{Character; prefix for data values.}

\item{data.label.suffix}{Character; suffix for data values.}

\item{data.label.position}{Character; where to place the source data
value in relation to the marker icon.  Can be "top left", "top center", "top
right", "middle left", "middle center", "middle right", "bottom left",
"bottom center", "bottom right". Only applicable for line and area charts.}

\item{opacity}{Opacity of area fill colors as an alpha value (0 to 1).}

\item{background.fill.color}{Background color in character format
(e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}

\item{background.fill.opacity}{Background opacity as an alpha value
(0 to 1).}

\item{charting.area.fill.color}{Charting area background color as
a named color in character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{charting.area.fill.opacity}{Charting area background
opacity as an alpha value (0 to 1).}

\item{legend.fill}{Same as \code{legend.fill.color}. Retained for backwards compatibility.}

\item{legend.fill.color}{Legend fill color as a named color in character format
(e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}

\item{legend.fill.opacity}{Legend fill opacity as an alpha value
(0 to 1).}

\item{legend.border.color}{Legend border color as a named color in character
format (e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}

\item{legend.border.line.width}{Width in pixels of the border
around the legend.  0 = no border.}

\item{legend.font.color}{Legend font color as a named color in character
format (e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}

\item{legend.font.family}{Character; legend font family.}

\item{legend.font.size}{Legend font size.}

\item{legend.position.y}{A numeric controlling the position of the legend.
Values range from 0 (bottom) to 1 (top).}

\item{legend.position.x}{A numeric controlling the position of the legend.
Values range from -0.5 (left) to 1.5 (right).}

\item{legend.ascending}{Logical; TRUE for ascending, FALSE for descending.
By default, we set it to to FALSE if the chart is stacked and TRUE otherwise.}

\item{margin.top}{Margin between plot area and the top of the
graphic in pixels}

\item{margin.bottom}{Margin between plot area and the bottom of the
graphic in pixels}

\item{margin.left}{Margin between plot area and the left of the
graphic in pixels}

\item{margin.right}{Margin between plot area and the right of the
graphic in pixels}

\item{margin.inner.pad}{Padding in pixels between plot proper
and axis lines}

\item{grid.show}{Logical; Whether to show grid lines.}

\item{y.title}{Character, y-axis title; defaults to chart input values;
to turn off set to "FALSE".}

\item{y.title.font.color}{y-axis title font color as a named color in
character format (e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, max = 255)).}

\item{y.title.font.family}{Character; y-axis title font family}

\item{y.title.font.size}{y-axis title font size}

\item{y.line.width}{y-axis line in pixels, 0 = no line}

\item{y.line.color}{y-axis line color as a named color in character format
(e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}

\item{y.tick.marks}{Character; whether and where to show tick marks on the
y axis.  Can be "outside", "inside", "none"}

\item{y.tick.mark.length}{Length of tick marks in pixels.}

\item{y.bounds.minimum}{Minimum of range for plotting;
NULL = no manual range set.  Must be less than y.bounds.maximum}

\item{y.bounds.maximum}{Maximum of range for
plotting; NULL = no manual range set.  Must be greater than y.bounds.minimum}

\item{y.tick.distance}{Tick mark distance.}

\item{y.zero}{Whether the y-axis should include zero.}

\item{y.zero.line.width}{Width in pixels of zero line; 0 = no zero line
shown}

\item{y.zero.line.color}{Color of horizontal zero line as a named
color in character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{y.position}{Character; set y-axis position; can be "left" or "right"}

\item{y.data.reversed}{Logical; whether to reverse y-axis or not}

\item{y.grid.width}{Width of y-grid lines in pixels; 0 = no line}

\item{y.grid.color}{Color of y-grid lines as a named color in character
format (e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}

\item{y.tick.show}{Whether to display the y-axis tick labels}

\item{y.tick.suffix}{y-axis tick label suffix}

\item{y.tick.prefix}{y-axis tick label prefix}

\item{y.tick.format}{Overrides tick.prefix, suffix and decimals;
See https://github.com/mbostock/d3/wiki/Formatting#numbers or
https://docs.python.org/release/3.1.3/library/string.html#formatspec}

\item{y.hovertext.format}{Overrides hovertext decimals;
See https://github.com/mbostock/d3/wiki/Formatting#numbers or
https://docs.python.org/release/3.1.3/library/string.html#formatspec}

\item{y.tick.angle}{y-axis tick label angle in degrees.
90 = vertical; 0 = horizontal}

\item{y.tick.font.color}{y-axis tick label font color as a named color
in character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{y.tick.font.family}{Character; y-axis tick label font family}

\item{y.tick.font.size}{y-axis tick label font size}

\item{x.title}{Character, x-axis title; defaults to chart input values;
to turn off set to "FALSE".}

\item{x.title.font.color}{x-axis title font color as a named color in
character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{x.title.font.family}{Character; x-axis title font family}

\item{x.title.font.size}{x-axis title font size}

\item{x.line.width}{x-axis line in pixels, 0 = no line}

\item{x.line.color}{x-axis line color as a named color in character format
(e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}

\item{x.tick.marks}{Character; whether and where to show tick marks on the
x-axis.  Can be "outside", "inside", "none"}

\item{x.tick.mark.length}{Length of tick marks in pixels.}

\item{x.bounds.minimum}{Minimum of range for plotting;
NULL = no manual range set.  Must be less than x.bounds.maximum}

\item{x.bounds.maximum}{Maximum of range for
plotting; NULL = no manual range set.  Must be greater than x.bounds.minimum}

\item{x.tick.distance}{Tick mark distance in
x-axis units between minimum and maximum for plotting; NULL = no manual
range set.}

\item{x.zero.line.width}{Width in pixels of zero line; 0 = no zero line
shown}

\item{x.zero.line.color}{Color of horizontal zero (origo) line as a named
color in character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{x.position}{Character; set x-axis position; can be "top" or "bottom"}

\item{x.data.reversed}{Logical; whether to reverse x-axis or not}

\item{x.grid.width}{Width of y-grid lines in pixels; 0 = no line}

\item{x.grid.color}{Color of y-grid lines as a named color in character
format (e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}

\item{x.tick.show}{Whether to display the x-axis tick labels}

\item{x.tick.suffix}{x-axis tick label suffix}

\item{x.tick.prefix}{x-axis tick label prefix}

\item{x.tick.format}{Overrides tick.prefix, suffix and decimals;
See https://github.com/mbostock/d3/wiki/Formatting#numbers or
https://docs.px.hon.org/release/3.1.3/librarx.string.html#formatspec}

\item{x.hovertext.format}{Overrides hovertext decimals;
See https://github.com/mbostock/d3/wiki/Formatting#numbers or
https://docs.px.hon.org/release/3.1.3/librarx.string.html#formatspec}

\item{x.tick.angle}{x-axis tick label angle in degrees.
90 = vertical; 0 = horizontal}

\item{x.tick.font.color}{X-axis tick label font color as a named color in
character format (e.g. "black") or an rgb value (e.g.
rgb(0, 0, 0, maxColorValue = 255)).}

\item{x.tick.font.family}{Character; x-axis tick label font family}

\item{x.tick.font.size}{x-axis tick label font size}

\item{x.tick.label.wrap}{Logical; whether to wrap long labels on the x-axis.}

\item{x.tick.label.wrap.nchar}{Integer; number of characters in each line when \code{label.wrap} is \code{TRUE}.}

\item{series.line.width}{Thickness, in pixels, of the series line}

\item{series.marker.border.width}{Width in pixels of border/line
around series markers; 0 is no line}

\item{series.marker.size}{Size in pixels of marker}

\item{series.marker.opacity}{Opacity for series markers as an alpha value (0 to 1).}

\item{series.marker.show}{Can be "none", "automatic" or a vector referencing
the plotly symbol dictionary using either numerics or strings.}

\item{series.marker.colors}{Character; a vector containing one or more named
colors from grDevices OR one or more specified hex value colors OR a single
named palette from grDevices, RColorBrewer, colorspace, or colorRamps.
be reversed.}

\item{swap.x.and.y}{Swap the x and y axis around on the chart.}

\item{...}{Extra arguments that are ignored.}
}
\description{
Scatter plot
}
