% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallmultiples.R
\name{SmallMultiples}
\alias{SmallMultiples}
\title{Plotting data as a small multiples}
\usage{
SmallMultiples(
  x,
  chart.type = "Area",
  nrows = 2,
  share.axes = TRUE,
  pad.left = 0.01,
  pad.right = 0.01,
  pad.top = 0.01,
  pad.bottom = 0.01,
  x.order = NULL,
  average.show = FALSE,
  average.color = rgb(153, 153, 153, maxColorValue = 255),
  y.bounds.maximum = NULL,
  y.bounds.minimum = NULL,
  x.bounds.maximum = NULL,
  x.bounds.minimum = NULL,
  x.tick.maxnum = NULL,
  y.tick.maxnum = NULL,
  values.bounds.maximum = NULL,
  values.bounds.minimum = NULL,
  colors = ChartColors(max(1, ncol(x), nrow(x), na.rm = TRUE)),
  fit.line.colors = colors,
  fit.CI.colors = colors,
  global.font.family = "Arial",
  global.font.color = rgb(44, 44, 44, maxColorValue = 255),
  title = "",
  title.font.family = global.font.family,
  title.font.color = global.font.color,
  title.font.size = 16,
  title.align = "center",
  subtitle = "",
  subtitle.font.family = global.font.family,
  subtitle.font.color = global.font.color,
  subtitle.font.size = 12,
  subtitle.align = "center",
  footer = "",
  footer.font.family = global.font.family,
  footer.font.color = global.font.color,
  footer.font.size = 8,
  footer.align = "center",
  footer.wrap = TRUE,
  footer.wrap.nchar = 100,
  panel.title.show = TRUE,
  panel.title.font.family = global.font.family,
  panel.title.font.color = global.font.color,
  panel.title.font.size = 14,
  panel.title.wrap = TRUE,
  panel.title.wrap.nchar = 20,
  x.title = "",
  x.title.font.size = 12,
  y.title = "",
  y.title.font.size = 12,
  data.label.show = FALSE,
  data.label.prefix = "",
  data.label.suffix = "",
  data.label.font.color = global.font.color,
  line.thickness = NULL,
  x.tick.show = TRUE,
  x.tick.angle = NULL,
  legend.show = TRUE,
  margin.autoexpand = TRUE,
  margin.left = NULL,
  margin.right = NULL,
  margin.top = NULL,
  margin.bottom = NULL,
  mapping.package = "plotly",
  scatter.x.column = 1,
  scatter.y.column = 2,
  scatter.sizes.column = 3,
  scatter.colors.column = 4,
  scatter.groups.column = NULL,
  scatter.colors.as.categorical = TRUE,
  scatter.sizes.as.diameter = FALSE,
  ...
)
}
\arguments{
\item{x}{Input data as a matrix or dataframe}

\item{chart.type}{Can be one of "Area", "Column", "Bar", "Line", "Radar" or "Geographic Map".}

\item{nrows}{Integer; Number of rows to arrange the charts in}

\item{share.axes}{Force range of the plot to be the same across all panels.}

\item{pad.left}{Numeric in [0,1]; Spacing to the left of chart (between panels)}

\item{pad.right}{Numeric in [0,1]; Spacing to the right chart (between panels)}

\item{pad.top}{Numeric in [0,1]; Spacing above chart (between panels)}

\item{pad.bottom}{Numeric in [0,1]; Spacing below chart (between panels)}

\item{x.order}{A vector containing the list index of the columns in the order which they are to be shown}

\item{average.show}{Logical; whether to show a second series in each panel containing
the data averaged across all series.}

\item{average.color}{The color in which the average series should be displayed}

\item{y.bounds.maximum}{Maximum of range for plotting; NULL = no manual range set.}

\item{y.bounds.minimum}{Minimum of range for plotting; For a date axis this should be supplied as a date string.
For a categorical axis, the index of the category (0-based) should be used.}

\item{x.bounds.maximum}{Maximum of range for
plotting; NULL = no manual range set.  Must be greater than x.bounds.minimum}

\item{x.bounds.minimum}{Minimum of range for plotting; For a date axis this should be supplied as a date string.
For a categorical axis, the index of the category (0-based) should be used.}

\item{x.tick.maxnum}{Maximum number of ticks shown on the axis.
It defaults to 11 which gives the same output from plotly as NULL.
This setting is ignored if \code{x.tick.distance} is set or
if the axis is categorical}

\item{y.tick.maxnum}{Maximum number of ticks shown on the axis.
This setting is ignored if \code{y.tick.distance} is set or
if the axis is categorical}

\item{values.bounds.maximum}{Numeric; specifies the maximum value in the colorscale.}

\item{values.bounds.minimum}{Numeric; specifies the minimum value in the colorscale.
If not specified, this will be automatically determined from the data.
If the value specified is larger than the minimum in the data then
the specified value will be ignored.}

\item{colors}{Character; a vector containing one or more colors specified as hex codes.}

\item{fit.line.colors}{Character; a vector containing one or more colors specified as hex codes.}

\item{fit.CI.colors}{Character; a vector containing one or more colors specified as hex codes.}

\item{global.font.family}{Character; font family for all occurrences of any
font attribute for the chart unless specified individually.}

\item{global.font.color}{Global font color as a named color in character format
(e.g. "black") or an a hex code.}

\item{title}{Character; chart title.}

\item{title.font.family}{Character; title font family. Can be "Arial Black",
"Arial", "Comic Sans MS", "Courier New", "Georgia", "Impact",
"Lucida Console", "Lucida Sans Unicode", "Marlett", "Symbol", "Tahoma",
"Times New Roman", "Trebuchet MS", "Verdana", "Webdings"}

\item{title.font.color}{Title font color as a named color in character
format (e.g. "black") or a hex code.}

\item{title.font.size}{Integer; Title font size; default = 10.}

\item{title.align}{Horizontal alignment of title.}

\item{subtitle}{Character}

\item{subtitle.font.family}{Character; subtitle font family}

\item{subtitle.font.color}{subtitle font color as a named color in
character format (e.g. "black") or an a hex code.}

\item{subtitle.font.size}{Integer; subtitle font size}

\item{subtitle.align}{Horizontal alignment of subtitle.}

\item{footer}{Character}

\item{footer.font.family}{Character; footer font family}

\item{footer.font.color}{footer font color as a named color in
character format (e.g. "black") or an a hex code.}

\item{footer.font.size}{Integer; footer font size}

\item{footer.align}{Horizontal alignment of footer.}

\item{footer.wrap}{Logical; whether the footer text should be wrapped.}

\item{footer.wrap.nchar}{Number of characters (approximately) in each
line of the footer when \code{footer.wrap} \code{TRUE}.}

\item{panel.title.show}{Logical; whether to show a title for each panel.}

\item{panel.title.font.family}{Font family of panel titles.}

\item{panel.title.font.color}{Font color of panel titles.}

\item{panel.title.font.size}{Font size of panel titles.}

\item{panel.title.wrap}{Logical; whether the panel title should be wrapped.}

\item{panel.title.wrap.nchar}{Number of characters (approximately) in each line
of the panel title when \code{panel.title.wordwrap} \code{TRUE}.}

\item{x.title}{Character, x-axis title; defaults to chart input values;
to turn off set to "FALSE".}

\item{x.title.font.size}{Integer; x-axis title font size}

\item{y.title}{Character, y-axis title; defaults to chart input values;
to turn off set to "FALSE".}

\item{y.title.font.size}{Integer; y-axis title font size}

\item{data.label.show}{Logical; whether to show data labels.}

\item{data.label.prefix}{Character; prefix for data values.}

\item{data.label.suffix}{Character; suffix for data values.}

\item{data.label.font.color}{Font color as a named color
in character format (e.g. "black") or an a hex code. This can be a single
color, a vector of colors (1 for each series/column), or a comma separated list
of colors}

\item{line.thickness}{Thickness, in pixels, of the series line}

\item{x.tick.show}{Whether to display the x-axis tick labels}

\item{x.tick.angle}{x-axis tick label angle in degrees.
90 = vertical; 0 = horizontal}

\item{legend.show}{Ignored except for with \code{GeographicMap}.}

\item{margin.autoexpand}{Logical; Whether extra space can be added to the margins
to allow space for axis/legend/data labels or other chart elements.}

\item{margin.left}{Margin between plot area and the left of the graphic in pixels}

\item{margin.right}{Margin between plot area and the right of the graphic in pixels}

\item{margin.top}{Margin between plot area and the top of the graphic in pixels}

\item{margin.bottom}{Margin between plot area and the bottom of the graphic in pixels}

\item{mapping.package}{Not used.}

\item{scatter.x.column}{When \code{x} is a dataframe or matrix, the index of the column (1-based) which contains the x-coordinate data.}

\item{scatter.y.column}{When \code{x} is a dataframe or matrix, the index of the column (1-based) which contains the y-coordinate data.}

\item{scatter.sizes.column}{When \code{x} is a dataframe or matrix, the index of the column (1-based) which contains \code{scatter.sizes} data.}

\item{scatter.colors.column}{When \code{x} is a dataframe or matrix, the index of the column (1-based) which contains \code{scatter.colors} data.}

\item{scatter.groups.column}{The column of \code{x} which is used to aggregate
the data for small multiples. By default this is the last column in \code{x}}

\item{scatter.colors.as.categorical}{Boolean; Whether to treat colors as a categorical groups, or a numeric scale.}

\item{scatter.sizes.as.diameter}{Whether to show the points with diameter (instead of area, which is the default) proportional to the sizes variable.}

\item{...}{Extra arguments passed to the charting function}
}
\value{
an HTML widget for \code{"leaflet"} or a \code{"plotly"} object.
}
\description{
A small multiple (sometimes called trellis chart, lattice chart,
   grid chart, or panel chart) is a series of similar graphs or charts using
   the same scale and axes, allowing them to be easily compared.
}
\examples{
x <- matrix(1:21, 7, 3, dimnames = list(letters[1:7], LETTERS[1:3]))
SmallMultiples(x, "Column", colors=c("red","green","blue"))
}
