% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographicmaphelper.R
\name{tidyCountryName}
\alias{tidyCountryName}
\title{Standardize country name}
\usage{
tidyCountryName(country)
}
\arguments{
\item{country}{The country to search for}
}
\description{
Standardize country name
}
