% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{TimeSeries}
\alias{TimeSeries}
\title{TimeSeries}
\usage{
TimeSeries(x = NULL, range.bars = FALSE, colors = NULL,
  line.thickness = NULL, legend.width = 250, window.start = NULL,
  global.font.family = "Arial", global.font.color = rgb(44, 44, 44,
  maxColorValue = 255), title = "",
  title.font.family = global.font.family,
  title.font.color = global.font.color, title.font.size = 16,
  x.title = "", x.title.font.color = global.font.color,
  x.title.font.family = global.font.family, x.title.font.size = 12,
  x.tick.font.color = global.font.color,
  x.tick.font.family = global.font.family, x.tick.font.size = 10,
  y.title = "", y.title.font.color = global.font.color,
  y.title.font.family = global.font.family, y.title.font.size = 12,
  y.tick.font.color = global.font.color,
  y.tick.font.family = global.font.family, y.tick.font.size = 10,
  y.tick.format = "", y.hovertext.font.size = 11,
  y.hovertext.font.color = global.font.color,
  y.hovertext.font.family = global.font.family,
  y.hovertext.format = y.tick.format)
}
\arguments{
\item{x}{Input data may be a matrix or a vector, wth dates as the rownames and data series along the columns.}

\item{range.bars}{Logical; whether the data consists of a single series with low, value, high in the columns, or
multiple series.}

\item{colors}{Character; a named color from grDevices OR a hex value color.}

\item{line.thickness}{Integer; The width of the lines connecting data points.}

\item{legend.width}{Integer; Width (in pixels) of the legend.}

\item{window.start}{The number of days before the end of the data series to start the range selector window.}

\item{global.font.family}{Character; font family for all occurrences of any
font attribute for the chart unless specified individually.}

\item{global.font.color}{Global font color as a named color in character format
(e.g. "black") or an a hex code.}

\item{title}{Character; chart title.}

\item{title.font.family}{Character; title font family. Can be "Arial Black",
"Arial", "Comic Sans MS", "Courier New", "Georgia", "Impact",
"Lucida Console", "Lucida Sans Unicode", "Marlett", "Symbol", "Tahoma",
"Times New Roman", "Trebuchet MS", "Verdana", "Webdings"}

\item{title.font.color}{Title font color as a named color in character
format (e.g. "black") or a hex code.}

\item{title.font.size}{Integer; Title font size; default = 10.}

\item{x.title}{Character, x-axis title; defaults to chart input values;
to turn off set to "FALSE".}

\item{x.title.font.color}{x-axis title font color as a named color in
character format (e.g. "black") or an a hex code.}

\item{x.title.font.family}{Character; x-axis title font family}

\item{x.title.font.size}{Integer; x-axis title font size}

\item{x.tick.font.color}{X-axis tick label font color as a named color in
character format (e.g. "black") or an a hex code.}

\item{x.tick.font.family}{Character; x-axis tick label font family}

\item{x.tick.font.size}{Integer; x-axis tick label font size}

\item{y.title}{Character, y-axis title; defaults to chart input values;
to turn off set to "FALSE".}

\item{y.title.font.color}{y-axis title font color as a named color in
character format (e.g. "black") or a hex code.}

\item{y.title.font.family}{Character; y-axis title font family}

\item{y.title.font.size}{Integer; y-axis title font size}

\item{y.tick.font.color}{y-axis tick label font color as a named color
in character format (e.g. "black") or an a hex code.}

\item{y.tick.font.family}{Character; y-axis tick label font family}

\item{y.tick.font.size}{Integer; y-axis tick label font size}

\item{y.tick.format}{A string representing a d3 formatting code.
See https://github.com/d3/d3/blob/master/API.md#number-formats-d3-format}

\item{y.hovertext.font.size}{Integer; Legend font size.}

\item{y.hovertext.font.color}{Legend font color as a named color in character
format (e.g. "black") or a hex code.}

\item{y.hovertext.font.family}{Character; legend font family.}

\item{y.hovertext.format}{A string representing a d3 formatting code
See https://github.com/d3/d3/blob/master/API.md#number-formats-d3-format}
}
\description{
Plots interactive time series. Either multiple series may be plotted, or a single series with high and low
range/error bars.
}
\examples{
z <- structure(c(1L, 2L, 3L, 4L, 5L, 2L, 3L, 4L, 5L, 6L),  .Dim = c(5L, 2L),
      .Dimnames = list(c("T", "U", "V", "W", "X"), c("A", "B")))
Area(z)
}
